% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Vegetation.R
\name{adj_phenology_by_temp}
\alias{adj_phenology_by_temp}
\title{Adjust a phenological pattern to a different temperature regime}
\usage{
adj_phenology_by_temp(x, ref_temp, target_temp)
}
\arguments{
\item{x}{A numeric vector of length 12. The values
reflect the phenological pattern that occur on average under
reference mean monthly temperature values \code{ref_temp}.}

\item{ref_temp}{A numeric vector of length 12. Reference mean monthly
temperature values in degree Celsius under which \code{x} was
determined / is valid.}

\item{target_temp}{A numeric vector of length 12. Mean monthly
temperature values in degree Celsius of a target site / condition
for which \code{x} is to be adjusted.}
}
\value{
An updated copy of \code{x} where the values have been adjusted
  to represent the phenology of a target climate described by
  \code{target_temp}.
}
\description{
Extract the characteristics of seasonal (mean monthly) values of biomass,
activity, etc. (i.e., phenology) that were observed for a specific
(reference) climate (at a specific site), and project those characteristics
to the mean monthly temperature values of a different site and/or different
climate conditions.
}
\details{
The algorithm attempts to adjust for the following patterns:
\itemize{
  \item Shifts in timing of highest/lowest seasonal biomass/activity
  \item Effects of residual temperature differences not accounted
        for by shifts in timing while maintaining a unimodal phenological
        pattern
  \item Maintain inactive period if present
  \item Maintain seasonality expressed as correlation between phenology
        and mean monthly temperature values
}
}
\examples{
sw_in <- rSOILWAT2::sw_exampleData
phen_reference <- as.data.frame(
  swProd_MonProd_grass(sw_in)[, c("Biomass", "Live_pct")]
)

clim <- calc_SiteClimate(weatherList = rSOILWAT2::weatherData)
ref_temp <- clim[["meanMonthlyTempC"]]
temp_randomwarmer3C <- ref_temp + stats::rnorm(12, 3, 1)

## Adjust phenology from the reference Mar-Oct to a target Nov-Jun
## growing season
phen_adj <- sapply(phen_reference, function(x) adj_phenology_by_temp(
  x = x,
  ref_temp = ref_temp,
  target_temp = temp_randomwarmer3C
))

## Plot reference and adjusted monthly values
\dontrun{
par_prev <- par(mfrow = c(2, 1))

plot(1:12, phen_reference[, 1], type = "l", col = "red",
  ylim = c(0, 250), xlab = "Month")
lines(1:12, phen_adj[, 1])
legend("bottom", ncol = 2,
  legend = c("reference", "adjusted"),
  lwd = 2, lty = 1, col = c("red", "black")
)

plot(1:12, phen_reference[, 2], type = "l", col = "red",
  ylim = c(0, 1), xlab = "Month")
lines(1:12, phen_adj[, 2])

par(par_prev)
}

}
