% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Vegetation.R
\name{adjBiom_by_ppt}
\alias{adjBiom_by_ppt}
\title{Adjust mean monthly biomass values by precipitation}
\usage{
adjBiom_by_ppt(
  biom_shrubs,
  biom_C3,
  biom_C4,
  biom_annuals,
  biom_maxs,
  map_mm_shrubs,
  map_mm_std_shrubs,
  map_mm_grasses,
  map_mm_std_grasses,
  vegcomp_std_shrubs,
  vegcomp_std_grass
)
}
\description{
Adjust mean monthly biomass values by precipitation
}
\section{Details}{
 Internally used by
  \code{\link{estimate_PotNatVeg_biomass}}.
}

