% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Vegetation.R
\name{adjBiom_by_temp}
\alias{adjBiom_by_temp}
\title{Adjust biomass phenology by pattern of growing/non-growing seasons to
site-specific climatic conditions}
\usage{
adjBiom_by_temp(x, mean_monthly_temp_C, growing_limit_C = 4,
  isNorth = TRUE)
}
\arguments{
\item{x}{A two-dimensional object or a list of such objects. Rows of each
such object correspond to month of the year, i.e., there must be 12 rows,
and columns correspond to variables that are adjusted.}

\item{mean_monthly_temp_C}{A numeric vector of length 12. Mean monthly
temperatures in Celsius. The default inputs have March-October as
growing season.}

\item{growing_limit_C}{A numeric value. \code{mean_monthly_temp_C} equal or
above this limit are considered suitable for growth (growing season).
Default value is 4 C.}

\item{isNorth}{A logical value. \code{TRUE} for locations on the northern
hemisphere, \code{FALSE} for locations on the southern hemisphere.}
}
\value{
An object as \code{x}, i.e., a two-dimensional object or a list of
  such objects, where the values were adjusted.
}
\description{
Adjust biomass phenology by pattern of growing/non-growing seasons to
site-specific climatic conditions
}
\examples{
sw_in <- rSOILWAT2::sw_exampleData
biomass_reference <- swProd_MonProd_grass(sw_in)[, 2:3]

## Adjust phenology from the reference Mar-Oct to a target Nov-Jun
## growing season
biomass_adj <- adjBiom_by_temp(
  x = biomass_reference,
  mean_monthly_temp_C = c(rep(10, 6), rep(0, 4), rep(10, 2)),
  growing_limit_C = 4)

## Plot reference and adjusted monthly values
\dontrun{
par_prev <- par(mfrow = c(2, 1))

plot(1:12, biomass_reference[, 1], type = "l", col = "red",
  ylim = c(0, 250), xlab = "Month")
lines(1:12, biomass_adj[, 1])

plot(1:12, biomass_reference[, 2], type = "l", col = "red",
  ylim = c(0, 1), xlab = "Month")
lines(1:12, biomass_adj[, 2])

par(par_prev)
}

}
