% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Vegetation.R
\name{adjust_TranspirationRegions}
\alias{adjust_TranspirationRegions}
\title{Adjust transpiration regions for roots and soil layers}
\usage{
adjust_TranspirationRegions(swInputData)
}
\arguments{
\item{swInputData}{A \pkg{rSOILWAT2} input object of class
\code{\linkS4class{swInputData}}.}
}
\value{
A transpiration region matrix.
}
\description{
Adjust transpiration regions for roots and soil layers
}
\examples{
sw_in <- rSOILWAT2::sw_exampleData

adjust_TranspirationRegions(sw_in)

}
