% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Vegetation.R
\name{biomass}
\alias{biomass}
\alias{Shrub_ANPP}
\alias{Grass_ANPP}
\title{Biomass equations}
\usage{
Shrub_ANPP(MAP_mm)

Grass_ANPP(MAP_mm)
}
\arguments{
\item{MAP_mm}{A numeric vector. Mean annual precipitation in
millimeters (mm).}
}
\description{
Biomass equations

Estimate shrub biomass density from mean annual precipitation

Estimate grass biomass density from mean annual precipitation
}
\references{
Milchunas & Lauenroth 1993 (Fig. 2):
  Y [g/m2/yr] = c1 * MAP [mm/yr] + c2
}
