% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{calc_dailyInputFlags}
\alias{calc_dailyInputFlags}
\title{Determine used weather variables based on values}
\usage{
calc_dailyInputFlags(x, name_data = weather_dataColumns())
}
\arguments{
\item{x}{Weather data, i.e.,
a list where each element is of class \code{\link{swWeatherData}}, or
a data frame with appropriate columns (see \code{\link[=dbW_weatherData_to_dataframe]{dbW_weatherData_to_dataframe()}}).}

\item{name_data}{A vector of character strings. The column names of \code{x}
with weather variables.}
}
\value{
A logical vector for each of the possible input variables with
\code{TRUE} if at least one value is not missing.
}
\description{
Determine used weather variables based on values
}
\examples{
calc_dailyInputFlags(rSOILWAT2::weatherData)
calc_dailyInputFlags(dbW_weatherData_to_dataframe(rSOILWAT2::weatherData))


}
