% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_SoilMTRegimes.R
\name{calc_RRs_Chambers2014}
\alias{calc_RRs_Chambers2014}
\title{Determine resilience & resistance classes (sensu Chambers et al. 2014) based
on soil moisture and soil temperature regimes}
\usage{
calc_RRs_Chambers2014(Tregime, Sregime, MAP_mm)
}
\arguments{
\item{Tregime}{A named numeric vector. The soil temperature regime
\var{\dQuote{STR}} element of the return object of
\code{\link{calc_SMTRs}}.}

\item{Sregime}{A named numeric vector. The soil moisture regime
\var{\dQuote{SMR}} element of the return object of
\code{\link{calc_SMTRs}}.}

\item{MAP_mm}{A numeric value. The mean annual precipitation in millimeters.}
}
\value{
A named numeric vector of \code{0s} and \code{1s} indicating
  the matching resilience and resistance class. Note: All elements may be
  0 if the input soil moisture/temperature regimes are not covered by
  Chambers et al. 2014.
}
\description{
Determine resilience & resistance classes (sensu Chambers et al. 2014) based
on soil moisture and soil temperature regimes
}
\examples{
# Calculate soil moisture and soil temperature regimes
sw_in <- rSOILWAT2::sw_exampleData
sw_out <- sw_exec(inputData = sw_in)
SMTR <- calc_SMTRs(sim_in = sw_in, sim_out = sw_out)

# Determine average across years and set aggregation agreement level
Tregime <- colMeans(SMTR[["STR"]]) >= 0.9
Sregime <- colMeans(SMTR[["SMR"]]) >= 0.9

# Calculate resilience and resistance categories
clim <- calc_SiteClimate(weatherList = get_WeatherHistory(sw_in))
calc_RRs_Chambers2014(Tregime, Sregime, MAP_mm = 10 * clim[["MAP_cm"]])

}
\references{
Chambers, J. C., D. A. Pyke, J. D. Maestas, M. Pellant,
  C. S. Boyd, S. B. Campbell, S. Espinosa, D. W. Havlina, K. E. Mayer, and
  A. Wuenschel. 2014. Using Resistance and Resilience Concepts to Reduce
  Impacts of Invasive Annual Grasses and Altered Fire Regimes on the
  Sagebrush Ecosystem and Greater Sage-Grouse: A Strategic Multi-Scale
  Approach. Gen. Tech. Rep. RMRS-GTR-326. U.S. Department of Agriculture,
  Forest Service, Rocky Mountain Research Station, Fort Collins, CO.
}
