% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_SoilMTRegimes.R
\name{calc_RRs_Maestas2016}
\alias{calc_RRs_Maestas2016}
\title{Determine resilience & resistance classes (sensu Maestas et al. 2016) based
on soil moisture and soil temperature regimes}
\usage{
calc_RRs_Maestas2016(Tregime, Sregime)
}
\arguments{
\item{Tregime}{A named numeric vector. The soil temperature regime
\var{\dQuote{STR}} element of the return object of
\code{\link{calc_SMTRs}}.}

\item{Sregime}{A named numeric vector. The soil moisture regime
\var{\dQuote{SMR}} element of the return object of
\code{\link{calc_SMTRs}}.}
}
\value{
A named numeric vector of \code{0s} and \code{1s} indicating
  the matching resilience and resistance class. Note: All elements may be
  0 if the input soil moisture/temperature regimes are not covered by
  Maestas et al. 2016.
}
\description{
Determine resilience & resistance classes (sensu Maestas et al. 2016) based
on soil moisture and soil temperature regimes
}
\examples{
# Calculate soil moisture and soil temperature regimes
sw_in <- rSOILWAT2::sw_exampleData
sw_out <- sw_exec(inputData = sw_in)
SMTR <- calc_SMTRs(sim_in = sw_in, sim_out = sw_out)

# Determine average across years and set aggregation agreement level
Tregime <- colMeans(SMTR[["STR"]]) >= 0.9
Sregime <- colMeans(SMTR[["SMR"]]) >= 0.9

# Calculate resilience and resistance categories
calc_RRs_Maestas2016(Tregime, Sregime)

}
\references{
Maestas, J.D., Campbell, S.B., Chambers, J.C., Pellant, M. &
  Miller, R.F. (2016). Tapping Soil Survey Information for Rapid Assessment
  of Sagebrush Ecosystem Resilience and Resistance. Rangelands, 38, 120-128.
}
