% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Miscellaneous_Functions.R
\name{calc_SiteClimate}
\alias{calc_SiteClimate}
\title{Calculate climate variables from daily weather}
\usage{
calc_SiteClimate(
  weatherList,
  year.start = NA,
  year.end = NA,
  do_C4vars = FALSE,
  do_Cheatgrass_ClimVars = FALSE,
  simTime2 = NULL,
  latitude = 90
)
}
\arguments{
\item{weatherList}{A list. Each element is an object of class
\code{\link[rSOILWAT2:swWeatherData-class]{rSOILWAT2::swWeatherData}}
containing daily weather data of a specific year.}

\item{year.start}{An integer value. The first year of the range of years for
which climate variables should be calculated.}

\item{year.end}{An integer value. The last year of the range of years for
which climate variables should be calculated.}

\item{do_C4vars}{A logical value. If \code{TRUE} then additional output is
returned.}

\item{do_Cheatgrass_ClimVars}{A logical value. If \code{TRUE} then additional
output is returned.}

\item{simTime2}{A list with two named elements. The elements are numeric
vectors \var{month_ForEachUsedDay_NSadj} and
\var{year_ForEachUsedDay_NSadj}; they are calculated internally
if \code{NULL}; alternatively, they can be generated by a call to the
function \code{\link[rSW2data]{simTiming_ForEachUsedTimeUnit}}.
They are only used if \code{isTRUE(do_C4vars)}.}

\item{latitude}{A numeric value. The latitude in degree of the site. Used
to adjust seasons between northern/southern hemisphere only if
\code{isTRUE(do_C4vars)} and \code{simTime2} has to be re-calculated.}
}
\value{
A list with named elements \itemize{
  \item{\var{\dQuote{meanMonthlyTempC}}} {A numeric vector of length 12.
   Mean monthly mean daily air temperature in degree Celsius.}
  \item{\var{\dQuote{minMonthlyTempC}}} {A numeric vector of length 12.
    Mean monthly minimum daily air temperature in degree Celsius.}
  \item{\var{\dQuote{maxMonthlyTempC}}} {A numeric vector of length 12.
    Mean monthly maximum daily air temperature in degree Celsius.}
  \item{\var{\dQuote{meanMonthlyPPTcm}}} {A numeric vector of length 12.
    Mean monthly precipitation in centimeters.}
  \item{\var{\dQuote{MAP_cm}}} {A numeric value. Mean annual precipitation
    in centimeters.}
  \item{\var{\dQuote{MAT_C}}} {A numeric value. Mean annual air temperature
    in degree Celsius.}
  \item{\var{\dQuote{dailyTempMin}}} {A numeric vector. If
    \code{isTRUE(do_C4vars)}, then minimum daily air temperature in degree
    Celsius for each day of time period between \code{year.start} and
    \code{year.end}. If \code{!isTRUE(do_C4vars)}, then \code{NA}.}
  \item{\var{\dQuote{dailyTempMean}}} {A numeric vector. Similar as for
    \code{dailyTempMin} but for mean daily air temperature.}
  \item{\var{\dQuote{dailyC4vars}}} {If \code{isTRUE(do_C4vars)}, then a
    named numeric vector containing the output of
    \code{\link{sw_dailyC4_TempVar}}, else \code{NA}.}
  \item{\var{\dQuote{monthlyCheatgrass_ClimVars}}} {
    If \code{isTRUE(do_Cheatgrass_ClimVars)}, then a named numeric vector
    containing the output of \code{\link{sw_Cheatgrass_ClimVar}},
    else \code{NA}.}
}
}
\description{
Calculate climate variables from daily weather
}
\examples{
## Load weather dataset from rSOILWAT2
utils::data("weatherData", package = "rSOILWAT2")
clim1 <- calc_SiteClimate(weatherList = weatherData)
clim2 <- calc_SiteClimate(weatherList = weatherData, do_C4vars = TRUE)
clim3 <- calc_SiteClimate(weatherList = weatherData,
  do_Cheatgrass_ClimVars = TRUE)

}
