% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_SoilMTRegimes.R
\name{Chambers2014_Table1}
\alias{Chambers2014_Table1}
\title{Table 1 in Chambers et al. 2014}
\usage{
Chambers2014_Table1()
}
\description{
Table 1 in Chambers et al. 2014
}
\references{
Chambers, J. C., D. A. Pyke, J. D. Maestas, M. Pellant,
  C. S. Boyd, S. B. Campbell, S. Espinosa, D. W. Havlina, K. E. Mayer, and
  A. Wuenschel. 2014. Using Resistance and Resilience Concepts to Reduce
  Impacts of Invasive Annual Grasses and Altered Fire Regimes on the
  Sagebrush Ecosystem and Greater Sage-Grouse: A Strategic Multi-Scale
  Approach. Gen. Tech. Rep. RMRS-GTR-326. U.S. Department of Agriculture,
  Forest Service, Rocky Mountain Research Station, Fort Collins, CO.
}
