% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swFunctions.R
\name{check_content}
\alias{check_content}
\alias{dbW_has_sites}
\alias{dbW_has_siteIDs}
\alias{dbW_has_scenarioIDs}
\alias{dbW_has_scenarios}
\alias{dbW_has_weatherData}
\title{Check availability of content in registered weather database}
\usage{
dbW_has_sites(Labels, ignore.case = FALSE)

dbW_has_siteIDs(Site_ids)

dbW_has_scenarioIDs(scenario_ids)

dbW_has_scenarios(scenarios, ignore.case = FALSE)

dbW_has_weatherData(Site_ids, Scenario_ids)
}
\arguments{
\item{Labels}{A vector of character strings. The names/labels of the queried sites.}

\item{ignore.case}{A vector of character strings. The names/labels of the queried scenarios.}

\item{Site_ids}{An integer vector. The IDs/database keys of the queried site.}

\item{scenario_ids}{An integer vector. The IDs/database keys of the queried scenario.}

\item{scenarios}{An integer vector. The IDs/database keys of the queried scenarios.}
}
\value{
\code{dbW_has_siteIDs} returns a logical vector of the length of queried sites.

\code{dbW_has_siteIDs} returns a logical vector of the length of queried sites.

\code{dbW_has_scenarios} returns a logical vector of the length of queried
 scenarios.

\code{dbW_has_scenarios} returns a logical vector of the length of queried
 scenarios.

code{dbW_has_weatherData} returns a logical matrix with rows corresponding to
 queried sites and columns to queried scenarios.
}
\description{
Check availability of content in registered weather database
}
\section{Details}{
 \code{dbW_has_siteIDs} checks whether sites are available.

 \code{dbW_has_siteIDs} checks whether sites are available.

 \code{dbW_has_scenarios} checks whether scenarios are available.

 \code{dbW_has_scenarios} checks whether scenarios are available.

 \code{dbW_has_weatherData} checks whether weather data are available.
}

