% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Pedotransfer_Functions.R
\name{check_pdf_availability}
\alias{check_pdf_availability}
\title{Check availability of \code{PDFs}}
\usage{
check_pdf_availability(pdfs = names(pdf_names()), verbose = interactive())
}
\arguments{
\item{pdfs}{A character vector. \code{PDF} names to be checked;
defaults to \code{\link[=pdf_names]{pdf_names()}}.}

\item{verbose}{A logical value.}
}
\value{
A named logical vector with current availability of \code{PDFs};
\code{pdfs} that are not implemented return \code{NA}.
}
\description{
PDFs implemented in \code{SOILWAT2} are always available;
PDFs implemented in \code{rSOILWAT2} may have additional requirements, e.g.,
live internet connection or access to specific data files.
}
\examples{
check_pdf_availability()
check_pdf_availability("neuroFX2021")
check_pdf_availability("nonexistent_PDF")

}
