% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Pedotransfer_Functions.R
\name{check_SWRC_vs_PDF}
\alias{check_SWRC_vs_PDF}
\title{Check whether PDF and SWRC are compatible and implemented}
\usage{
check_SWRC_vs_PDF(swrc_name, pdf_name, fail = FALSE)
}
\arguments{
\item{swrc_name}{An character string or vector.
The selected \code{SWRC} name
(one of \code{\link[=swrc_names]{swrc_names()}}, with default \code{"Campbell1974"}).}

\item{pdf_name}{An character string or vector.
The selected \code{PDF} name
(one of \code{\link[=pdf_names]{pdf_names()}}, with default \code{"Cosby1984AndOthers"}).}

\item{fail}{A logical value.
Issue a warning (\code{FALSE}, default) or throw an error (\code{TRUE})
if request fails.}
}
\value{
A logical vector.
}
\description{
Check whether PDF and SWRC are compatible and implemented
}
\examples{
check_SWRC_vs_PDF("Campbell1974", c("Cosby1984", "Rosetta3"))

}
