% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Pedotransfer_Functions.R
\name{check_swrcp}
\alias{check_swrcp}
\title{Check Soil Water Retention Curve parameters}
\usage{
check_swrcp(swrc_type = 1L, swrcp)
}
\arguments{
\item{swrc_type}{An integer value or vector.
The identification number of selected \code{SWRC}.}

\item{swrcp}{A numeric vector or matrix.
The parameters of a selected \code{SWRC};
each row represents one \code{SWRC}, e.g., one per soil layer.}
}
\description{
Check Soil Water Retention Curve parameters
}
\section{Notes}{

The argument selecting \code{SWRC} (\code{swrc_type}) is recycled
for multiple parameter sets, i.e., rows of \code{swrcp}.
}

\section{Details}{

See \link{SWRCs} for suitable values of \code{swrc_type}.
}

\examples{
swrc_type <- 1
swrcp <- pdf_estimate(
  sand = c(0.5, 0.3),
  clay = c(0.2, 0.1),
  fcoarse = c(0, 0),
  swrc_type = swrc_type
)

check_swrcp(swrc_type, swrcp)
check_swrcp(rep(swrc_type, 2), swrcp)
check_swrcp(swrc_type, swrcp[1, ])

swrcp2 <- swrcp
swrcp2[1, 1] <- -10
check_swrcp(rep(swrc_type, 2), swrcp2)

}
\seealso{
\code{\link[=pdf_estimate]{pdf_estimate()}}
}
