% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_upgrade.R
\name{check_updatedDB}
\alias{check_updatedDB}
\title{Run database integrity checks on a weather database}
\usage{
check_updatedDB(con)
}
\arguments{
\item{con}{A \link[DBI:DBIConnection-class]{DBI::DBIConnection} object.
The connection to a weather database.}
}
\description{
Run database integrity checks on a weather database
}
