% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Mathematical_Functions.R
\name{circ_minus}
\alias{circ_minus}
\title{Calculate the circular subtraction \var{x - y}}
\usage{
circ_minus(x, y, int, type = c("minusPiPlusPi", "ZeroPlus2Pi"))
}
\arguments{
\item{x}{A numeric vector or array.}

\item{y}{A numeric vector or array.}

\item{int}{A numeric value. The number of units of \code{x} in a full circle,
e.g., for unit days: \code{int = 365}; for unit months: \code{int = 12}.}

\item{type}{A character string. If \code{type == "minusPiPlusPi"},
then the resulting value lies between \code{-int / 2} and \code{int / 2}.
If \code{type == "ZeroPlus2Pi"},
then the resulting value lies between \code{0} and \code{int}.}
}
\description{
Calculate the circular subtraction \var{x - y}
}
\examples{
# Days of year
circ_minus(260, 240, int = 365) ## expected: +20
circ_minus(240, 260, int = 365) ## expected: -20
circ_minus(240, 260, int = 365, type = "ZeroPlus2Pi") ## expected: 345
circ_minus(10, 360, int = 365) ## expected: +15
circ_minus(360, 10, int = 365) ## expected: -15
circ_minus(360, 10, int = 365, type = "ZeroPlus2Pi") ## expected: 350
circ_minus(0, 360, int = 365) ## expected: +5
circ_minus(360, 0, int = 365) ## expected: -5
circ_minus(360, 0, int = 365, type = "ZeroPlus2Pi") ## expected: 360

# Matrix examples
x <- matrix(c(260, 240, 10, 360, 0, 360), nrow = 3, ncol = 2)
y <- matrix(c(240, 260, 360, 10, 360, 0), nrow = 3, ncol = 2)
circ_minus(x, y, int = 365)
y2 <- y
y2[1, 1] <- NA
circ_minus(y2, x, int = 365)

}
