% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Mathematical_Functions.R
\name{circ_seq}
\alias{circ_seq}
\title{Sequence generation for circular data}
\usage{
circ_seq(from, to, int, by, length.out = NULL)
}
\arguments{
\item{from}{the starting and (maximal) end values of the
    sequence.  Of length \code{1} unless just \code{from} is supplied as
    an unnamed argument.}

\item{to}{the starting and (maximal) end values of the
    sequence.  Of length \code{1} unless just \code{from} is supplied as
    an unnamed argument.}

\item{int}{A numeric value. The number of units of \code{x} in a full circle,
e.g., for unit days: \code{int = 365}; for unit months: \code{int = 12}.}

\item{by}{number: increment of the sequence.}

\item{length.out}{desired length of the sequence.  A
    non-negative number, which for \code{seq} and \code{seq.int} will be
    rounded up if fractional.}
}
\description{
Sequence generation for circular data
}
\examples{
circ_seq(5, 8, int = 12)          ## expected c(5, 6, 7, 8)
circ_seq(-7, 8, int = 12)         ## expected c(5, 6, 7, 8)
circ_seq(-2, 3, int = 12)         ## expected c(10, 11, 12, 1, 2, 3)
circ_seq(-2, 3, int = 12, by = 2) ## expected c(10, 12, 2)
circ_seq(-2, 3, int = 12, length.out = 3) ## expected c(10, 0.5, 3)

}
\seealso{
\code{\link{seq}}
}
