% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Mathematical_Functions.R
\name{circular}
\alias{circular}
\alias{circ_mean}
\alias{circ_range}
\alias{circ_sd}
\title{Functions for circular descriptive statistics}
\usage{
circ_mean(x, int, type = c("minusPiPlusPi", "ZeroPlus2Pi"), na.rm = FALSE)

circ_range(x, int, na.rm = FALSE)

circ_sd(x, int, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector or a matrix. If a data.frame is supplied, then
\code{x} is coerced to a matrix.}

\item{int}{A numeric value. The number of units of \code{x} in a full circle,
e.g., for unit days: \code{int = 365}; for unit months: \code{int = 12}.}

\item{type}{A character string. If \code{type == "minusPiPlusPi"},
then the resulting value lies between \code{-int / 2} and \code{int / 2}.
If \code{type == "ZeroPlus2Pi"},
then the resulting value lies between \code{0} and \code{int}.}

\item{na.rm}{A logical value indicating whether \code{NA} values should be
stripped before the computation proceeds.}
}
\value{
A numeric value or \code{NA}.
}
\description{
Functions for circular descriptive statistics
}
\examples{
x <- 1:3
circ_mean(x, int = 12, type = "minusPiPlusPi") ## expected 2
circ_mean(x, int = 12, type = "ZeroPlus2Pi") ## expected 2
x <- (-2):0
circ_mean(x, int = 12, type = "minusPiPlusPi") ## expected -1
circ_mean(x, int = 12, type = "ZeroPlus2Pi") ## expected 11
x <- (-5):5
circ_mean(x, int = 12, type = "minusPiPlusPi") ## expected 0
circ_mean(x, int = 12, type = "ZeroPlus2Pi") ## expected 12
x <- (-5):8
circ_mean(x, int = 12, type = "minusPiPlusPi") ## expected -4.5
circ_mean(x, int = 12, type = "ZeroPlus2Pi") ## expected 7.5

}
\seealso{
\code{\link[circular]{mean.circular}},
  \code{\link[circular]{range.circular}},
  \code{\link[circular]{sd.circular}}
}
