% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swWeatherGenerator.R
\name{compare_weather}
\alias{compare_weather}
\title{Compare two weather datasets: produces comparison plots for means, quantiles,
and Markov weather generator input parameters for all time steps}
\usage{
compare_weather(ref_weather, weather, N, WET_limit_cm = 0, path, tag)
}
\arguments{
\item{ref_weather}{A \code{list} of or a two-dimensional numerical object,
e.g., \code{data.frame} or \code{matrix}. If a \code{list} then four
suitable elements one for each \code{time_step}. Otherwise, daily weather
data with columns \var{Year}, \var{DOY}, \var{Tmax_C}, \var{Tmin_C},
and \var{PPT_cm}; for instance, the result of
function \code{\link{dbW_weatherData_to_dataframe}}. This represents the
reference weather against which \code{weather} is compared.}

\item{weather}{A \code{list} of elements such as described for
\code{ref_weather} or an object such as \code{ref_weather}. This represents
the weather data (potentially from many sites and/or runs) that are
compared against \code{ref_weather}.}

\item{N}{An integer number representing the number of runs or sites contained
in \code{weather}.}

\item{WET_limit_cm}{A numeric value. A day with more precipitation than
this value is considered \var{wet} instead of \var{dry}. Default is 0.
This values should be equal to the corresponding value used in
\var{SOILWAT2}'s function \code{SW_MKV_today}.}

\item{path}{A character string. The directory path in which output figures
will be saved.}

\item{tag}{A character string to uniquely identify a set of output figures.}
}
\value{
This function is called for its side effects of producing figures
  that are saved as \var{png} files on the disk.
}
\description{
Compare two weather datasets: produces comparison plots for means, quantiles,
and Markov weather generator input parameters for all time steps
}
\section{Notes}{
 The number of days represented by \code{ref_weather} and by
  \code{weather} does not need to be the same.

 See also the Weather generator integration tests (in file
  \var{"tests/testthat/test_WeatherGenerator_functionality.R}).
}

\examples{
path <- tempdir()

## Example with default rSOILWAT2 weather data
w0 <- dbW_weatherData_to_dataframe(rSOILWAT2::weatherData)
w1 <- w0
w1[, "Tmax_C"] <- w1[, "Tmax_C"] + 2
w1[, "Tmin_C"] <- w1[, "Tmin_C"] - 2

compare_weather(
  ref_weather = w0,
  weather = w1,
  N = 1,
  path = path,
  tag = "Example1-Silly"
)

## Example with STEPWAT2 output data averaged across iterations (`-o` option)
fname_main <- "file_STEPWAT2_main.csv"
fname_dev <- "file_STEPWAT2_dev.csv"
if (all(file.exists(fname_main, fname_dev))) {
  cols_STEPWAT2 <- c(
    "Year", "Day", "PRECIP_ppt_Mean", "TEMP_max_C_Mean", "TEMP_min_C_Mean"
  )
  cols_rSOILWAT2 <- c("Year", "Day", "PPT_cm", "Tmax_C", "Tmin_C")
  w0 <- utils::read.csv(fname_main)[, cols_STEPWAT2]
  colnames(w0) <- cols_rSOILWAT2
  w1 <- utils::read.csv(fname_dev)[, cols_STEPWAT2]
  colnames(w1) <- cols_rSOILWAT2

  # Note: Since values are averages across many iterations, most days
  # have average precipitation values > 0; thus, we need to adjust
  # `WET_limit_cm` accordingly (here, with a guess)
  compare_weather(
    ref_weather = w0,
    weather = w1,
    N = 1,
    WET_limit_cm = 0.1,
    path = path,
    tag = "Example2-STEPWAT2"
  )
}

## Cleanup
unlink(list.files(path), force = TRUE)

}
