% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_SoilMTRegimes.R
\name{dailyNRCS_Chambers2014_ResilienceResistance}
\alias{dailyNRCS_Chambers2014_ResilienceResistance}
\title{Determine resilience & resistance classes (sensu Chambers et al. 2014) based
on soil moisture and soil temperature regimes}
\usage{
dailyNRCS_Chambers2014_ResilienceResistance(Tregime, Sregime, MAP_mm)
}
\description{
Determine resilience & resistance classes (sensu Chambers et al. 2014) based
on soil moisture and soil temperature regimes
}
\references{
Chambers, J. C., D. A. Pyke, J. D. Maestas, M. Pellant,
  C. S. Boyd, S. B. Campbell, S. Espinosa, D. W. Havlina, K. E. Mayer, and
  A. Wuenschel. 2014. Using Resistance and Resilience Concepts to Reduce
  Impacts of Invasive Annual Grasses and Altered Fire Regimes on the
  Sagebrush Ecosystem and Greater Sage-Grouse: A Strategic Multi-Scale
  Approach. Gen. Tech. Rep. RMRS-GTR-326. U.S. Department of Agriculture,
  Forest Service, Rocky Mountain Research Station, Fort Collins, CO.
}
