% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_SoilMTRegimes.R
\name{dailyNRCS_Maestas2016_ResilienceResistance}
\alias{dailyNRCS_Maestas2016_ResilienceResistance}
\title{Determine resilience & resistance classes (sensu Maestas et al. 2016) based
on soil moisture and soil temperature regimes}
\usage{
dailyNRCS_Maestas2016_ResilienceResistance(Tregime, Sregime)
}
\description{
Determine resilience & resistance classes (sensu Maestas et al. 2016) based
on soil moisture and soil temperature regimes
}
\references{
Maestas, J.D., Campbell, S.B., Chambers, J.C., Pellant, M. &
  Miller, R.F. (2016). Tapping Soil Survey Information for Rapid Assessment
  of Sagebrush Ecosystem Resilience and Resistance. Rangelands, 38, 120-128.
}
