% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Time_SimulationWorld.R
\name{days_in_years}
\alias{days_in_years}
\title{Create a sequence of each day between and including two calendar years}
\usage{
days_in_years(start_year, end_year)
}
\arguments{
\item{start_year}{An integer value. The first year.}

\item{end_year}{An integer value. The last year.}
}
\description{
Create a sequence of each day between and including two calendar years
}
\examples{
length(days_in_years(1980, 1980)) == 366

}
