% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_addFromFolders}
\alias{dbW_addFromFolders}
\title{Add weather data from \pkg{SOILWAT2} weather "folders" to a weather database}
\usage{
dbW_addFromFolders(
  MetaData = NULL,
  FoldersPath = ".",
  ScenarioName = "Current",
  weather_tag = "weath"
)
}
\arguments{
\item{MetaData}{A data.frame with suitable columns.}

\item{FoldersPath}{A character string.}

\item{ScenarioName}{A character string.}

\item{weather_tag}{A character string.}
}
\description{
Add weather data from \pkg{SOILWAT2} weather "folders" to a weather database
}
