% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSOILWAT2_deprecated.R
\name{dbW_addSite}
\alias{dbW_addSite}
\title{Add a new site description to a weather database}
\usage{
dbW_addSite(Site_id = NULL, lat = NULL, long = NULL, Label = NULL)
}
\arguments{
\item{Site_id}{An integer number.}

\item{lat}{A numeric value.}

\item{long}{A numeric value.}

\item{Label}{A character string.}
}
\description{
Add a new site description to a weather database
}
