% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_addSites}
\alias{dbW_addSites}
\title{Adds new sites to a registered weather database}
\usage{
dbW_addSites(site_data, ignore.case = FALSE, verbose = FALSE)
}
\arguments{
\item{site_data}{A data.frame. The site data with column names
\var{Latitude}, \var{Longitude}, and \var{Label}.}

\item{ignore.case}{A logical value.}

\item{verbose}{A logical value.}
}
\value{
An invisible logical value indicating success with \code{TRUE} and
 failure with \code{FALSE}.
}
\description{
Adds new sites to a registered weather database
}
