% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_addWeatherData}
\alias{dbW_addWeatherData}
\title{Adds daily weather data to a registered weather database}
\usage{
dbW_addWeatherData(
  Site_id = NULL,
  lat = NULL,
  long = NULL,
  tol_xy = 1e-04,
  weatherFolderPath = NULL,
  weatherData = NULL,
  Label = NULL,
  Scenario_id = NULL,
  Scenario = "Current",
  weather_tag = "weath",
  elevation = NA,
  ignore.case = FALSE,
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{Site_id}{An integer value. The IDs/database key of the queried site.}

\item{lat}{A numeric vector or \code{NULL}. The latitude in decimal degrees
of \code{WGS84}. Northern latitude are positive, sites on the southern
hemisphere have negative values.}

\item{long}{A numeric vector or \code{NULL}. The longitude in decimal degrees
of \code{WGS84}. Eastern longitudes are positive, sites on the western
hemisphere have negative values.}

\item{tol_xy}{A numeric value. The tolerance used to match requested
longitude and latitude values.}

\item{weatherFolderPath}{A character string. The path to the parent folder.}

\item{weatherData}{A list of elements of class \code{\link[=swWeatherData-class]{swWeatherData}}
that each hold daily weather data for one calendar year.}

\item{Label}{A character string. The name/label of the queried site.}

\item{Scenario_id}{An integer value The ID/database key of the queried
scenario.}

\item{Scenario}{A character string. The name/label of a climate scenario.}

\item{weather_tag}{A character string. The base file name without extension
for \code{SOILWAT2}-formatted input files; default is \code{"weath"}}

\item{elevation}{A numeric value. Site elevation above sea level `[m]`.
Used only if specific humidity is provided as input
for calculating relative humidity.}

\item{ignore.case}{A logical value.}

\item{overwrite}{A logical value. Should weather data that already exists
in the data base be overwritten?}

\item{verbose}{A logical value.}
}
\value{
An invisible logical value indicating success with \code{TRUE} and
 failure with \code{FALSE}.
}
\description{
Adds daily weather data to a registered weather database
}
