% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_check_version}
\alias{dbW_check_version}
\title{Check that version of registered weather database is up-to-date}
\usage{
dbW_check_version(dbW_min_version = NULL)
}
\value{
A logical value.
}
\description{
Check that version of registered weather database is up-to-date
}
