% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_check_weatherData}
\alias{dbW_check_weatherData}
\title{Check that weather data is well-formed}
\usage{
dbW_check_weatherData(x, check_all = TRUE)
}
\arguments{
\item{x}{An object.}

\item{check_all}{A logical value}
}
\value{
A logical value.
}
\description{
Check that weather data is organized in a list
where each element is of class \code{\linkS4class{swWeatherData}}, and
represents daily data for one Gregorian year
}
\examples{
dbW_check_weatherData(rSOILWAT2::weatherData)
dbW_check_weatherData(weatherHistory())
dbW_check_weatherData(weatherHistory(), check_all = FALSE)


}
