% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swFunctions.R
\name{dbW_compression}
\alias{dbW_compression}
\title{Query compression type of registered weather database}
\usage{
dbW_compression()
}
\value{
A character string.
}
\description{
Query compression type of registered weather database
}
