% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_dataframe_aggregate}
\alias{dbW_dataframe_aggregate}
\title{Aggregate daily weather data.frame to weekly, monthly, or yearly values}
\usage{
dbW_dataframe_aggregate(dailySW, time_step = c("Year", "Month", "Week",
  "Day"), na.rm = FALSE)
}
\description{
Aggregate daily weather data.frame to weekly, monthly, or yearly values
}
