% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_dataframe_to_monthly}
\alias{dbW_dataframe_to_monthly}
\title{Conversion: object of daily weather data.frame to matrix of monthly values
(\var{mean Tmax}, \var{mean Tmin}, \var{sum PPT})}
\usage{
dbW_dataframe_to_monthly(dailySW, na.rm = FALSE)
}
\description{
Conversion: object of daily weather data.frame to matrix of monthly values
(\var{mean Tmax}, \var{mean Tmin}, \var{sum PPT})
}
