% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_dataframe_to_weatherData}
\alias{dbW_dataframe_to_weatherData}
\title{Conversion: data.frame to object of class \code{\linkS4class{swWeatherData}}}
\usage{
dbW_dataframe_to_weatherData(weatherDF, years = NULL,
  weatherDF_dataColumns = c("DOY", "Tmax_C", "Tmin_C", "PPT_cm"),
  round = 2)
}
\description{
Conversion: data.frame to object of class \code{\linkS4class{swWeatherData}}
}
