% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_delete_duplicated_weatherData}
\alias{dbW_delete_duplicated_weatherData}
\title{Remove duplicated weather data records}
\usage{
dbW_delete_duplicated_weatherData(
  site_id = NULL,
  check_values = TRUE,
  carefully = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{site_id}{A numeric value.
If \code{NULL}, then duplicates of any sites are deleted.
Otherwise, duplicates for the requested site only are deleted.}

\item{check_values}{A logical value. See details.}

\item{carefully}{A logical value. If \code{TRUE} and \code{site_id} is specified,
then first count all and unique weather data records to determine
if there could be duplicate records before attempting to delete them.
Counting instead of identifying duplicates can be faster in certain
situations.}

\item{verbose}{A logical value.}
}
\value{
The number of deleted records
}
\description{
Remove duplicated weather data records
}
\section{Details}{

A weather data record is considered a duplicate if \code{site_id}, \code{scenario_id},
\code{start_year}, and \code{end_year} agree; if \code{check_values} is \code{TRUE}, then
the daily weather values must also exactly agree (checked as blobs)
to be considered a duplicate entry.
}

