% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_deleteSiteData}
\alias{dbW_deleteSiteData}
\title{Delete a weather data record from a registered weather database}
\usage{
dbW_deleteSiteData(Site_id, Scenario_id = NULL)
}
\arguments{
\item{Site_id}{An integer value. The IDs/database key of the queried site.}

\item{Scenario_id}{An integer value The ID/database key of the queried
scenario.}
}
\value{
An invisible logical value indicating success with \code{TRUE} and
  failure with \code{FALSE}.
}
\description{
Delete a weather data record from a registered weather database
}
