% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_disconnectConnection}
\alias{dbW_disconnectConnection}
\title{De-registers/deconnects a SQLite weather database from the package}
\usage{
dbW_disconnectConnection()
}
\value{
An invisible logical value indicating success with \code{TRUE} and failure
 with \code{FALSE}.
}
\description{
De-registers/deconnects a SQLite weather database from the package
}
