% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swWeatherGenerator.R
\name{dbW_fixWeather}
\alias{dbW_fixWeather}
\title{Fix weather data}
\usage{
dbW_fixWeather(
  weatherData,
  subData = NULL,
  new_startYear = NULL,
  new_endYear = NULL,
  nmax_interp = 7L,
  precip_lt_nmax = 0,
  elevation = NA,
  correctWeatherValues = FALSE,
  fillMissingValues = TRUE,
  return_weatherDF = FALSE
)
}
\arguments{
\item{weatherData}{A list of elements of class \code{\link{swWeatherData}}
or a data frame as returned by \code{\link[=dbW_weatherData_to_dataframe]{dbW_weatherData_to_dataframe()}}.}

\item{subData}{A weather data object.}

\item{new_startYear}{An integer value. The first Calendar year of the new
time period. If \code{NULL}, then the first year of \code{weatherData}.}

\item{new_endYear}{An integer value. The last Calendar year of the new
time period. If \code{NULL}, then the last year of \code{weatherData}.}

\item{nmax_interp}{An integer value. Maximum spell length of missing values
for which linear interpolation is applied.}

\item{precip_lt_nmax}{A numeric value. Should short spells of
missing precipitation values be
linearly interpolated (if \code{NA}),
substituted with values from \code{subData} (if \code{Inf}), or
replaced by a fixed numeric value (default is 0)?}

\item{elevation}{A numeric value. Site elevation above sea level \verb{[m]}.
Used only if specific humidity is provided as input
for calculating relative humidity.}

\item{correctWeatherValues}{A logical value.
Apply all available corrections to weather values (see details).}

\item{fillMissingValues}{A logical value.
Fill in missing values (see details).}

\item{return_weatherDF}{A logical value. See section "Value".}
}
\value{
A list with two named elements
\itemize{
\item \code{"weatherData"}: An updated copy of the input \code{weatherData}
where missing values have been replaced and/or values corrected.
If \code{return_weatherDF} is \code{TRUE}, then the result is converted to a
data frame where columns represent weather variables.
If \code{return_weatherDF} is \code{FALSE}, then the result is
a list of elements of class \code{\link{swWeatherData}}.
\item \code{"meta"}: a data frame with the same dimensions as \code{"weatherData"}
with tags indicating which approach was used to replaced missing values
in corresponding cells of \code{weatherData} and/or values corrected
(see section \code{Details}).
}
}
\description{
This function applies the following 1-3 steps to the input weather data:
\enumerate{
\item \code{weatherData} is formatted for \code{rSOILWAT2}, i.e., converted to a
Gregorian calendar and required but missing variables added.
\item If \code{correctWeatherValues} is requested, then apply all available
corrections to weather values (meta data tag \code{"correctedValue"})
including (see \code{SOILWAT2} \code{"weathsetup.in"})
* Swap min/max values if min > max
(air temperature, relative humidity).
* Reset percentages to 100\% if they are larger than 100\%
(relative humidity, cloud cover).
* Reset observed solar radiation to extraterrestrial radiation if
the observed value is larger than expected.
\item If \code{fillMissingValues} is requested, then missing values are \code{"fixed"}
a) Short spells of missing values
(consecutive days shorter than \code{nmax_interp}) are linearly
interpolated from adjacent non-missing values
(meta data tag \code{"interpolateLinear (<= X days)"})
b) Short spells of missing precipitation values are
\itemize{
\item Linearly interpolated if \code{precip_lt_nmax} is \code{NA}
\item Set to a fixed numeric value of \code{precip_lt_nmax}
(meta data tag \code{"fixedValue"})
\item Substituted with values from \code{subData} if \code{precip_lt_nmax} is \code{Inf}
(see next point)
c) Values from a second weather data object \code{subData} are used to
replace (meta data tag \verb{substituteData"}):
\item Missing precipitation values (if \code{precip_lt_nmax} is \code{Inf})
\item Values before first day with any non-missing values
\item Variables absent in \code{weatherData} and present in \code{subData}
d) Long-term daily means are used to replace any remaining missing
values (meta data tag \code{"longTermDailyMean"});
for instance, this approach may be applied for
\item Values of variables that are present in \code{weatherData} and
absent in \code{subData}, before first day with any non-missing values
\item Values after end of available values in both \code{weatherData} and
\code{subData}
}
}
}
\examples{
# Example with missing values
x0 <- x <- dbW_weatherData_to_dataframe(rSOILWAT2::weatherData)

tmp <- x[, "Year"] == 1981
ids_to_interp <- tmp & x[, "DOY"] >= 144 & x[, "DOY"] <= 145
x[ids_to_interp, -(1:2)] <- NA

tmp <- x[, "Year"] == 1980
ids_to_sub <- tmp & x[, "DOY"] >= 153 & x[, "DOY"] <= 244
x[ids_to_sub, -(1:2)] <- NA

xf <- dbW_fixWeather(x, x0, return_weatherDF = TRUE)
all.equal(
  xf[["weatherData"]][!ids_to_interp, ],
  as.data.frame(x0)[!ids_to_interp, ]
)
table(xf[["meta"]])


# Example fixing weather values: switched min/max temperature
x <- x0

ids <- x[, "Year"] == 1981 & x[, "DOY"] >= 144 & x[, "DOY"] <= 145
x[ids, c("Tmin_C", "Tmax_C")] <- x[ids, c("Tmax_C", "Tmin_C")]

xf <- dbW_fixWeather(x, correctWeatherValues = TRUE, return_weatherDF = TRUE)
all.equal(xf[["weatherData"]], as.data.frame(x0))
table(xf[["meta"]])

}
\seealso{
\code{\link[=dbW_imputeWeather]{dbW_imputeWeather()}}, \code{\link[=dbW_substituteWeather]{dbW_substituteWeather()}},
\code{\link[=dbW_generateWeather]{dbW_generateWeather()}}
}
