% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_getIDs}
\alias{dbW_getIDs}
\title{Locate keys for weather database tables in the registered weather database}
\usage{
dbW_getIDs(
  site_id = NULL,
  site_label = NULL,
  long = NULL,
  lat = NULL,
  scenario = NULL,
  scenario_id = NULL,
  add_if_missing = FALSE,
  ignore.case = FALSE,
  verbose = FALSE
)
}
\value{
A list with two elements \code{site_id} and \code{scenario_id}.
}
\description{
Obtain database table keys 'Site_id' and 'Scenario_id' using alternative
information and optionally add missing sites and scenarios.
A site will be identified alternatively
by \itemize{
   \item its identification number \code{site_id},
   \item its name \code{site_label}, or
   \item its geographic location \code{long} and \code{lat}.
}
A scenario will be identified alternatively
by \itemize{
   \item its identification number \code{scenario_id}, or
   \item its name \code{scenario}.
}
}
