% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swFunctions.R
\name{dbW_getScenarioId}
\alias{dbW_getScenarioId}
\title{Extract table key to connect a scenario with weather data in the registered weather database}
\usage{
dbW_getScenarioId(Scenario, ignore.case = FALSE, verbose = FALSE)
}
\arguments{
\item{ignore.case}{A logical value.}
}
\value{
An integer value or \code{NULL}.
}
\description{
Extract table key to connect a scenario with weather data in the registered weather database
}
