% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_getScenariosTable}
\alias{dbW_getScenariosTable}
\title{Read entire table of Scenarios from the registered weather database}
\usage{
dbW_getScenariosTable()
}
\value{
A data.frame.
}
\description{
Read entire table of Scenarios from the registered weather database
}
