% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swFunctions.R
\name{dbW_getSiteId}
\alias{dbW_getSiteId}
\title{Extract table key to connect a site with weather data in the registered weather database}
\usage{
dbW_getSiteId(lat = NULL, long = NULL, Labels = NULL,
  ignore.case = FALSE, verbose = FALSE)
}
\arguments{
\item{lat}{A numeric value or \code{NULL}. The latitude in decimal degrees of WGS84.
Northern latitude are positive, sites on the southern hemisphere have negative values.}

\item{long}{A numeric value or \code{NULL}. The longitude in decimal degrees of WGS84.
Eastern longitudes are positive, sites on the western hemisphere have negative values.}

\item{Labels}{A vector of character strings. The names/labels of the queried sites.}

\item{ignore.case}{A logical value.}
}
\value{
An integer value or \code{NULL}.
}
\description{
Extract table key to connect a site with weather data in the registered weather database
}
\details{
The key (Site_id) can be located by either providing a \code{Label} or
by providing \code{lat} and \code{long} of the requested site.
}
