% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swFunctions.R
\name{dbW_getSiteTable}
\alias{dbW_getSiteTable}
\title{Read entire table of sites from the registered weather database}
\usage{
dbW_getSiteTable()
}
\value{
A data.frame.
}
\description{
Read entire table of sites from the registered weather database
}
