% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_getWeatherData}
\alias{dbW_getWeatherData}
\title{Extracts daily weather data from a registered weather database}
\usage{
dbW_getWeatherData(
  Site_id = NULL,
  lat = NULL,
  long = NULL,
  tol_xy = 1e-04,
  Label = NULL,
  startYear = NULL,
  endYear = NULL,
  Scenario = "Current",
  Scenario_id = NULL,
  ignore.case = FALSE,
  stop_if_missing = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{Site_id}{An integer value. The IDs/database key of the queried site.}

\item{lat}{A numeric vector or \code{NULL}. The latitude in decimal degrees
of \code{WGS84}. Northern latitude are positive, sites on the southern
hemisphere have negative values.}

\item{long}{A numeric vector or \code{NULL}. The longitude in decimal degrees
of \code{WGS84}. Eastern longitudes are positive, sites on the western
hemisphere have negative values.}

\item{tol_xy}{A numeric value. The tolerance used to match requested
longitude and latitude values.}

\item{Label}{A character string. The name/label of the queried site.}

\item{startYear}{A numeric value. First calendar year of the weather data.}

\item{endYear}{A numeric value. Last calendar year of the weather data.}

\item{Scenario}{A character string. The name/label of a climate scenario.}

\item{Scenario_id}{An integer value The ID/database key of the queried
scenario.}

\item{ignore.case}{A logical value.}

\item{stop_if_missing}{A logical value. If \code{TRUE}, then throws an
error if at least one requested weather data object is not available
in the current weather database. If \code{FALSE}, then returns \code{NULL}
for those requested site scenario combinations.}

\item{verbose}{A logical value.}
}
\value{
If one site and one scenario were requested, then returns
weather data as list. Each element is an object of class
\code{\linkS4class{swWeatherData}} and contains data for one year.
If more than one site or more than scenario were requested, then returns
a list of such weather data lists.
Elements of the returned list may be \code{NULL} if there is no
weather data object for the requested site scenario combination and if
\code{stop_if_missing} is \code{FALSE}.
}
\description{
Weather data for the soil water simulation run can be stored in the input
data or it can be separate to keep the input data size down for multiple
variations of the same site. This function is used to return the weather
data from a predefined weather database. Using the database was faster then
reading in multiple weather files from disk.
}
\details{
If there is missing data, then impute or use the built-in Markov
weather generator (see examples for \code{\link{sw_exec}}).
}
\section{Notes}{

This function returns the first record of weather data for a
site x scenario combination even if duplicate entries match the query.
}

\seealso{
\code{\link{getWeatherData_folders}}
}
