% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_IsValid}
\alias{dbW_IsValid}
\title{Check whether registered weather database connection is valid}
\usage{
dbW_IsValid()
}
\value{
A logical value.
}
\description{
Check whether registered weather database connection is valid
}
