% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_setConnection}
\alias{dbW_setConnection}
\alias{.dbW_setConnection}
\title{Registers/connects a SQLite weather database with the package}
\usage{
dbW_setConnection(
  dbFilePath,
  create_if_missing = FALSE,
  check_version = FALSE,
  verbose = FALSE
)

.dbW_setConnection(dbFilePath)
}
\arguments{
\item{dbFilePath}{A character string. The file path of the weather database.
This will be a file of type \code{sqlite3}. In-memory databases are not
supported.}

\item{create_if_missing}{A logical value. If \code{TRUE} and file
\code{dbFilePath} does not exist then create a new database file.}

\item{check_version}{A logical value. If \code{TRUE} then check database
version against currently implemented version by the package.}

\item{verbose}{A logical value.}
}
\value{
An invisible logical value indicating success/failure.
}
\description{
Registers/connects a SQLite weather database with the package
}
\section{Details}{

\code{\link[=.dbW_setConnection]{.dbW_setConnection()}} is a bare-bones version of \code{\link[=dbW_setConnection]{dbW_setConnection()}}.
It doesn't carry out any checks that make sure the database works
correctly.
}

