% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swWeatherGenerator.R
\name{dbW_substituteWeather}
\alias{dbW_substituteWeather}
\title{Replace missing values with values from another weather data set}
\usage{
dbW_substituteWeather(
  weatherData,
  subData,
  vars_substitute = NULL,
  by = NULL,
  by_weatherData = by,
  by_subData = by,
  return_weatherDF = FALSE
)
}
\arguments{
\item{weatherData}{A list of elements of class \code{\link{swWeatherData}}
or a data frame as returned by \code{\link[=dbW_weatherData_to_dataframe]{dbW_weatherData_to_dataframe()}}.}

\item{subData}{A weather data object.}

\item{vars_substitute}{Names of variables for which missing values
in \code{weatherData} should be replaced by values from \code{subData}.
If \code{NULL}, then all weather variables (i.e., \code{\link[=weather_dataColumns]{weather_dataColumns()}}).}

\item{by}{Names of variables used to match days (rows) between
\code{weatherData} and \code{subData}. If \code{NULL}, then all variables occurring both
in \code{weatherData} and \code{subData} that are not weather variables.}

\item{by_weatherData}{See \code{by}.}

\item{by_subData}{See \code{by}.}

\item{return_weatherDF}{A logical value. See section "Value".}
}
\value{
An updated copy of \code{weatherData} where missing values have been
replaced by corresponding values from \code{subData}.
If \code{return_weatherDF} is \code{TRUE}, then the result is converted to a
data frame where columns represent weather variables.
If \code{return_weatherDF} is \code{FALSE}, then the result is
a list of elements of class \code{\link{swWeatherData}}.
}
\description{
Replace missing values with values from another weather data set
}
\examples{
# Load example data
path_demo <- system.file("extdata", "example1", package = "rSOILWAT2")
dif <- c(rep(TRUE, 3L), rep(FALSE, 11L))
dif[13L] <- TRUE # ACTUAL_VP
dif[14L] <- TRUE # SHORT_WR, desc_rsds = 2
wdata <- getWeatherData_folders(
  LookupWeatherFolder = file.path(path_demo, "Input"),
  weatherDirName = "data_weather_daymet",
  filebasename = "weath",
  startYear = 1980,
  endYear = 1981,
  dailyInputFlags = dif,
  method = "C"
)
x0 <- x <- dbW_weatherData_to_dataframe(wdata)
dif0 <- calc_dailyInputFlags(x0)

# Set June-August of 1980 as missing
ids_1980 <- x[, "Year"] == 1980
ids_missing <- ids_1980 & x[, "DOY"] >= 153 & x[, "DOY"] <= 244
x[ids_missing, -(1:2)] <- NA

# Substitute missing values
all.equal(
  dbW_substituteWeather(x, x0[ids_1980, ], return_weatherDF = TRUE),
  x0
)


}
\seealso{
\code{\link[=dbW_generateWeather]{dbW_generateWeather()}}, \code{\link[=dbW_imputeWeather]{dbW_imputeWeather()}}
}
