% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_upgrade.R
\name{dbW_upgrade}
\alias{dbW_upgrade}
\alias{dbW_upgrade_to_rSOILWAT2}
\alias{dbW_upgrade_v3to31}
\alias{dbW_upgrade_v2to3}
\alias{dbW_upgrade_v1to2}
\alias{backup_copy}
\title{Weather database upgrade functions}
\usage{
dbW_upgrade_to_rSOILWAT2(dbWeatherDataFile, fbackup = NULL,
  check_all = FALSE)

dbW_upgrade_v3to31(dbWeatherDataFile, fbackup = NULL, type_new = "gzip")

dbW_upgrade_v2to3(dbWeatherDataFile, fbackup = NULL)

dbW_upgrade_v1to2(dbWeatherDataFile, fbackup = NULL, SWRunInformation)

backup_copy(dbWeatherDataFile, fbackup = NULL)
}
\arguments{
\item{dbWeatherDataFile}{A character string. The path to the weather database file.}

\item{fbackup}{A character string. The path to where the weather database should be
backed up. If \code{NULL}, then '_copy' is appended to \code{dbWeatherDataFile}.}

\item{check_all}{A logical value. If \code{TRUE}, then every record is checked;
otherwise, only the first record is checked for the package version.}

\item{type_new}{The type of compression used to compress the weather blobs. See
\code{\link{memCompress}}.}
}
\description{
Weather database upgrade functions
}
\section{Details}{
 \code{dbW_upgrade_to_rSOILWAT2} upgrades a weather database that was
created under'Rsoilwat31' to the current package version 'rSOILWAT2'

 \code{dbW_upgrade_v3to31} upgrades a weather database from version
 '3.0.x' to '3.1.0'

 \code{dbW_upgrade_v2to3} upgrades a weather database from version
 '2.x.y' to '3.0.0'

 \code{dbW_upgrade_v1to2} upgrades a weather database from version
 '1.x.y' to '2.0.0'

 \code{backup_copy} creates a backup copy of a weather database file
 if not already present
}

