% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_version}
\alias{dbW_version}
\title{Query version number of registered weather database}
\usage{
dbW_version()
}
\value{
A numeric version number.
}
\description{
Query version number of registered weather database
}
