% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_weatherData_to_monthly}
\alias{dbW_weatherData_to_monthly}
\title{Conversion: object of class \code{\linkS4class{swWeatherData}} to
matrix of monthly values (\var{mean Tmax}, \var{mean Tmin}, \var{sum PPT})}
\usage{
dbW_weatherData_to_monthly(dailySW, na.rm = FALSE, valNA = NULL)
}
\arguments{
\item{valNA}{The (numerical) value of missing weather data.
If \code{NULL}, then default values are interpreted as missing.}
}
\description{
Conversion: object of class \code{\linkS4class{swWeatherData}} to
matrix of monthly values (\var{mean Tmax}, \var{mean Tmin}, \var{sum PPT})
}
