% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Vegetation.R
\name{estimate_PotNatVeg_roots}
\alias{estimate_PotNatVeg_roots}
\title{Calculate rooting profile for a soil profile from \var{lookup} tables}
\usage{
estimate_PotNatVeg_roots(
  layers_depth,
  trco_type_by_veg = list(grass_C3 = "SchenkJackson2003_PCdry_grasses", grass_C4 =
    "SchenkJackson2003_PCdry_grasses", grass_annuals = "Jacksonetal1996_crops", shrub =
    "SchenkJackson2003_PCdry_shrubs", forb = "SchenkJackson2003_PCdry_forbs", tree =
    "Bradfordetal2014_LodgepolePine"),
  trco_adj_by_veg = list(grass_C3 = "positive", grass_C4 = "positive", grass_annuals =
    "positive", shrub = "positive", forb = "positive", tree = "positive"),
  fgrass_c3c4ann = c(grass_C3 = NA, grass_C4 = NA, grass_annuals = NA),
  trco_table = rSOILWAT2::sw2_trco_table
)
}
\arguments{
\item{layers_depth}{A numeric vector. Values describe
the lower soil layer depths [cm].}

\item{trco_type_by_veg}{A named list of character strings.
The rooting profiles, i.e., column names in the \var{lookup} table,
for each vegetation type. A \code{NA} indicates that a rooting profile
is not calculated for that vegetation type.
The values will be passed to argument \var{trco_type} of function
\code{\link{TranspCoeffByVegType}}.}

\item{trco_adj_by_veg}{A named list of character strings.
The type of adjustment from the full rooting profile
to the \code{layers_depth}.
The values will be passed to argument \var{adjustType} of function
\code{\link{TranspCoeffByVegType}}.}

\item{fgrass_c3c4ann}{A named, numeric vector of length 3.
Relative contribution [0-1] of the C3-grasses, C4-grasses, and
annuals functional groups. The sum of \code{fgrass_c3c4ann} is 1.}

\item{trco_table}{A named list with two elements. Default values
are taken from \code{\link{sw2_trco_table}}.}
}
\description{
Calculate rooting profile for a soil profile from \var{lookup} tables
}
\examples{
sw_in <- rSOILWAT2::sw_exampleData

clim <- calc_SiteClimate(weatherList = rSOILWAT2::weatherData)

veg_cover <- rSOILWAT2::estimate_PotNatVeg_composition(
  MAP_mm = 10 * clim[["MAP_cm"]],
  MAT_C = clim[["MAT_C"]],
  mean_monthly_ppt_mm = 10 * clim[["meanMonthlyPPTcm"]],
  mean_monthly_Temp_C = clim[["meanMonthlyTempC"]],
  dailyC4vars = clim[["dailyC4vars"]]
)

estimate_PotNatVeg_roots(
  layers_depth = c(5, 10, 20, 30, 40, 50, 100, 200),
  fgrass_c3c4ann = veg_cover[["Grasses"]]
)

}
\seealso{
\code{\link{TranspCoeffByVegType}}
}
