% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{format_timestamp}
\alias{format_timestamp}
\title{Format time stamp of an input or output object}
\usage{
format_timestamp(object)
}
\description{
Format time stamp of an input or output object
}
\seealso{
\code{\link{get_timestamp}}

export
}
