% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_OutputDerived_Functions.R
\name{get_derived_output}
\alias{get_derived_output}
\title{Calculate derived variables from output}
\arguments{
\item{x}{An object of class \code{\linkS4class{swOutput}}.}

\item{timestep}{A character string. One of the \pkg{rSOILWAT2} time steps.}

\item{keep_time}{A logical value. Include time information in the returned
object.}
}
\description{
Calculate derived variables from output
}
