% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_OutputDerived_Functions.R
\name{get_soiltemp_avg}
\alias{get_soiltemp_avg}
\title{Extract average soil temperature}
\usage{
get_soiltemp_avg(x, timestep = c("Day", "Week", "Month", "Year"))
}
\arguments{
\item{x}{An object of class \code{\linkS4class{swOutput}}.}

\item{timestep}{A character string. One of the \pkg{rSOILWAT2} time steps.}
}
\value{
A numeric matrix of soil temperature [C] for each time step
  at depth of each soil layer.
}
\description{
Extract average soil temperature
}
\examples{
sw_out <- sw_exec(inputData = rSOILWAT2::sw_exampleData)
get_soiltemp_avg(sw_out, "Month")

}
