% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_OutputDerived_Functions.R
\name{get_surfacetemp_avg}
\alias{get_surfacetemp_avg}
\title{Extract average soil surface temperature}
\usage{
get_surfacetemp_avg(x, timestep = c("Day", "Week", "Month", "Year"))
}
\arguments{
\item{x}{An object of class \code{\linkS4class{swOutput}}.}

\item{timestep}{A character string. One of the \pkg{rSOILWAT2} time steps.}
}
\value{
A numeric vector of soil surface temperature [C] for each time step.
}
\description{
Extract average soil surface temperature
}
\examples{
sw_out <- sw_exec(inputData = rSOILWAT2::sw_exampleData)
get_surfacetemp_avg(sw_out, "Month")

}
