% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{get_swCarbon}
\alias{get_swCarbon}
\title{get_swCarbon}
\usage{
get_swCarbon(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swCarbon}} or \code{\linkS4class{swInputData}}.}
}
\description{
get_swCarbon
}
\seealso{
\code{\linkS4class{swCarbon}} and \code{\linkS4class{swInputData}}
}
