% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{get_swFiles}
\alias{get_swFiles}
\title{\code{get_swFiles}}
\usage{
get_swFiles(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swFiles}} or
\code{\linkS4class{swInputData}}.}
}
\description{
\code{get_swFiles}
}
\seealso{
\code{\linkS4class{swFiles}} and \code{\linkS4class{swInputData}}
}
