% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{get_swMarkov}
\alias{get_swMarkov}
\title{\code{get_swMarkov}}
\usage{
get_swMarkov(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swMarkov}} or
\code{\linkS4class{swInputData}}.}
}
\description{
\code{get_swMarkov}
}
\seealso{
\code{\linkS4class{swMarkov}} and \code{\linkS4class{swInputData}}
}
