% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{get_swOUT}
\alias{get_swOUT}
\title{\code{get_swOUT}}
\usage{
get_swOUT(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swOUT}} or \code{\linkS4class{swInputData}}.}
}
\description{
\code{get_swOUT}
}
\seealso{
\code{\linkS4class{swOUT}} and \code{\linkS4class{swInputData}}
}
