% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{get_swProd}
\alias{get_swProd}
\title{\code{get_swProd}}
\usage{
get_swProd(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swProd}} or
\code{\linkS4class{swInputData}}.}
}
\description{
\code{get_swProd}
}
\section{Notes}{

The method for class \code{\linkS4class{swProd}} is deprecated,
see \code{\link{get_swProd,swProd-method}}; please use class
\code{\linkS4class{swProd2}} instead, see
\code{\link{get_swProd,swProd2-method}}.
}

\seealso{
\code{\linkS4class{swProd}} and \code{\linkS4class{swInputData}}
}
