% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{get_swSoils}
\alias{get_swSoils}
\title{\code{get_swSoils}}
\usage{
get_swSoils(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swSoils}} or \code{\linkS4class{swInputData}}.}
}
\description{
\code{get_swSoils}
}
\seealso{
\code{\linkS4class{swSoils}} and \code{\linkS4class{swInputData}}
}
