% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{get_swSpinup}
\alias{get_swSpinup}
\title{\code{get_swSpinup}}
\usage{
get_swSpinup(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swSpinup}} or
\code{\linkS4class{swInputData}}.}
}
\description{
\code{get_swSpinup}
}
\seealso{
\code{\linkS4class{swSpinup}} and \code{\linkS4class{swInputData}}
}
