% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{get_swWeatherData}
\alias{get_swWeatherData}
\title{get_swWeatherData}
\usage{
get_swWeatherData(object, year)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swWeatherData}} or \code{\linkS4class{swInputData}}.}
}
\description{
get_swWeatherData
}
\seealso{
\code{\linkS4class{swWeatherData}} and \code{\linkS4class{swInputData}}
}
