% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{get_swYears}
\alias{get_swYears}
\title{\code{get_swYears}}
\usage{
get_swYears(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swYears}} or
\code{\linkS4class{swInputData}}.}
}
\description{
\code{get_swYears}
}
\seealso{
\code{\linkS4class{swYears}} and \code{\linkS4class{swInputData}}
}
