% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{get_timestamp}
\alias{get_timestamp}
\title{\code{get_timestamp}}
\usage{
get_timestamp(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swInputData}} or
\code{\linkS4class{swOutput}}.}
}
\description{
\code{get_timestamp}
}
