% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{get_version}
\alias{get_version}
\alias{get_version,ANY-method}
\title{Retrieve version of \pkg{rSOILWAT2} that was used to create object}
\usage{
get_version(object)

\S4method{get_version}{ANY}(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swInputData}} or
\code{\linkS4class{swOutput}}.}
}
\value{
A character string representing the version number (or \code{NA}).
}
\description{
Retrieve version of \pkg{rSOILWAT2} that was used to create object
}
\examples{
get_version(rSOILWAT2::sw_exampleData)
get_version(sw_exec(rSOILWAT2::sw_exampleData))

}
\seealso{
\code{\link{check_version}}
}
