% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{get_WeatherHistory}
\alias{get_WeatherHistory}
\title{\code{get_WeatherHistory}}
\usage{
get_WeatherHistory(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swInputData}}.}
}
\description{
\code{get_WeatherHistory}
}
\seealso{
\code{\linkS4class{swInputData}}
}
