% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{get_years_from_weatherData}
\alias{get_years_from_weatherData}
\title{Extract years from a \var{weatherData} object}
\usage{
get_years_from_weatherData(wd)
}
\arguments{
\item{wd}{A list of elements of class \code{\link[=swWeatherData-class]{swWeatherData}}
that each hold daily weather data for one calendar year.}
}
\description{
Extract years from a \var{weatherData} object
}
