% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Time_SimulationWorld.R
\name{getStartYear}
\alias{getStartYear}
\title{Calculate start year of simulation period after \var{spin-up}}
\usage{
getStartYear(simstartyr, spinup_N = 1L)
}
\description{
Calculate start year of simulation period after \var{spin-up}
}
