% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsw.R
\name{has_soilTemp_failed}
\alias{has_soilTemp_failed}
\title{Check if soil temperature simulation failed}
\usage{
has_soilTemp_failed()
}
\value{
A logical value
}
\description{
Check if soil temperature simulation failed
}
