% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swWeatherGenerator.R
\name{impute_df}
\alias{impute_df}
\title{Imputes missing values in a \code{data.frame}}
\usage{
impute_df(
  x,
  imputation_type = c("none", "mean", "locf"),
  imputation_span = 5L,
  cyclic = FALSE
)
}
\arguments{
\item{x}{A \code{\link{data.frame}} with numerical columns. Imputation
works on each column separately.}

\item{imputation_type}{A character string describing the imputation method;
 currently, one of three values: \describe{
   \item{\var{"none"}}{no imputation is carried out;}
   \item{\var{"mean"}}{missing values will be replaced by the average
     of \code{imputation_span} non-missing values before and
     \code{imputation_span} non-missing values after; note:
     this may fail if there are less than \code{2 * imputation_span}
     non-missing values;}
   \item{\var{"locf"}}{missing values will be replaced by the
     \var{"last-observation-carried-forward"} imputation method.}
}}

\item{imputation_span}{An integer value. The number of non-missing values
considered if \code{imputation_type = "mean"}.}

\item{cyclic}{A logical value. If \code{TRUE}, then the last row of \code{x}
is considered to be a direct neighbor of the first row, e.g., rows of
\code{x} represent day of year for an average year.}
}
\value{
An updated version of \code{x}.
}
\description{
Imputes missing values in a \code{data.frame}
}
