% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inputData}
\alias{inputData}
\title{Input data object for a site}
\format{An object of class \code{\linkS4class{swInputData}}. Values correspond to
  \code{example1}; see example code.}
\usage{
inputData
}
\description{
A dataset containing complete input data for an unspecified location.
}
\examples{
data(inputData)
path_demo <- system.file("extdata", "example1", package = "rSOILWAT2")
x <- sw_inputDataFromFiles(dir = path_demo, files.in = "files.in")
identical(x, inputData) ## TRUE
}
\keyword{datasets}
