% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{is_missing_weather}
\alias{is_missing_weather}
\title{Check which weather values are missing}
\usage{
is_missing_weather(x)
}
\arguments{
\item{x}{A two-dimensional numeric object.}
}
\value{
A logical object with same dimensions as \code{x}
}
\description{
Check which weather values are missing
}
\examples{
x <- data.frame(
  Tmax = c(-1.5, 2, NA, 999),
  Tmin = c(-5, NaN, 999, -5)
)

is_missing_weather(x)

}
