% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Pedotransfer_Functions.R
\name{list_matched_swrcs_pdfs}
\alias{list_matched_swrcs_pdfs}
\title{Matching pairs of implemented \code{SWRCs} and \code{PDFs}}
\usage{
list_matched_swrcs_pdfs(swrc_name = names(swrc_names()))
}
\arguments{
\item{swrc_name}{An character string or vector.
The selected \code{SWRC} name
(one of \code{\link[=swrc_names]{swrc_names()}}, with default \code{"Campbell1974"}).}
}
\value{
A \code{data.frame} with two columns \code{SWRC} and \code{PDF} where each
row contains a matching pair of \code{SWRC} and \code{PDF} that are implemented.
}
\description{
Matching pairs of implemented \code{SWRCs} and \code{PDFs}
}
\examples{
# Data frame of SWRC-PDF combinations
df_swrc_pdfs <- rSOILWAT2::list_matched_swrcs_pdfs()

# List of SWRC-PDF combinations
list_swrcs_pdfs <- unname(as.list(as.data.frame(t(df_swrc_pdfs))))

# Available SWRC-PDF combinations
has_pdf <- check_pdf_availability(df_swrc_pdfs[, "PDF"])
df_swrc_pdfs[has_pdf, , drop = FALSE]
list_swrcs_pdfs[has_pdf]

}
