% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Pedotransfer_Functions.R
\name{list_matched_swrcs_ptfs}
\alias{list_matched_swrcs_ptfs}
\title{Matching pairs of implemented \code{SWRCs} and \code{PTFs}}
\usage{
list_matched_swrcs_ptfs(swrc_name = names(swrc_names()))
}
\arguments{
\item{swrc_name}{An character string or vector.
The selected \code{SWRC} name
(one of \code{\link[=swrc_names]{swrc_names()}}, with default \code{"Campbell1974"}).}
}
\value{
A \code{data.frame} with two columns \code{SWRC} and \code{PTF} where each
row contains a matching pair of \code{SWRC} and \code{PTF} that are implemented.
}
\description{
Matching pairs of implemented \code{SWRCs} and \code{PTFs}
}
\examples{
# Data frame of SWRC-PTF combinations
df_swrc_ptfs <- rSOILWAT2::list_matched_swrcs_ptfs()

# List of SWRC-PTF combinations
list_swrcs_ptfs <- unname(as.list(as.data.frame(t(df_swrc_ptfs))))

# Available SWRC-PTF combinations
has_ptf <- check_ptf_availability(df_swrc_ptfs[, "PTF"])
df_swrc_ptfs[has_ptf, , drop = FALSE]
list_swrcs_ptfs[has_ptf]

}
