% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Vegetation.R
\name{locate_intersections}
\alias{locate_intersections}
\title{Locate intersections between a spline function and a curve function}
\usage{
locate_intersections(psp, linfun, xlim, expected_nroots = 1)
}
\arguments{
\item{psp}{Spline function}

\item{linfun}{Linear function}

\item{xlim}{Search interval}

\item{expected_nroots}{Maximum number of roots,
i.e., the interval \code{xlim} is split into this number of
contiguous sections and each is checked for the presence of a root}
}
\description{
Locate intersections between a spline function and a curve function
}
