% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_SoilMTRegimes.R
\name{Maestas2016_Table1}
\alias{Maestas2016_Table1}
\title{Table 1 in Maestas et al. 2016}
\usage{
Maestas2016_Table1()
}
\description{
Table 1 in Maestas et al. 2016
}
\section{Details}{
 We need to make the following additional assumptions:
  \itemize{
    \item "Dry-Xeric" == "Xeric bordering on Aridic"
    \item "Weak-Aridic" == "Aridic bordering on Xeric"
  }
}

\references{
Maestas, J.D., Campbell, S.B., Chambers, J.C., Pellant, M. &
  Miller, R.F. (2016). Tapping Soil Survey Information for Rapid Assessment
  of Sagebrush Ecosystem Resilience and Resistance. Rangelands, 38, 120-128.
}
