% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Pedotransfer_Functions.R
\name{pdf_estimate}
\alias{pdf_estimate}
\title{Estimate \code{SWRC} parameters from soil texture with a pedotransfer function}
\usage{
pdf_estimate(sand, clay, fcoarse, swrc_type = 1L, pdf_type = 1L)
}
\arguments{
\item{sand}{A numeric value or vector.
Sand content of the matric soil component
(< 2 mm fraction; units of \verb{[g/g]}) of each soil layer.}

\item{clay}{A numeric value or vector.
Clay content of the matric soil component
(< 2 mm fraction; units of \verb{[g/g]}) of each soil layer.}

\item{fcoarse}{A numeric value or vector.
Coarse fragments, e.g., gravel, (> 2 mm; units of \verb{[m3/m3]})
relative to the whole soil of each soil layer.
\code{fcoarse} is required, for instance, to translate between
values relative to the matric soil component (< 2 mm fraction) and
relative to the whole soil (matric soil plus coarse fragments).}

\item{swrc_type}{An integer value or vector.
The identification number of selected \code{SWRC}.}

\item{pdf_type}{An integer value or vector.
The identification number of selected \code{PDF}.}
}
\value{
\code{swrcp}, i.e,.
a numeric matrix where rows represent soil (layers) and
columns represent a fixed number of \code{SWRC} parameters.
The interpretation is dependent on the selected \code{SWRC}.
}
\description{
Estimate \code{SWRC} parameters from soil texture with a pedotransfer function
}
\section{Details}{

Implemented PDFs (\code{pdf_type}):
\enumerate{
\item Cosby et al. 1984 PDF estimates parameters of Campbell 1974 SWRC
}


See \link{SWRCs} for suitable values of \code{swrc_type}.
}

\section{Notes}{

The soil parameters \code{sand}, \code{clay}, and \code{fcoarse} must be of
the same length, i.e., represent one soil (length 1) or
multiple soil (layers) (length > 1).
The arguments selecting \code{SWRC} (\code{swrc_type}) and \code{PDF} (\code{pdf_type})
are recycled for multiple soil layers.
}

\examples{
pdf_estimate(sand = c(0.5, 0.3), clay = c(0.2, 0.1), fcoarse = c(0, 0))

soils <- swSoils_Layers(rSOILWAT2::sw_exampleData)
pdf_estimate(
  sand = soils[, "sand_frac"],
  clay = soils[, "clay_frac"],
  fcoarse = soils[, "gravel_content"]
)

}
\references{
Cosby, B. J., G. M. Hornberger, R. B. Clapp, and T. R. Ginn.
1984. A statistical exploration of the relationships of soil moisture
characteristics to the physical properties of soils.
Water Resources Research, 20:682-690, \doi{10.1029/WR020i006p00682}
}
