% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Pedotransfer_Functions.R
\name{pdf_names}
\alias{pdf_names}
\title{List Pedotransfer Functions \code{PDFs}}
\usage{
pdf_names()
}
\value{
An integer vector with names of implemented \code{PDFs}
}
\description{
List Pedotransfer Functions \code{PDFs}
}
\details{
Notes:
The integer values may change with new versions of \code{SOILWAT2.}
}
\seealso{
\code{\link{SWRCs}}, \code{\link[=swrc_names]{swrc_names()}}
}
