% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Pedotransfer_Functions.R
\name{pdf_Rosetta_for_vanGenuchten1980}
\alias{pdf_Rosetta_for_vanGenuchten1980}
\title{Estimate van Genuchten 1980 \code{SWRC} parameters using \code{Rosetta} live \code{API}}
\usage{
pdf_Rosetta_for_vanGenuchten1980(
  sand,
  clay,
  bdensity = NULL,
  version = c("3", "1", "2"),
  verbose = interactive(),
  ...
)
}
\arguments{
\item{sand}{A numeric value or vector.
Sand content of the matric soil component
(< 2 mm fraction; units of \verb{[g/g]}) of each soil layer.}

\item{clay}{A numeric value or vector.
Clay content of the matric soil component
(< 2 mm fraction; units of \verb{[g/g]}) of each soil layer.}

\item{bdensity}{A numeric value or vector.
Density of the whole soil
(matric soil plus coarse fragments; units \verb{[g/cm3]}).}

\item{version}{A character string that selects a \code{Rosetta} version.}

\item{verbose}{A logical value. If \code{TRUE}, then display
\code{SOILWAT2} internal warnings and other messages.}

\item{...}{Additional function arguments passed on or ignored.}
}
\value{
\code{swrcp}, i.e,.
a numeric matrix where rows represent soil (layers) and
columns represent a fixed number of \code{SWRC} parameters: \itemize{
\item \code{swrcp[0]} (\code{theta_r}): residual volumetric water content
of the matric component (units of \verb{[cm / cm]})
\item \code{swrcp[1]} (\code{theta_s}): saturated volumetric water content
of the matric component (units of \verb{[cm / cm]})
\item \code{swrcp[2]} (\code{alpha}): related to the inverse of
air entry suction (units of \verb{[cm-1]})
\item \code{swrcp[3]} (\code{n}): measure of the pore-size distribution \verb{[-]}
\item \code{swrcp[4]} (\code{K_sat}): saturated hydraulic conductivity \verb{[cm / day]}
}
}
\description{
Estimate van Genuchten 1980 \code{SWRC} parameters using \code{Rosetta} live \code{API}
}
\section{Details}{

\code{\link[=pdf_estimate]{pdf_estimate()}} is the function that should be directly called; this here
is an internal helper function.
}

\section{Notes}{

This function calls \code{\link[soilDB:ROSETTA]{soilDB::ROSETTA()}} and
a live internet connection is required to access \code{Rosetta}.
}

\references{
Mualem, Y. 1976. A new model for predicting the hydraulic conductivity of
unsaturated porous media.
Water Resources Research, 12:513-522, \doi{10.1029/WR012i003p00513}

van Genuchten, M. T. 1980. A Closed-form Equation for Predicting the
Hydraulic Conductivity of Unsaturated Soils.
Soil Science Society of America Journal, 44:892-898,
\doi{10.2136/sssaj1980.03615995004400050002x}

Zhang, Y., & Schaap, M. G. 2017. Weighted recalibration of the
Rosetta pedotransfer model with improved estimates of
hydraulic parameter distributions and summary statistics (Rosetta3).
Journal of Hydrology, 547:39-53, \doi{10.1016/j.jhydrol.2017.01.004}
}
\seealso{
\code{\link[soilDB:ROSETTA]{soilDB::ROSETTA()}}
}
