% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Pedotransfer_Functions.R
\name{pdfs_implemented_by_rSW2}
\alias{pdfs_implemented_by_rSW2}
\title{List PDFs implemented by \code{rSOILWAT2}}
\usage{
pdfs_implemented_by_rSW2()
}
\description{
List PDFs implemented by \code{rSOILWAT2}
}
