% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Pedotransfer_Functions.R
\name{pdfs_implemented_by_SW2}
\alias{pdfs_implemented_by_SW2}
\title{List PDFs implemented by \code{SOILWAT2}}
\usage{
pdfs_implemented_by_SW2(names_only = FALSE)
}
\arguments{
\item{names_only}{A logical value, see \code{return} value.}
}
\value{
An named integer vector (if not \code{names_only})
with or a character vector (if \code{names_only}) names of implemented \code{PDFs}.
}
\description{
List PDFs implemented by \code{SOILWAT2}
}
