% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Pedotransfer_Functions.R
\name{pedotransfer}
\alias{pedotransfer}
\alias{pdf_to_vwc}
\alias{pdf_to_swp}
\alias{SWPtoVWC}
\alias{VWCtoSWP}
\title{Deprecated pedotransfer functions to convert between soil moisture
(volumetric water content, \var{VWC}) and soil water potential (\var{SWP})}
\usage{
pdf_to_vwc(
  swp,
  sand,
  clay,
  thetas,
  psis,
  b,
  MPa_toBar = -10,
  bar_conversion = 1024
)

pdf_to_swp(
  vwc,
  sand,
  clay,
  thetas,
  psis,
  b,
  bar_toMPa = -0.1,
  bar_conversion = 1024
)

SWPtoVWC(swp, sand, clay, ...)

VWCtoSWP(vwc, sand, clay, ...)
}
\arguments{
\item{swp}{A numeric value, vector, or 2-dimensional object
(matrix or data.frame). The soil water potential (of the soil matrix) in
units of \var{MPa}, i.e., the soil without the volume of rock and gravel.}

\item{sand}{A numeric value or vector. Sand content of the soil layer(s) as
fractional value in \code{[0,1]}.}

\item{clay}{A numeric value or vector. Clay content of the soil layer(s) as
fractional value in \code{[0,1]}.}

\item{thetas}{Soon obsolete ... (see \code{feature_swrc})}

\item{psis}{Soon obsolete ... (see \code{feature_swrc})}

\item{b}{Soon obsolete ... (see \code{feature_swrc})}

\item{MPa_toBar}{Soon obsolete ... (see \code{feature_swrc})}

\item{bar_conversion}{Soon obsolete ... (see \code{feature_swrc})}

\item{vwc}{A numeric value, vector, or 2-dimensional object
(matrix or data.frame). The matric soil moisture, i.e., reduced by the
volume of rock and gravel.}

\item{bar_toMPa}{Soon obsolete ... (see \code{feature_swrc})}

\item{...}{Additional arguments.}
}
\value{
Volumetric water content in units of m^3 (of water) / m^3 (of soil)
 \code{[0, 1]}. There are six use cases:\enumerate{
   \item 1) \itemize{
     \item Input: \code{SWP} [single value]; \code{sand} and \code{clay}
       [single values]
     \item Output: \code{VWC} [single value]}
   \item 2) \itemize{
     \item Input: \code{SWP} [single value]; \code{sand} and \code{clay}
       [vectors of length d]
     \item Output: \code{VWC} [vector of length d]}
   \item 3) \itemize{
     \item Input: \code{SWP} [vector of length l]; \code{sand} and
       \code{clay} infraction [single values]
     \item Output: \code{VWC} [vector of length l]}
   \item 4) \itemize{
     \item Input: \code{SWP} [vector of length l]; \code{sand} and
       \code{clay} [vectors of length d]
     \item Output: \code{VWC} [l x d matrix] where \code{SWP} is
       repeated for each column}
   \item 5) \itemize{
     \item Input: \code{SWP} [l x d matrix]; \code{sand} and \code{clay}
       [single values]
     \item Output: \code{VWC} [l x d matrix]}
   \item 6) \itemize{
     \item Input: \code{SWP} [l x d matrix]; \code{sand} and \code{clay}
       [vectors of length d]
     \item Output: \code{VWC} [l x d matrix], \code{sand} and \code{clay}
       vectors are repeated for each row}
 }

Soil water potential in units of \var{MPa} \code{[-Inf, 0]}.
  There are six use cases: \enumerate{
   \item 1) \itemize{
     \item Input: \code{VWC} [single value]; \code{sand} and \code{clay}
       [single values]
     \item Output: \code{SWP} [single value]}
   \item 2) \itemize{
     \item Input: \code{VWC} [single value]; \code{sand} and \code{clay}
       [vectors of length d]
     \item Output: \code{SWP} [vector of length d]}
   \item 3) \itemize{
     \item Input: \code{VWC} [vector of length l]; \code{sand} and
       \code{clay} in fraction [single values]
     \item Output: \code{SWP} [vector of length l]}
   \item 4) \itemize{
     \item Input: \code{VWC} [vector of length l]; \code{sand} and
       \code{clay} [vectors of length d]
     \item Output: \code{SWP} [l x d matrix] where \code{VWC} is repeated for
       each column}
   \item 5) \itemize{
     \item Input: \code{VWC} [l x d matrix]; \code{sand} and \code{clay}
       [single values]
     \item Output: \code{SWP} [l x d matrix]}
   \item 6) \itemize{
     \item Input: \code{VWC} [l x d matrix]; \code{sand} and \code{clay}
       [vectors of length d]
     \item Output: \code{SWP} [l x d matrix], \code{sand} and \code{clay}
       vectors are repeated for each row}
 }
}
\description{
Deprecated pedotransfer functions to convert between soil moisture
(volumetric water content, \var{VWC}) and soil water potential (\var{SWP})

Calculate volumetric water content from soil water potential and soil texture

Calculate soil water potential from volumetric water content and soil texture
}
\section{Note}{
 either \code{swp} or \code{sand}/\code{clay} needs be a
  single value

 either \code{vwc} or \code{sand}/\code{clay} needs be a
  single value
}

\references{
Cosby, B. J., G. M. Hornberger, R. B. Clapp, and T. R. Ginn.
1984. A statistical exploration of the relationships of soil moisture
characteristics to the physical properties of soils. Water Resources Research
20:682-690.
}
