% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Vegetation.R
\name{predict_season}
\alias{predict_season}
\title{Predict seasonal phenology}
\usage{
predict_season(x, ref_season, target_season)
}
\arguments{
\item{x}{A two-dimensional object. Rows correspond to month of the year,
i.e., there must be 12 rows, and columns correspond to variables that are
adjusted.}

\item{ref_season}{A numeric vector. The season as months (integers in 1-12)
for which the values of \code{x} serve as reference (used for fitting).}

\item{target_season}{A numeric vector. The season as months (1-12)
for which the prediction is made.}
}
\value{
A two-dimensional object with rows that correspond to
  \code{target_season} and columns that correspond to the adjusted variable
  values from \code{x}.
}
\description{
Fits a loess-curve to a seasonal subset of monthly values and uses the fit
to make a prediction for a different subset of months.
}
\section{Details}{
 You may want to eliminate negative values that may arise
  from the fitting/prediction process.
}

\examples{
sw_in <- rSOILWAT2::sw_exampleData
biomass_reference <- swProd_MonProd_grass(sw_in)[, 2:3]
biomass_adj <- biomass_reference
biomass_adj[] <- NA

## Adjust phenology from a reference Nov-Feb to a target Jul-Oct
## non-growing season
ref_season <- c(11:12, 1:2) # November-February
target_season <- 7:10 # July-October
biomass_adj[target_season, ] <- predict_season(biomass_reference,
  ref_season, target_season)

## Adjust phenology from a reference Mar-Oct to a target Nov-Jun
## growing season
ref_season <- 3:10 # March-October
target_season <- c(11:12, 1:6) # November-June
biomass_adj[target_season, ] <- predict_season(biomass_reference,
  ref_season, target_season)

## Plot reference and adjusted monthly values
\dontrun{
par_prev <- par(mfrow = c(2, 1))

plot(1:12, biomass_reference[, 1], type = "l", col = "red",
  ylim = c(0, 250), xlab = "Month")
lines(1:12, biomass_adj[, 1])

plot(1:12, biomass_reference[, 2], type = "l", col = "red",
  ylim = c(0, 1), xlab = "Month")
lines(1:12, biomass_adj[, 2])

par(par_prev)
}

}
