% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Vegetation.R
\name{prepare_TranspirationRegions}
\alias{prepare_TranspirationRegions}
\title{Prepare transpiration regions}
\usage{
prepare_TranspirationRegions(tr_lyrs)
}
\arguments{
\item{tr_lyrs}{An integer vector.
The transpiration region for each soil layer.}
}
\value{
A transpiration region matrix.
}
\description{
Translate a soil vector of transpiration region values into a
transpiration region matrix.
}
\examples{
# Example values correspond to `CONUSSOIL_BSE_EVERY10cm` of \pkg{rSFSW2}:
prepare_TranspirationRegions(tr_lyrs = c(1, 1, 1, 2, 2, 2, 2, 3, 3, 3))

}
\seealso{
\code{\link{adjust_TranspirationRegions}} and
  \code{\link{check_TranspirationRegions}}
}
