% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Pedotransfer_Functions.R
\name{ptf_estimate}
\alias{ptf_estimate}
\title{Estimate \code{SWRC} parameters from soil texture with a pedotransfer function}
\usage{
ptf_estimate(
  sand,
  clay,
  fcoarse,
  bdensity = NULL,
  swrc_name,
  ptf_name,
  fail = FALSE,
  ...
)
}
\arguments{
\item{sand}{A numeric value or vector.
Sand content of the matric soil component
(< 2 mm fraction; units of \verb{[g/g]}) of each soil layer.}

\item{clay}{A numeric value or vector.
Clay content of the matric soil component
(< 2 mm fraction; units of \verb{[g/g]}) of each soil layer.}

\item{fcoarse}{A numeric value or vector.
Coarse fragments, e.g., gravel, (> 2 mm; units of \verb{[m3/m3]})
relative to the whole soil of each soil layer.
\code{fcoarse} is required, for instance, to translate between
values relative to the matric soil component (< 2 mm fraction) and
relative to the whole soil (matric soil plus coarse fragments).}

\item{bdensity}{A numeric value or vector.
Density of the whole soil
(matric soil plus coarse fragments; units \verb{[g/cm3]}).}

\item{swrc_name}{An character string or vector.
The selected \code{SWRC} name
(one of \code{\link[=swrc_names]{swrc_names()}}, with default \code{"Campbell1974"}).}

\item{ptf_name}{An character string or vector.
The selected \code{PTF} name
(one of \code{\link[=ptf_names]{ptf_names()}}, with default \code{"Cosby1984AndOthers"}).}

\item{fail}{A logical value.
Issue a warning (\code{FALSE}, default) or throw an error (\code{TRUE})
if request fails.}

\item{...}{Additional parameters passed to selected \code{PTF} function.}
}
\value{
\code{swrcp}, i.e,.
a numeric matrix where rows represent soil (layers) and
columns represent a fixed number of \code{SWRC} parameters.
The interpretation is dependent on the selected \code{SWRC}, see
\code{SOILWAT2} input file \code{swrc_param.in}
(
\code{system.file("extdata", "example1", "Input", "swrc_params.in", package = "rSOILWAT2")}
).
}
\description{
Estimate \code{SWRC} parameters from soil texture with a pedotransfer function
}
\section{Notes}{

\code{\link[=swrc_names]{swrc_names()}} lists implemented \code{SWRCs};
\code{\link[=ptf_names]{ptf_names()}} lists implemented \code{PTFs}; and
\code{\link[=check_ptf_availability]{check_ptf_availability()}} checks availability of \code{PTFs}.


The soil parameters \code{sand}, \code{clay}, \code{fcoarse}, and \code{bdensity} must be of
the same length, i.e., represent one soil (length 1) or
multiple soil (layers) (length > 1); however, \code{bdensity} may be \code{NULL}.
The arguments selecting \code{SWRC} (\code{swrc_name}) and \code{PTF} (\code{ptf_name})
are recycled for multiple soil layers.
}

\examples{
ptf_estimate(sand = c(0.5, 0.3), clay = c(0.2, 0.1), fcoarse = c(0, 0))

soils <- swSoils_Layers(rSOILWAT2::sw_exampleData)

# Use PTF "Cosby1984" to estimate parameters of SWRC "Campbell1974"
ptf_estimate(
  sand = soils[, "sand_frac"],
  clay = soils[, "clay_frac"],
  fcoarse = soils[, "gravel_content"],
  swrc_name = "Campbell1974",
  ptf_name = "Cosby1984"
)

# Use PTF "Rosetta3" to estimate parameters of SWRC "vanGenuchten1980"
if (check_ptf_availability("Rosetta3")) {
  ptf_estimate(
    sand = soils[, "sand_frac"],
    clay = soils[, "clay_frac"],
    fcoarse = soils[, "gravel_content"],
    bdensity = soils[, "bulkDensity_g/cm^3"],
    swrc_name = "vanGenuchten1980",
    ptf_name = "Rosetta3"
  )
}

# Use PTF "neuroFX2021" to estimate parameters of SWRC `FXW`
\dontrun{
# Set neuroFX2021 file path, see details in `ptf_neuroFX2021_for_FXW()`
options(RSW2_FILENEUROFX2021 = "path/to/sscbd.RData")
}

if (check_ptf_availability("neuroFX2021")) {
  ptf_estimate(
    sand = soils[, "sand_frac"],
    clay = soils[, "clay_frac"],
    fcoarse = soils[, "gravel_content"],
    bdensity = soils[, "bulkDensity_g/cm^3"],
    swrc_name = "FXW",
    ptf_name = "neuroFX2021"
  )
}

}
\references{
Cosby, B. J., G. M. Hornberger, R. B. Clapp, & T. R. Ginn. 1984.
A statistical exploration of the relationships of soil moisture
characteristics to the physical properties of soils.
Water Resources Research, 20:682-690, \doi{10.1029/WR020i006p00682}
}
