% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Pedotransfer_Functions.R
\name{ptf_names}
\alias{ptf_names}
\title{List Pedotransfer Functions \code{PTFs}}
\usage{
ptf_names()
}
\value{
An named integer vector with names of implemented \code{PTFs}
}
\description{
List Pedotransfer Functions \code{PTFs}
}
\details{
Notes:
The integer values may change with new versions of \code{SOILWAT2.}
}
\seealso{
\code{\link{SWRCs}}, \code{\link[=swrc_names]{swrc_names()}}, \code{\link[=check_ptf_availability]{check_ptf_availability()}}
}
