% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Pedotransfer_Functions.R
\name{ptf_neuroFX2021_for_FXW}
\alias{ptf_neuroFX2021_for_FXW}
\title{Estimate \code{FXW} \code{SWRC} parameters using \code{neuroFX2021}}
\usage{
ptf_neuroFX2021_for_FXW(
  sand,
  clay,
  bdensity = NULL,
  file_neuroFX2021 = getOption("RSW2_FILENEUROFX2021", NULL),
  ...
)
}
\arguments{
\item{sand}{A numeric value or vector.
Sand content of the matric soil component
(< 2 mm fraction; units of \verb{[g/g]}) of each soil layer.}

\item{clay}{A numeric value or vector.
Clay content of the matric soil component
(< 2 mm fraction; units of \verb{[g/g]}) of each soil layer.}

\item{bdensity}{A numeric value or vector.
Density of the whole soil
(matric soil plus coarse fragments; units \verb{[g/cm3]}).}

\item{file_neuroFX2021}{A character string that contains the file name with
full path of the \code{neuroFX2021} R object provided by Rudiyanto et al. 2021;
The path to the appropriate file can be set per R session
via option \code{"RSW2_FILENEUROFX2021"}, see additional details.}

\item{...}{Additional function arguments passed on or ignored.}
}
\value{
\code{swrcp}, i.e,.
a numeric matrix where rows represent soil (layers) and
columns represent a fixed number of \code{SWRC} parameters: \itemize{
\item \code{swrcp[0]} (\code{theta_s}): saturated volumetric water content
of the matric component (units of \verb{[cm / cm]})
\item \code{swrcp[1]} (\code{alpha}): shape parameter (units of \verb{[cm-1]})
\item \code{swrcp[2]} (\code{n}): shape parameter \verb{[-]}
\item \code{swrcp[3]} (\code{m}): shape parameter \verb{[-]}
\item \code{swrcp[4]} (\code{K_sat}): saturated hydraulic conductivity \verb{[cm / day]}
\item \code{swrcp[5]} (\code{L}): tortuosity/connectivity parameter \verb{[-]}
}
}
\description{
Estimate \code{FXW} \code{SWRC} parameters using \code{neuroFX2021}
}
\section{Details}{

\code{\link[=ptf_estimate]{ptf_estimate()}} is the function that should be directly called; this here
is an internal helper function.


This function requires that users download
the fitted \code{neuroFX2021} neural networks published by Rudiyanto et al. 2021
in Supplementary Material 1 (resulting in a local file named \code{xxx_mmc1.zip}).
This needs to be unzipped and the resulting \code{tar} file unpacked;
this produces a folder \verb{R code for neuroFX2021}.
This folder contains two R data files : \code{ssc.RData} and \code{sscbd.RData}.
The argument \code{file_neuroFX2021} is the file name (with path) to \code{sscbd.RData}
if soil density data are available and to \code{ssc.RData} otherwise
(see Rudiyanto et al. 2021).
The path to the appropriate file can be set per R session
via option \code{"RSW2_FILENEUROFX2021"}
(and avoid passing it directly as argument to the function);
this can be useful, for example, if \code{ptf_estimate()} is used for \code{FXW}.
}

\references{
Rudiyanto, Minasny, B., Chaney, N. W., Maggi, F., Goh Eng Giap, S.,
Shah, R. M., Fiantis, D., & Setiawan, B. I. 2021.
Pedotransfer functions for estimating soil hydraulic properties from
saturation to dryness.
Geoderma, 403:115194, \doi{10.1016/j.geoderma.2021.115194}

Fredlund, D. G., & Xing, A. 1994.
Equations for the soil-water characteristic curve.
Canadian Geotechnical Journal, 31: 512–532, \doi{10.1139/t94-061}

Wang, Y., Jin, M., & Deng, Z. 2018.
Alternative model for predicting soil hydraulic conductivity over
the complete moisture range.
Water Resources Research, 54:6860–6876, \doi{10.1029/2018WR023037}
}
