% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Pedotransfer_Functions.R
\name{ptfs_implemented_by_rSW2}
\alias{ptfs_implemented_by_rSW2}
\title{List \code{PTFs} implemented by \code{rSOILWAT2}}
\usage{
ptfs_implemented_by_rSW2()
}
\description{
List \code{PTFs} implemented by \code{rSOILWAT2}
}
