% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSOILWAT2-defunct.R
\name{rSOILWAT2-defunct}
\alias{rSOILWAT2-defunct}
\alias{SMR_names}
\alias{SMRq_names}
\alias{STR_names}
\alias{calc_RRs_Chambers2014}
\alias{calc_RRs_Maestas2016}
\alias{calc_SMTRs}
\alias{getStartYear}
\alias{setup_time_simulation_run}
\alias{simTiming_ForEachUsedTimeUnit}
\alias{update_requested_years}
\title{Defunct functions in package \pkg{rSOILWAT2}}
\usage{
SMR_names(...)

SMRq_names(...)

STR_names(...)

calc_RRs_Chambers2014(...)

calc_RRs_Maestas2016(...)

calc_SMTRs(...)

getStartYear(...)

setup_time_simulation_run(...)

simTiming_ForEachUsedTimeUnit(...)

update_requested_years(...)
}
\description{
Executing a defunct function will fail and tell you which function
replaces them.
}
