% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSOILWAT2_deprecated.R
\name{rSOILWAT2-deprecated}
\alias{rSOILWAT2-deprecated}
\alias{sw_outputData}
\alias{swProd_MonProd_grass}
\alias{swProd_MonProd_shrub}
\alias{swProd_MonProd_tree}
\alias{swProd_MonProd_forb}
\alias{swProd_MonProd_grass<-}
\alias{swProd_MonProd_shrub<-}
\alias{swProd_MonProd_tree<-}
\alias{swProd_MonProd_forb<-}
\alias{swProd_MonProd_grass,swProd-method}
\alias{swProd_MonProd_shrub,swProd-method}
\alias{swProd_MonProd_tree,swProd-method}
\alias{swProd_MonProd_forb,swProd-method}
\alias{swProd_MonProd_grass<-,swProd-method}
\alias{swProd_MonProd_shrub<-,swProd-method}
\alias{swProd_MonProd_tree<-,swProd-method}
\alias{swProd_MonProd_forb<-,swProd-method}
\alias{swProd_MonProd_grass,swInputData-method}
\alias{swProd_MonProd_shrub,swInputData-method}
\alias{swProd_MonProd_tree,swInputData-method}
\alias{swProd_MonProd_forb,swInputData-method}
\alias{swProd_MonProd_grass<-,swInputData-method}
\alias{swProd_MonProd_shrub<-,swInputData-method}
\alias{swProd_MonProd_tree<-,swInputData-method}
\alias{swProd_MonProd_forb<-,swInputData-method}
\title{Deprecated functions in package \pkg{rSOILWAT2}}
\usage{
sw_outputData(inputData)

swProd_MonProd_grass(object)

swProd_MonProd_shrub(object)

swProd_MonProd_tree(object)

swProd_MonProd_forb(object)

swProd_MonProd_grass(object) <- value

swProd_MonProd_shrub(object) <- value

swProd_MonProd_tree(object) <- value

swProd_MonProd_forb(object) <- value

\S4method{swProd_MonProd_grass}{swProd}(object)

\S4method{swProd_MonProd_shrub}{swProd}(object)

\S4method{swProd_MonProd_tree}{swProd}(object)

\S4method{swProd_MonProd_forb}{swProd}(object)

\S4method{swProd_MonProd_grass}{swProd}(object) <- value

\S4method{swProd_MonProd_shrub}{swProd}(object) <- value

\S4method{swProd_MonProd_tree}{swProd}(object) <- value

\S4method{swProd_MonProd_forb}{swProd}(object) <- value

\S4method{swProd_MonProd_grass}{swInputData}(object)

\S4method{swProd_MonProd_shrub}{swInputData}(object)

\S4method{swProd_MonProd_tree}{swInputData}(object)

\S4method{swProd_MonProd_forb}{swInputData}(object)

\S4method{swProd_MonProd_grass}{swInputData}(object) <- value

\S4method{swProd_MonProd_shrub}{swInputData}(object) <- value

\S4method{swProd_MonProd_tree}{swInputData}(object) <- value

\S4method{swProd_MonProd_forb}{swInputData}(object) <- value
}
\arguments{
\item{inputData}{an object of the \var{S4} class
\code{\linkS4class{swInputData}} which is generated from
\code{\link{sw_inputData}} or \code{\link{sw_inputDataFromFiles}}.}

\item{object}{An object of a \code{rSOILWAT2} class.}

\item{value}{A value to assign to a specific slot of the object.}
}
\description{
Executing a deprecated function will warn and tell you which function
replaces them.
}
