% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsoilwat31-package.R
\docType{package}
\name{Rsoilwat31}
\alias{Rsoilwat31}
\alias{Rsoilwat31-package}
\title{Package `Rsoilwat31': summary information}
\description{
The package `Rsoilwat' / 'Rsoilwat31' is a R implementation of SoilWat
Modeling Software.  The C code was converted into a library with
input/output functions to R.  SoilWat is a simulator of the dynamics of soil
water. Originally written in the 1900's by a highly trained team of Botany's
finest. It is now maintained by Dr. William Lauenroth's Lab and Dr. Daniel
Schlaepfer.  The input data structure is S4. Generic functions that
apply to the input container or individual object control setting/getting
input data. Output data is in the form of lists.
}
\section{Version}{
 The version level of the package is given by the command
\code{packageDescription("Rsoilwat31")}. The most recent version of the
package can be obtained from github at
\url{https://github.com/Burke-Lauenroth-Lab/Rsoilwat}
}
\references{
Bradford, J. B., D. R. Schlaepfer, and W. K. Lauenroth (2014)
Ecohydrology of adjacent sagebrush and lodgepole pine ecosystems: The
consequences of climate change and disturbance.  \emph{Ecosystems}
\bold{17}:590--605.

Schlaepfer, D. R., W. K. Lauenroth, and J. B. Bradford (2012)
Ecohydrological niche of sagebrush ecosystems.  \emph{Ecohydrology}
\bold{5}:453--466.

Parton, W.J. (1978).  \emph{Abiotic section of ELM}. In: Grassland
simulation model (ed. Innis, G.S.).  Springer New York, NY, pp. 31--53.

Sala, O.E., Lauenroth, W.K. & Parton, W.J. (1992) Long-term soil-water
dynamics in the shortgrass steppe.  \emph{Ecology} \bold{73}:1175--1181.
}
\seealso{
\itemize{ \item \code{\link{sw_exec}} for running a simulation
\item \code{\link{sw_inputData}} and \code{\link{sw_inputDataFromFiles}} for
data input \item \code{\link{dbW_getWeatherData}} and
\code{\link{getWeatherData_folders}} for weather data input }
}

