% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Pedotransfer_Functions.R
\name{rSW2_SWRC_PDF_estimate_parameters}
\alias{rSW2_SWRC_PDF_estimate_parameters}
\title{Estimate parameters of selected soil water retention curve (\code{SWRC})
using selected pedotransfer function (\code{PDF}) that are implemented in \code{R}}
\usage{
rSW2_SWRC_PDF_estimate_parameters(pdf_name, sand, clay, fcoarse, fail = TRUE)
}
\arguments{
\item{pdf_name}{An character string or vector.
The selected \code{PDF} name
(one of \code{\link[=pdf_names]{pdf_names()}}, with default \code{"Cosby1984AndOthers"}).}

\item{sand}{A numeric value or vector.
Sand content of the matric soil component
(< 2 mm fraction; units of \verb{[g/g]}) of each soil layer.}

\item{clay}{A numeric value or vector.
Clay content of the matric soil component
(< 2 mm fraction; units of \verb{[g/g]}) of each soil layer.}

\item{fcoarse}{A numeric value or vector.
Coarse fragments, e.g., gravel, (> 2 mm; units of \verb{[m3/m3]})
relative to the whole soil of each soil layer.
\code{fcoarse} is required, for instance, to translate between
values relative to the matric soil component (< 2 mm fraction) and
relative to the whole soil (matric soil plus coarse fragments).}

\item{fail}{A logical value. If \code{TRUE} (default) and
requested \code{PDF} is not implemented in \code{R}, then throw an error;
otherwise, return silently.}
}
\value{
\code{swrcp}, i.e,.
a numeric matrix where rows represent soil (layers) and
columns represent a fixed number of \code{SWRC} parameters.
The interpretation is dependent on the selected \code{SWRC}.
However, return value is \code{NULL}
only if \code{fail} is \code{FALSE} and requested \code{PDF} is not implemented in \code{R}.
}
\description{
Estimate parameters of selected soil water retention curve (\code{SWRC})
using selected pedotransfer function (\code{PDF}) that are implemented in \code{R}
}
\section{Details}{

\code{\link[=pdf_estimate]{pdf_estimate()}} is the function that should be directly called; this here
is an internal helper function.
}

\section{Notes}{

See \code{SWRC_PDF_estimate_parameters()} in \code{SOILWAT2} for \code{PDFs}
implemented in C.
}

\references{
Cosby, B. J., G. M. Hornberger, R. B. Clapp, & T. R. Ginn. 1984.
A statistical exploration of the relationships of soil moisture
characteristics to the physical properties of soils.
Water Resources Research, 20:682-690, \doi{10.1029/WR020i006p00682}
}
