% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Time_SimulationWorld.R
\name{seq_month_ofeach_day}
\alias{seq_month_ofeach_day}
\title{The sequence of month numbers for each day in the period from - to}
\usage{
seq_month_ofeach_day(
  from = list(year = 1900, month = 1, day = 1),
  to = list(year = 1900, month = 12, day = 31),
  tz = "UTC"
)
}
\description{
The sequence of month numbers for each day in the period from - to
}
\examples{
\dontrun{
 month1 <- function() as.POSIXlt(seq(from = ISOdate(1980, 1, 1, tz = "UTC"),
    to = ISOdate(2010, 12, 31, tz = "UTC"), by = "1 day"))$mon + 1
 month2 <- function() seq_month_ofeach_day(list(1980, 1, 1),
   list(2010, 12, 31), tz = "UTC")

   if (requireNamespace("microbenchmark", quietly = TRUE))
     # barely any difference
     microbenchmark::microbenchmark(month1(), month2())
 }
}
