% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{set_Markov<-}
\alias{set_Markov<-}
\title{set_Markov<-}
\usage{
set_Markov(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swMarkov}} or \code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
set_Markov<-
}
\seealso{
\code{\linkS4class{swMarkov}} and \code{\linkS4class{swInputData}}
}
