% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{set_Markov}
\alias{set_Markov}
\title{set_Markov
Need to define and export this generic method -- otherwise, \code{\link{set_Markov<-}} doesn't work.}
\usage{
set_Markov(object, value)
}
\description{
set_Markov
Need to define and export this generic method -- otherwise, \code{\link{set_Markov<-}} doesn't work.
}
