% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{set_missing_weather}
\alias{set_missing_weather}
\title{Deal with missing weather values: convert to NAs}
\usage{
set_missing_weather(data, valNA = NULL)
}
\arguments{
\item{data}{A numerical object.}

\item{valNA}{The (numerical) value of missing weather data.
If \code{NULL}, then default values are interpreted as missing.}
}
\value{
\code{data} where \pkg{SOILWAT2} missing values are converted to
  R-compatible \code{NA}s.
}
\description{
Missing weather values may be coded with \code{NA},
  with the corresponding \var{SOILWAT2} value (i.e.,
  \code{rSOILWAT2:::rSW2_glovars[["kSOILWAT2"]][["kNUM"]][["SW_MISSING"]]}),
  or with the value of the argument \code{valNA}.
}
