% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsw.R
\name{set_requested_flags}
\alias{set_requested_flags}
\title{Assign requested values to (scalar) input flags}
\usage{
set_requested_flags(swIn, tag, use, values, fun, reset = TRUE, default = NA)
}
\arguments{
\item{reset}{A logical value. If \code{TRUE}, then reset flags identified by \code{tag}
and turned off as identified by \code{use} to \code{default}. If \code{FALSE}, then
set  flags identified by \code{tag} and turned on as identified by \code{use} to
corresponding elements of \code{values}; other flags are not changed.}
}
\description{
Assign requested values to (scalar) input flags
}
