% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{set_swCarbon<-}
\alias{set_swCarbon<-}
\title{\code{set_swCarbon<-}}
\usage{
set_swCarbon(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swCarbon}} or
\code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
\code{set_swCarbon<-}
}
\seealso{
\code{\linkS4class{swCarbon}} and \code{\linkS4class{swInputData}}
}
