% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{set_swCarbon}
\alias{set_swCarbon}
\title{\code{set_swCarbon}
Need to define and export this generic method -- otherwise, \code{\link{set_swCarbon<-}} doesn't work.}
\usage{
set_swCarbon(object, value)
}
\description{
\code{set_swCarbon}
Need to define and export this generic method -- otherwise, \code{\link{set_swCarbon<-}} doesn't work.
}
