% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{set_swCloud<-}
\alias{set_swCloud<-}
\title{\code{set_swCloud<-}}
\usage{
set_swCloud(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swCloud}} or
\code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
\code{set_swCloud<-}
}
\seealso{
\code{\linkS4class{swCloud}} and \code{\linkS4class{swInputData}}
}
