% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{set_swCloud}
\alias{set_swCloud}
\title{\code{set_swCloud}
Need to define and export this generic method -- otherwise, \code{\link{set_swCloud<-}} doesn't work.}
\usage{
set_swCloud(object, value)
}
\description{
\code{set_swCloud}
Need to define and export this generic method -- otherwise, \code{\link{set_swCloud<-}} doesn't work.
}
