% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{set_swFiles<-}
\alias{set_swFiles<-}
\title{set_swFiles<-}
\usage{
set_swFiles(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swFiles}} or \code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
set_swFiles<-
}
\seealso{
\code{\linkS4class{swFiles}} and \code{\linkS4class{swInputData}}
}
