% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{set_swFiles}
\alias{set_swFiles}
\title{\code{set_swFiles}
Need to define and export this generic method -- otherwise, \code{\link{set_swFiles<-}} doesn't work.}
\usage{
set_swFiles(object, value)
}
\description{
\code{set_swFiles}
Need to define and export this generic method -- otherwise, \code{\link{set_swFiles<-}} doesn't work.
}
