% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{set_swMarkov<-}
\alias{set_swMarkov<-}
\title{\code{set_swMarkov<-}}
\usage{
set_swMarkov(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swMarkov}} or
\code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
\code{set_swMarkov<-}
}
\seealso{
\code{\linkS4class{swMarkov}} and \code{\linkS4class{swInputData}}
}
