% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{set_swOUT<-}
\alias{set_swOUT<-}
\title{\code{set_swOUT<-}}
\usage{
set_swOUT(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swOUT}} or
\code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
\code{set_swOUT<-}
}
\seealso{
\code{\linkS4class{swOUT}} and \code{\linkS4class{swInputData}}
}
