% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{set_swOUT}
\alias{set_swOUT}
\title{set_swOUT
Need to define and export this generic method -- otherwise, \code{\link{set_swOUT<-}} doesn't work.}
\usage{
set_swOUT(object, value)
}
\description{
set_swOUT
Need to define and export this generic method -- otherwise, \code{\link{set_swOUT<-}} doesn't work.
}
