% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{set_swSoils}
\alias{set_swSoils}
\title{\code{set_swSoils}
Need to define and export this generic method -- otherwise, \code{\link{set_swSoils<-}} doesn't work.}
\usage{
set_swSoils(object, value)
}
\description{
\code{set_swSoils}
Need to define and export this generic method -- otherwise, \code{\link{set_swSoils<-}} doesn't work.
}
