% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{set_swSpinup}
\alias{set_swSpinup}
\title{\code{set_swSpinup}}
\usage{
set_swSpinup(object, value)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swSpinup}} or
\code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
\code{set_swSpinup}
}
