% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{set_swSWC<-}
\alias{set_swSWC<-}
\title{\code{set_swSWC<-}}
\usage{
set_swSWC(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swSWC}} or
\code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
\code{set_swSWC<-}
}
\seealso{
\code{\linkS4class{swSWC}} and \code{\linkS4class{swInputData}}
}
