% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{set_swWeather}
\alias{set_swWeather}
\title{\code{set_swWeather}
Need to define and export this generic method -- otherwise,
\code{\link{set_swWeather<-}} doesn't work.}
\usage{
set_swWeather(object, value)
}
\description{
\code{set_swWeather}
Need to define and export this generic method -- otherwise,
\code{\link{set_swWeather<-}} doesn't work.
}
