% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{set_swWeatherData<-}
\alias{set_swWeatherData<-}
\title{\code{set_swWeatherData<-}}
\usage{
set_swWeatherData(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swWeatherData}} or \code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
\code{set_swWeatherData<-}
}
\seealso{
\code{\linkS4class{swWeatherData}} and \code{\linkS4class{swInputData}}
}
