% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{set_swWeatherData}
\alias{set_swWeatherData}
\title{\code{set_swWeatherData}
Need to define and export this generic method -- otherwise,
\code{\link{set_swWeatherData<-}} doesn't work.}
\usage{
set_swWeatherData(object, value)
}
\description{
\code{set_swWeatherData}
Need to define and export this generic method -- otherwise,
\code{\link{set_swWeatherData<-}} doesn't work.
}
