% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{set_swYears}
\alias{set_swYears}
\title{\code{set_swYears}
Need to define and export this generic method -- otherwise, \code{\link{set_swYears<-}} doesn't work.}
\usage{
set_swYears(object, value)
}
\description{
\code{set_swYears}
Need to define and export this generic method -- otherwise, \code{\link{set_swYears<-}} doesn't work.
}
