% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{set_WeatherHistory<-}
\alias{set_WeatherHistory<-}
\title{set_WeatherHistory<-}
\usage{
set_WeatherHistory(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
set_WeatherHistory<-
}
\seealso{
\code{\linkS4class{swInputData}}
}
