% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{set_WeatherHistory}
\alias{set_WeatherHistory}
\title{\code{set_WeatherHistory}
Need to define and export this generic method -- otherwise,
\code{\link{set_WeatherHistory<-}} doesn't work.}
\usage{
set_WeatherHistory(object, value)
}
\description{
\code{set_WeatherHistory}
Need to define and export this generic method -- otherwise,
\code{\link{set_WeatherHistory<-}} doesn't work.
}
