% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Time_SimulationWorld.R
\name{setup_time_simulation_run}
\alias{setup_time_simulation_run}
\title{Describe the time of a simulation run}
\usage{
setup_time_simulation_run(
  sim_time = list(spinup_N = NULL, simstartyr = NA, startyr = NULL, endyr = NA)
)
}
\arguments{
\item{sim_time}{A list with at least values for three named elements:
\var{\dQuote{endyr}} and two of the following three:
\var{\dQuote{startyr}}, \var{\dQuote{simstartyr}}, and
\var{\dQuote{spinup_N}}.}
}
\value{
A named list, i.e., the updated version of \code{sim_time} with
  additional elements: \itemize{
  \item \var{\dQuote{useyrs}}: numeric vector with included calendar years,
        i.e., years between \var{\dQuote{startyr}} and \var{\dQuote{endyr}},
        but without the \var{\dQuote{spinup_N}} following
        \var{\dQuote{simstartyr}};
  \item \var{\dQuote{no.useyr}}: number of included years;
  \item \var{\dQuote{no.usemo}}: number of included months;
  \item \var{\dQuote{no.usedy}}: number of included days;
  \item \var{\dQuote{index.useyr}}: indices of included years;
  \item \var{\dQuote{index.usemo}}: indices of included months;
  \item \var{\dQuote{index.usedy}}: indices of included days.
  }
}
\description{
Describe the time of a simulation run
}
\examples{
st1 <- setup_time_simulation_run(sim_time =
  list(simstartyr = 1979, startyr = 1980, endyr = 2010))
st2 <- setup_time_simulation_run(sim_time =
  list(spinup_N = 1, startyr = 1980, endyr = 2010))
st3 <- setup_time_simulation_run(sim_time =
  list(simstartyr = 1979, spinup_N = 1, endyr = 2010))

}
