% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Time_SimulationWorld.R
\name{simTiming_ForEachUsedTimeUnit}
\alias{simTiming_ForEachUsedTimeUnit}
\title{Calculate indices along different time steps for simulation time}
\usage{
simTiming_ForEachUsedTimeUnit(useyrs, sim_tscales = c("daily", "weekly",
  "monthly", "yearly"), use_doy_range = FALSE, doy_ranges = list(),
  latitude = 90, account_NorthSouth = TRUE)
}
\arguments{
\item{useyrs}{A numeric vector. The ordered vector of simulation years.}

\item{sim_tscales}{A vector of character strings. One or multiple of
\code{c("daily", "weekly", "monthly", "yearly")}.}

\item{use_doy_range}{A logical value. If \code{TRUE}, then the result is
additional daily indices indicating whether the \var{DOY} is within the
days indicated in the \code{doy_ranges}.}

\item{doy_ranges}{A named list. Aggregation output variables and the daily
\code{c(min, max)} of days you wish to calculate the aggregation over.}

\item{latitude}{A numeric value. The latitude in decimal degrees for which a
hemispheric adjustment is requested; however, the code extracts only the
sign. Positive values are interpreted as from the northern hemisphere;
negative latitudes as from the southern hemisphere.}

\item{account_NorthSouth}{A logical value. If \code{TRUE}, then the result is
corrected for locations on the southern vs. northern hemisphere.}
}
\value{
A named list.
}
\description{
Calculate indices along different time steps for simulation time
}
