% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_swSoils.R
\name{soilLayer_dataColumns}
\alias{soilLayer_dataColumns}
\title{List names of currently implemented soil properties}
\usage{
soilLayer_dataColumns()
}
\value{
A vector of names of soil properties.
}
\description{
List names of currently implemented soil properties
}
