% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_SoilMTRegimes.R
\name{STMR}
\alias{STMR}
\alias{STR_names}
\alias{SMR_names}
\alias{SMRq_names}
\title{Categories of soil temperature regimes and soil moisture regimes}
\usage{
STR_names()

SMR_names()

SMRq_names()
}
\description{
Categories of soil temperature regimes and soil moisture regimes

Soil temperature regime categories

Soil moisture regime categories

Soil moisture regime categories including qualifiers
}
\section{Definitions}{
 Soil temperature and moisture regimes are defined in
  SSS 2014. Our operationalization is explained in the vignette
  \var{SoilMoistureRegimes_SoilTemperatureRegimes}.
}

\examples{
 vignette("SoilMoistureRegimes_SoilTemperatureRegimes",
   package = "rSOILWAT2")
}
\references{
Soil Survey Staff. 2014. Keys to soil taxonomy, 12th ed., USDA
  Natural Resources Conservation Service, Washington, DC.
}
