% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_SoilMTRegimes.R
\name{STR_logic}
\alias{STR_logic}
\title{Soil temperature regime: based on Soil Survey Staff 2014 (Key to Soil
Taxonomy): p.31 we ignore distinction between iso- and not iso-}
\usage{
STR_logic(MAST, MSST, SatSoilSummer_days, has_permafrost, has_Ohorizon)
}
\description{
Soil temperature regime: based on Soil Survey Staff 2014 (Key to Soil
Taxonomy): p.31 we ignore distinction between iso- and not iso-
}
