% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sw2_list_omSWRCp}
\alias{sw2_list_omSWRCp}
\title{Parameters of fibric and sapric peat for implemented \code{SWRCs}}
\format{
A named list with a matrix for each implemented
soil water retention curve (\code{SWRC})
\itemize{
\item Columns represent the \code{SWRC} parameters
\item Rows represent (1) fibric peat and (2) sapric peat
}
}
\usage{
sw2_list_omSWRCp
}
\description{
Parameters of fibric and sapric peat for implemented \code{SWRCs}
}
\examples{
  swin <- rSOILWAT2::sw_exampleData
  swSoils_omSWRCp(swin) <- rSOILWAT2::sw2_list_omSWRCp[["Campbell1974"]]

}
\keyword{datasets}
