% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_swOutput.R
\name{sw_out_flags}
\alias{sw_out_flags}
\title{Slot names of \linkS4class{swOutput}}
\usage{
sw_out_flags()
}
\value{
Standardized named vector for easier access to slots of class
 \linkS4class{swOutput}.
}
\description{
Slot names of \linkS4class{swOutput}
}
