% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsw.R
\name{sw_outputData}
\alias{sw_outputData}
\title{Return output data}
\usage{
sw_outputData(inputData)
}
\arguments{
\item{inputData}{An object of class \code{\linkS4class{swInputData}}.}
}
\value{
An object of class \code{\linkS4class{swOutput}}.
}
\description{
Return output data
}
