% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsw.R
\name{sw_verbosity}
\alias{sw_verbosity}
\title{Turn on/off `SOILWAT2` notes and warnings}
\usage{
sw_verbosity(verbose = TRUE)
}
\arguments{
\item{verbose}{A logical value.}
}
\value{
The previous logical value.
}
\description{
Turn on/off `SOILWAT2` notes and warnings
}
