% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{sw_weather_database}
\alias{sw_weather_database}
\title{Weather data base structure}
\arguments{
\item{dbFilePath}{A character string. The file path of the weather database.
This will be a file of type \code{sqlite3}. In-memory databases are not
supported.}

\item{site_data}{A data.frame. The site data with column names
\code{Longitude}, \code{Latitude}, and \code{Label}.}

\item{Site_id}{An integer value. The IDs/database key of the queried site.}

\item{site_id}{An integer value. The IDs/database key of the queried site.}

\item{Site_ids}{An integer vector. The IDs/database keys of the queried sites}

\item{site_ids}{An integer vector. The IDs/database keys of the queried sites}

\item{Labels}{A vector of character strings. The names/labels of
queried sites.}

\item{Label}{A character string. The name/label of the queried site.}

\item{site_labels}{A vector of character string. The names/labels of
queried sites.}

\item{site_label}{A character string. The name/label of the queried site.}

\item{lat}{A numeric vector or \code{NULL}. The latitude in decimal degrees
of \code{WGS84}. Northern latitude are positive, sites on the southern
hemisphere have negative values.}

\item{long}{A numeric vector or \code{NULL}. The longitude in decimal degrees
of \code{WGS84}. Eastern longitudes are positive, sites on the western
hemisphere have negative values.}

\item{Scenario_ids}{An integer vector. The IDs/database keys of the queried
scenario.}

\item{scen_ids}{An integer vector. The IDs/database keys of the queried
scenario.}

\item{Scenario_id}{An integer value The ID/database key of the queried
scenario.}

\item{scenario_id}{An integer value The ID/database key of the queried
scenario.}

\item{Scenarios}{A vector of character strings. The climate scenarios of
which the first one is enforced to be \code{scen_ambient}.}

\item{scen_labels}{A vector of character strings. The climate scenarios of
which the first one is enforced to be \code{scen_ambient}.}

\item{Scenario}{A character string. The name/label of a climate scenario.}

\item{scenario}{A character string. The name/label of a climate scenario.}

\item{scen_ambient}{A character string. The first/default climate scenario.}

\item{startYear}{A numeric value. First calendar year of the weather data.}

\item{endYear}{A numeric value. Last calendar year of the weather data.}

\item{ignore.case}{A logical value.}

\item{verbose}{A logical value.}
}
\description{
Weather data base structure
}
