% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swCarbon_CO2ppm}
\alias{swCarbon_CO2ppm}
\title{swCarbon_CO2ppm}
\usage{
swCarbon_CO2ppm(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swCarbon}} or \code{\linkS4class{swInputData}}.}
}
\description{
swCarbon_CO2ppm
}
\seealso{
\code{\linkS4class{swCarbon}} and \code{\linkS4class{swInputData}}
}
