% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_swCloud.R
\docType{class}
\name{swCloud-class}
\alias{swCloud-class}
\alias{initialize,swCloud-method}
\alias{get_swCloud,swCloud-method}
\alias{swCloud_SkyCover,swCloud-method}
\alias{swCloud_WindSpeed,swCloud-method}
\alias{swCloud_Humidity,swCloud-method}
\alias{swCloud_SnowDensity,swCloud-method}
\alias{swCloud_RainEvents,swCloud-method}
\alias{set_swCloud<-,swCloud-method}
\alias{swCloud_SkyCover<-,swCloud-method}
\alias{swCloud_WindSpeed<-,swCloud-method}
\alias{swCloud_Humidity<-,swCloud-method}
\alias{swCloud_SnowDensity<-,swCloud-method}
\alias{swCloud_RainEvents<-,swCloud-method}
\alias{swReadLines,swCloud,character-method}
\title{Class \code{"swCloud"}}
\usage{
\S4method{initialize}{swCloud}(.Object, ...)

\S4method{get_swCloud}{swCloud}(object)

\S4method{swCloud_SkyCover}{swCloud}(object)

\S4method{swCloud_WindSpeed}{swCloud}(object)

\S4method{swCloud_Humidity}{swCloud}(object)

\S4method{swCloud_SnowDensity}{swCloud}(object)

\S4method{swCloud_RainEvents}{swCloud}(object)

\S4method{set_swCloud}{swCloud}(object) <- value

\S4method{swCloud_SkyCover}{swCloud}(object) <- value

\S4method{swCloud_WindSpeed}{swCloud}(object) <- value

\S4method{swCloud_Humidity}{swCloud}(object) <- value

\S4method{swCloud_SnowDensity}{swCloud}(object) <- value

\S4method{swCloud_RainEvents}{swCloud}(object) <- value

\S4method{swReadLines}{swCloud,character}(object, file)
}
\arguments{
\item{.Object}{An object of class \code{\linkS4class{swCloud}}.}

\item{...}{Further arguments to methods.}

\item{object}{An object of class \code{\linkS4class{swCloud}}.}

\item{value}{A value to assign to a specific slot of the object.}

\item{file}{A character string. The file name from which to read.}
}
\description{
The methods listed below work on this class and the proper slot of the class
  \code{\linkS4class{swInputData}}.
}
\examples{
showClass("swCloud")
x <- new("swCloud")

}
\seealso{
\code{\linkS4class{swInputData}} \code{\linkS4class{swFiles}}
\code{\linkS4class{swWeather}} \code{\linkS4class{swInputData}}
\code{\linkS4class{swMarkov}} \code{\linkS4class{swProd}}
\code{\linkS4class{swSite}} \code{\linkS4class{swSoils}}
\code{\linkS4class{swEstab}} \code{\linkS4class{swOUT}}
\code{\linkS4class{swSWC}} \code{\linkS4class{swLog}}
}
