% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swCloud_Humidity}
\alias{swCloud_Humidity}
\title{\code{swCloud_Humidity}}
\usage{
swCloud_Humidity(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swCloud}} or
\code{\linkS4class{swInputData}}.}
}
\description{
\code{swCloud_Humidity}
}
\seealso{
\code{\linkS4class{swCloud}} and \code{\linkS4class{swInputData}}
}
