% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swCloud_SkyCover}
\alias{swCloud_SkyCover}
\title{\code{swCloud_SkyCover}}
\usage{
swCloud_SkyCover(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swCloud}} or \code{\linkS4class{swInputData}}.}
}
\description{
\code{swCloud_SkyCover}
}
\seealso{
\code{\linkS4class{swCloud}} and \code{\linkS4class{swInputData}}
}
