% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swCloud_SnowDensity<-}
\alias{swCloud_SnowDensity<-}
\title{\code{swCloud_SnowDensity<-}}
\usage{
swCloud_SnowDensity(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swCloud}} or
\code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
\code{swCloud_SnowDensity<-}
}
\seealso{
\code{\linkS4class{swCloud}} and \code{\linkS4class{swInputData}}
}
