% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swCloud_Transmissivity}
\alias{swCloud_Transmissivity}
\title{swCloud_Transmissivity}
\usage{
swCloud_Transmissivity(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swCloud}} or \code{\linkS4class{swInputData}}.}
}
\description{
swCloud_Transmissivity
}
\seealso{
\code{\linkS4class{swCloud}} and \code{\linkS4class{swInputData}}
}
