% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swCloud_WindSpeed}
\alias{swCloud_WindSpeed}
\title{\code{swCloud_WindSpeed}}
\usage{
swCloud_WindSpeed(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swCloud}} or \code{\linkS4class{swInputData}}.}
}
\description{
\code{swCloud_WindSpeed}
}
\seealso{
\code{\linkS4class{swCloud}} and \code{\linkS4class{swInputData}}
}
