% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_swFiles.R
\docType{class}
\name{swFiles-class}
\alias{swFiles-class}
\alias{initialize,swFiles-method}
\alias{swFiles_ProjDir,swFiles-method}
\alias{swFiles_WeatherPrefix,swFiles-method}
\alias{swFiles_OutputPrefix,swFiles-method}
\alias{swFiles_filesIn,swFiles-method}
\alias{swFiles_Years,swFiles-method}
\alias{swFiles_LogFile,swFiles-method}
\alias{swFiles_SiteParams,swFiles-method}
\alias{swFiles_Soils,swFiles-method}
\alias{swFiles_WeatherSetup,swFiles-method}
\alias{swFiles_MarkovProbs,swFiles-method}
\alias{swFiles_MarkovCov,swFiles-method}
\alias{swFiles_Cloud,swFiles-method}
\alias{swFiles_Prod,swFiles-method}
\alias{swFiles_Estab,swFiles-method}
\alias{swFiles_SWCsetup,swFiles-method}
\alias{swFiles_Carbon,swFiles-method}
\alias{swFiles_Output,swFiles-method}
\alias{swFiles_ProjDir<-,swFiles-method}
\alias{swFiles_WeatherPrefix<-,swFiles-method}
\alias{swFiles_OutputPrefix<-,swFiles-method}
\alias{swFiles_filesIn<-,swFiles-method}
\alias{swFiles_Years<-,swFiles-method}
\alias{swFiles_LogFile<-,swFiles-method}
\alias{swFiles_SiteParams<-,swFiles-method}
\alias{swFiles_Soils<-,swFiles-method}
\alias{swFiles_WeatherSetup<-,swFiles-method}
\alias{swFiles_MarkovProbs<-,swFiles-method}
\alias{swFiles_MarkovCov<-,swFiles-method}
\alias{swFiles_Cloud<-,swFiles-method}
\alias{swFiles_Prod<-,swFiles-method}
\alias{swFiles_Estab<-,swFiles-method}
\alias{swFiles_SWCsetup<-,swFiles-method}
\alias{swFiles_Carbon<-,swFiles-method}
\alias{swFiles_Output<-,swFiles-method}
\alias{swReadLines,swFiles,character-method}
\title{Class \code{"swFiles"}}
\usage{
\S4method{initialize}{swFiles}(.Object, ...)

\S4method{swFiles_ProjDir}{swFiles}(object)

\S4method{swFiles_WeatherPrefix}{swFiles}(object)

\S4method{swFiles_OutputPrefix}{swFiles}(object)

\S4method{swFiles_filesIn}{swFiles}(object)

\S4method{swFiles_Years}{swFiles}(object)

\S4method{swFiles_LogFile}{swFiles}(object)

\S4method{swFiles_SiteParams}{swFiles}(object)

\S4method{swFiles_Soils}{swFiles}(object)

\S4method{swFiles_WeatherSetup}{swFiles}(object)

\S4method{swFiles_MarkovProbs}{swFiles}(object)

\S4method{swFiles_MarkovCov}{swFiles}(object)

\S4method{swFiles_Cloud}{swFiles}(object)

\S4method{swFiles_Prod}{swFiles}(object)

\S4method{swFiles_Estab}{swFiles}(object)

\S4method{swFiles_SWCsetup}{swFiles}(object)

\S4method{swFiles_Carbon}{swFiles}(object)

\S4method{swFiles_Output}{swFiles}(object)

\S4method{swFiles_ProjDir}{swFiles}(object) <- value

\S4method{swFiles_WeatherPrefix}{swFiles}(object) <- value

\S4method{swFiles_OutputPrefix}{swFiles}(object) <- value

\S4method{swFiles_filesIn}{swFiles}(object) <- value

\S4method{swFiles_Years}{swFiles}(object) <- value

\S4method{swFiles_LogFile}{swFiles}(object) <- value

\S4method{swFiles_SiteParams}{swFiles}(object) <- value

\S4method{swFiles_Soils}{swFiles}(object) <- value

\S4method{swFiles_WeatherSetup}{swFiles}(object) <- value

\S4method{swFiles_MarkovProbs}{swFiles}(object) <- value

\S4method{swFiles_MarkovCov}{swFiles}(object) <- value

\S4method{swFiles_Cloud}{swFiles}(object) <- value

\S4method{swFiles_Prod}{swFiles}(object) <- value

\S4method{swFiles_Estab}{swFiles}(object) <- value

\S4method{swFiles_SWCsetup}{swFiles}(object) <- value

\S4method{swFiles_Carbon}{swFiles}(object) <- value

\S4method{swFiles_Output}{swFiles}(object) <- value

\S4method{swReadLines}{swFiles,character}(object, file)
}
\arguments{
\item{.Object}{An object of class \code{\linkS4class{swFiles}}.}

\item{...}{Further arguments to methods.}

\item{object}{An object of class \code{\linkS4class{swFiles}}.}

\item{value}{A value to assign to a specific slot of the object.}

\item{file}{A character string. The file name from which to read.}
}
\description{
The methods listed below work on this class and the proper slot of the class
  \code{\linkS4class{swInputData}}.
}
\examples{
showClass("swFiles")
x <- new("swFiles")

}
\seealso{
\code{\linkS4class{swInputData}} \code{\linkS4class{swYears}}
\code{\linkS4class{swWeather}} \code{\linkS4class{swCloud}}
\code{\linkS4class{swMarkov}} \code{\linkS4class{swProd}}
\code{\linkS4class{swSite}} \code{\linkS4class{swSoils}}
\code{\linkS4class{swEstab}} \code{\linkS4class{swOUT}}
\code{\linkS4class{swSWC}} \code{\linkS4class{swLog}}
\code{\linkS4class{swCarbon}}
}
