% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swFiles_Carbon}
\alias{swFiles_Carbon}
\title{swFiles_Carbon}
\usage{
swFiles_Carbon(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swFiles}} or \code{\linkS4class{swInputData}}.}
}
\description{
swFiles_Carbon
}
\seealso{
\code{\linkS4class{swFiles}} and \code{\linkS4class{swInputData}}
}
