% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swFiles_OutputPrefix<-}
\alias{swFiles_OutputPrefix<-}
\title{\code{swFiles_OutputPrefix<-}}
\usage{
swFiles_OutputPrefix(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swFiles}} or \code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
\code{swFiles_OutputPrefix<-}
}
\seealso{
\code{\linkS4class{swFiles}} and \code{\linkS4class{swInputData}}
}
