% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swFiles_Prod}
\alias{swFiles_Prod}
\title{\code{swFiles_Prod}}
\usage{
swFiles_Prod(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swFiles}} or
\code{\linkS4class{swInputData}}.}
}
\description{
\code{swFiles_Prod}
}
\seealso{
\code{\linkS4class{swFiles}} and \code{\linkS4class{swInputData}}
}
