% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swFiles_SiteParams}
\alias{swFiles_SiteParams}
\title{\code{swFiles_SiteParams}}
\usage{
swFiles_SiteParams(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swFiles}} or \code{\linkS4class{swInputData}}.}
}
\description{
\code{swFiles_SiteParams}
}
\seealso{
\code{\linkS4class{swFiles}} and \code{\linkS4class{swInputData}}
}
