% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swFiles_SWRCp<-}
\alias{swFiles_SWRCp<-}
\title{\code{swFiles_SWRCp<-}}
\usage{
swFiles_SWRCp(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swFiles}} or
\code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
\code{swFiles_SWRCp<-}
}
\seealso{
\code{\linkS4class{swFiles}} and \code{\linkS4class{swInputData}}
}
