% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_swMarkov.R
\docType{class}
\name{swMarkov-class}
\alias{swMarkov-class}
\alias{initialize,swMarkov-method}
\alias{get_Markov,swMarkov-method}
\alias{swMarkov_Prob,swMarkov-method}
\alias{swMarkov_Conv,swMarkov-method}
\alias{set_Markov<-,swMarkov-method}
\alias{swMarkov_Prob<-,swMarkov-method}
\alias{swMarkov_Conv<-,swMarkov-method}
\alias{swReadLines,swMarkov,character-method}
\title{Class \code{"swMarkov"}}
\usage{
\S4method{initialize}{swMarkov}(.Object, ...)

\S4method{get_Markov}{swMarkov}(object)

\S4method{swMarkov_Prob}{swMarkov}(object)

\S4method{swMarkov_Conv}{swMarkov}(object)

\S4method{set_Markov}{swMarkov}(object) <- value

\S4method{swMarkov_Prob}{swMarkov}(object) <- value

\S4method{swMarkov_Conv}{swMarkov}(object) <- value

\S4method{swReadLines}{swMarkov,character}(object, file)
}
\arguments{
\item{.Object}{An object of class \code{\linkS4class{swMarkov}}.}

\item{...}{Further arguments to methods.}

\item{object}{An object of class \code{\linkS4class{swMarkov}}.}

\item{value}{A value to assign to a specific slot of the object.}

\item{file}{A character string. The file name from which to read.}
}
\description{
The methods listed below work on this class and the proper slot of the class
  \code{\linkS4class{swInputData}}.
}
\examples{
showClass("swMarkov")
x <- new("swMarkov")

}
\seealso{
\code{\linkS4class{swInputData}} \code{\linkS4class{swFiles}}
\code{\linkS4class{swWeather}} \code{\linkS4class{swCloud}}
\code{\linkS4class{swInputData}} \code{\linkS4class{swProd}}
\code{\linkS4class{swSite}} \code{\linkS4class{swSoils}}
\code{\linkS4class{swEstab}} \code{\linkS4class{swOUT}}
\code{\linkS4class{swSWC}} \code{\linkS4class{swLog}}
}
