% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swMarkov_Conv<-}
\alias{swMarkov_Conv<-}
\title{swMarkov_Conv<-}
\usage{
swMarkov_Conv(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swMarkov}} or \code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
swMarkov_Conv<-
}
\seealso{
\code{\linkS4class{swMarkov}} and \code{\linkS4class{swInputData}}
}
