% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swMarkov_Conv}
\alias{swMarkov_Conv}
\title{swMarkov_Conv}
\usage{
swMarkov_Conv(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swMarkov}} or \code{\linkS4class{swInputData}}.}
}
\description{
swMarkov_Conv
}
\seealso{
\code{\linkS4class{swMarkov}} and \code{\linkS4class{swInputData}}
}
