% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swMarkov_Prob}
\alias{swMarkov_Prob}
\title{\code{swMarkov_Prob}}
\usage{
swMarkov_Prob(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swMarkov}} or
\code{\linkS4class{swInputData}}.}
}
\description{
\code{swMarkov_Prob}
}
\seealso{
\code{\linkS4class{swMarkov}} and \code{\linkS4class{swInputData}}
}
