% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_swOut.R, R/rSOILWAT2-package.R
\docType{methods}
\name{swOUT_TimeStepsForEveryKey<-,swOUT-method}
\alias{swOUT_TimeStepsForEveryKey<-,swOUT-method}
\alias{swOUT-class}
\alias{set_swOUT<-,swOUT-method}
\alias{get_swOUT,swOUT-method}
\alias{swOUT_OutputSeparator<-,swOUT-method}
\alias{swOUT_OutputSeparator,swOUT-method}
\alias{swOUT_TimeStep<-,swOUT-method}
\alias{swOUT_TimeStep,swOUT-method}
\alias{swReadLines,swOUT-method}
\title{Class \code{swOUT}}
\usage{
\S4method{swOUT_TimeStepsForEveryKey}{swOUT}(object) <- value
}
\description{
The generic functions listed work on this and the proper sw Class in the container's slots.
}
\details{
Output can be generated for four different time steps: daily (DY),
 weekly (WK), monthly (MO), and yearly (YR) periods.
 We have two options to specify time steps:\itemize{
   \item The same time step(s) for every output; this option corresponds to specifying
       a line with `TIMESTEP ...` in the SOILWAT2 input file `outsetup.in`. The matrix
       in slot `timeSteps` should have `SW_OUTNKEYS` rows and `used_SW_OUTNPERIODS`
       columns where each row contains identical values.
   \item A different time step for each output; however, only one time step per
       output variable can be specified. this option corresponds to specifying the
       time step in the column `PERIOD` in the SOILWAT2 input file `outsetup.in`. The
       matrix in slot `timeSteps` should have `SW_OUTNKEYS` rows and 1 column.
}
}
\section{Methods (by generic)}{
\itemize{
\item \code{swOUT_TimeStepsForEveryKey<-}: Set time steps to the same set of values for each output key.
}}

\section{Slots}{

\describe{
\item{\code{outputSeparator}}{A character string. Currently, only "\\t" is functional.}

\item{\code{timeSteps}}{An integer matrix. See details.}
}}

\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("swOUT", ...)}.
}

\examples{
x <- new("swOUT")
swOUT_TimeStepsForEveryKey(x) <- c(2, 3)
identical(as.vector(unique(swOUT_TimeStep(x))), as.integer(c(2, 3)))

showClass("swOUT")

}
\references{
https://github.com/DrylandEcology/rSOILWAT2
}
\seealso{
\code{\linkS4class{swInputData}}
}
\author{
Ryan Murphy
}
\keyword{classes}
