% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_swOut.R
\docType{class}
\name{swOUT_key-class}
\alias{swOUT_key-class}
\alias{initialize,swOUT_key-method}
\title{Class \code{"swOUT_key"}}
\usage{
\S4method{initialize}{swOUT_key}(.Object, ...)
}
\arguments{
\item{.Object}{An object of class \code{\linkS4class{swOUT_key}}.}

\item{...}{Further arguments to methods.}
}
\description{
The methods listed below work on this class and the proper slot of the class
  \code{\linkS4class{swInputData}}.
}
\examples{
showClass("swOUT_key")
x <- new("swOUT_key")

}
