% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swOUT_OutputSeparator<-}
\alias{swOUT_OutputSeparator<-}
\title{swOUT_OutputSeparator<-}
\usage{
swOUT_OutputSeparator(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swOUT}} or \code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
swOUT_OutputSeparator<-
}
\seealso{
\code{\linkS4class{swOUT}} and \code{\linkS4class{swInputData}}
}
