% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swOUT_TimeStep<-}
\alias{swOUT_TimeStep<-}
\title{swOUT_TimeStep<-}
\usage{
swOUT_TimeStep(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swOUT}} or \code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
swOUT_TimeStep<-
}
\seealso{
\code{\linkS4class{swOUT}} and \code{\linkS4class{swInputData}}
}
