% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swOutput_getKEY}
\alias{swOutput_getKEY}
\title{\code{swOutput_getKEY}}
\usage{
swOutput_getKEY(object, index)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swOutput}} or
\code{\linkS4class{swInputData}}.}
}
\description{
\code{swOutput_getKEY}
}
\seealso{
\code{\linkS4class{swOutput}} and \code{\linkS4class{swInputData}}
}
