% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_swOutput.R
\docType{class}
\name{swOutput_KEY-class}
\alias{swOutput_KEY-class}
\alias{swOutput_KEY}
\alias{swOutput_KEY_Period,swOutput_KEY-method}
\alias{swOutput_KEY_TimeStep,swOutput_KEY-method}
\alias{swOutput_KEY_Columns,swOutput_KEY-method}
\alias{swOutput_KEY_Period<-,swOutput_KEY-method}
\title{Class \code{"swOutput_KEY"}}
\usage{
\S4method{swOutput_KEY_Period}{swOutput_KEY}(object, index)

\S4method{swOutput_KEY_TimeStep}{swOutput_KEY}(object)

\S4method{swOutput_KEY_Columns}{swOutput_KEY}(object)

\S4method{swOutput_KEY_Period}{swOutput_KEY}(object, index) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swOutput_KEY}}.}

\item{index}{An integer value. One of the four possible time steps.}

\item{value}{A value to assign to a specific slot of the object.}
}
\description{
The methods listed below work on this class and the proper slot of the class
  \code{\linkS4class{swInputData}}.
}
\examples{
showClass("swOutput_KEY")
x <- new("swOutput_KEY")

}
\seealso{
\code{\linkS4class{swInputData}}
}
