% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swOutput_KEY_Columns}
\alias{swOutput_KEY_Columns}
\title{swOutput_KEY_Columns}
\usage{
swOutput_KEY_Columns(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swOutput}} or \code{\linkS4class{swInputData}}.}
}
\description{
swOutput_KEY_Columns
}
\seealso{
\code{\linkS4class{swOutput}} and \code{\linkS4class{swInputData}}
}
