% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swOutput_KEY_Period<-}
\alias{swOutput_KEY_Period<-}
\title{\code{swOutput_KEY_Period<-}}
\usage{
swOutput_KEY_Period(object, index) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swOutput}} or
\code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
\code{swOutput_KEY_Period<-}
}
\seealso{
\code{\linkS4class{swOutput}} and \code{\linkS4class{swInputData}}
}
