% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swOutput_KEY_TimeStep<-}
\alias{swOutput_KEY_TimeStep<-}
\title{swOutput_KEY_TimeStep<-}
\usage{
swOutput_KEY_TimeStep(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swOutput}} or \code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
swOutput_KEY_TimeStep<-
}
\seealso{
\code{\linkS4class{swOutput}} and \code{\linkS4class{swInputData}}
}
