% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swProd_CritSoilWaterPotential}
\alias{swProd_CritSoilWaterPotential}
\title{swProd_CritSoilWaterPotential}
\usage{
swProd_CritSoilWaterPotential(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swProd}} or \code{\linkS4class{swInputData}}.}
}
\description{
swProd_CritSoilWaterPotential
}
\seealso{
\code{\linkS4class{swProd}} and \code{\linkS4class{swInputData}}
}
