% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swProd_HydrRedstro<-}
\alias{swProd_HydrRedstro<-}
\title{swProd_HydrRedstro<-}
\usage{
swProd_HydrRedstro(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swProd}} or \code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
swProd_HydrRedstro<-
}
\seealso{
\code{\linkS4class{swProd}} and \code{\linkS4class{swInputData}}
}
