% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swProd_HydrRedstro_use}
\alias{swProd_HydrRedstro_use}
\title{swProd_HydrRedstro_use}
\usage{
swProd_HydrRedstro_use(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swProd}} or \code{\linkS4class{swInputData}}.}
}
\description{
swProd_HydrRedstro_use
}
\seealso{
\code{\linkS4class{swProd}} and \code{\linkS4class{swInputData}}
}
