% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swProd_MonProd_forb}
\alias{swProd_MonProd_forb}
\title{swProd_MonProd_forb}
\usage{
swProd_MonProd_forb(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swProd}} or \code{\linkS4class{swInputData}}.}
}
\description{
swProd_MonProd_forb
}
\seealso{
\code{\linkS4class{swProd}} and \code{\linkS4class{swInputData}}
}
