% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swProd_MonProd_grass}
\alias{swProd_MonProd_grass}
\title{\code{swProd_MonProd_grass}}
\usage{
swProd_MonProd_grass(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swProd}} or
\code{\linkS4class{swInputData}}.}
}
\description{
\code{swProd_MonProd_grass}
}
\seealso{
\code{\linkS4class{swProd}} and \code{\linkS4class{swInputData}}
}
