% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Pedotransfer_Functions.R
\name{swrc_names}
\alias{swrc_names}
\title{List Soil Water Retention Curves \code{SWRCs}}
\usage{
swrc_names()
}
\value{
An integer vector with names of implemented \code{SWRCs}
}
\description{
List Soil Water Retention Curves \code{SWRCs}
}
\details{
Notes:
The integer values may change with new versions of \code{SOILWAT2.}
}
\seealso{
\code{\link{SWRCs}}, \code{\link[=pdf_names]{pdf_names()}}
}
