% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Pedotransfer_Functions.R
\name{SWRCs}
\alias{SWRCs}
\title{Functionality for Soil Water Retention Curves (\code{SWRC})}
\arguments{
\item{sand}{A numeric value or vector.
Sand content of the matric soil component
(< 2 mm fraction; units of \verb{[g/g]}) of each soil layer.}

\item{clay}{A numeric value or vector.
Clay content of the matric soil component
(< 2 mm fraction; units of \verb{[g/g]}) of each soil layer.}

\item{fcoarse}{A numeric value or vector.
Coarse fragments, e.g., gravel, (> 2 mm; units of \verb{[m3/m3]})
relative to the whole soil of each soil layer.
\code{fcoarse} is required, for instance, to translate between
values relative to the matric soil component (< 2 mm fraction) and
relative to the whole soil (matric soil plus coarse fragments).}

\item{layer_width}{A numeric value or vector.
Depth interval, width, of each soil layer (units of \code{cm}).
\code{layer_width} is required to translate between
soil water content of a soil layer and volumetric water content.}

\item{swrc_type}{An integer value or vector.
The identification number of selected \code{SWRC}.}

\item{pdf_type}{An integer value or vector.
The identification number of selected \code{PDF}.}

\item{swrcp}{A numeric vector or matrix.
The parameters of a selected \code{SWRC};
each row represents one \code{SWRC}, e.g., one per soil layer.}

\item{swrc}{A named list.
Contains all necessary elements of a \code{SWRC},
i.e., \code{type} (short for \code{swrc_type}) and \code{swrcp},
or all necessary elements to estimate parameters of a \code{SWRC} given
soil parameters, i.e., \code{swrc_type} and \code{pdf_type}.}
}
\description{
\code{SWRCs} convert between soil water content and soil water potential
using a set of parameters, see \code{\link[=swrc_swp_to_vwc]{swrc_swp_to_vwc()}} and \code{\link[=swrc_vwc_to_swp]{swrc_vwc_to_swp()}}.

The \code{SWRC} parameters may be estimated from soil properties with suitable
pedotransfer functions \code{PDFs}, see \code{\link[=pdf_estimate]{pdf_estimate()}}.

The \code{SWRC} parameters can be checked for consistency with \code{\link[=check_swrcp]{check_swrcp()}}.
}
\section{Details}{

Implemented SWRCs (\code{swrc_type}):
\enumerate{
\item Campbell 1974
}
}

\references{
Campbell, G. S. 1974.
A simple method for determining unsaturated conductivity
from moisture retention data.
Soil Science, 117(6):311-314. \doi{10.1097/00010694-197406000-00001}
}
\seealso{
\code{\link[=pdf_estimate]{pdf_estimate()}},
\code{\link[=check_swrcp]{check_swrcp()}},
\code{\link[=swrc_swp_to_vwc]{swrc_swp_to_vwc()}},
\code{\link[=swrc_vwc_to_swp]{swrc_vwc_to_swp()}}
}
