% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R, R/F_swSite.R,
%   R/K_swContainer.R
\name{swSite_depthSapric}
\alias{swSite_depthSapric}
\alias{swSite_depthSapric<-}
\alias{swSite_depthSapric,swSite-method}
\alias{swSite_depthSapric<-,swSite-method}
\alias{swSite_depthSapric,swInputData-method}
\alias{swSite_depthSapric<-,swInputData-method}
\title{\code{swSite_depthSapric}}
\usage{
swSite_depthSapric(object)

swSite_depthSapric(object) <- value

\S4method{swSite_depthSapric}{swSite}(object)

\S4method{swSite_depthSapric}{swSite}(object) <- value

\S4method{swSite_depthSapric}{swInputData}(object)

\S4method{swSite_depthSapric}{swInputData}(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swSite}} or
\code{\linkS4class{swInputData}}.}

\item{value}{A numeric value.}
}
\description{
\code{swSite_depthSapric}
}
\seealso{
\code{\linkS4class{swSite}} and \code{\linkS4class{swInputData}}
}
