% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swSite_EvapCoefficients}
\alias{swSite_EvapCoefficients}
\title{\code{swSite_EvapCoefficients}}
\usage{
swSite_EvapCoefficients(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swSite}} or
\code{\linkS4class{swInputData}}.}
}
\description{
\code{swSite_EvapCoefficients}
}
\seealso{
\code{\linkS4class{swSite}} and \code{\linkS4class{swInputData}}
}
