% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swSite_ModelFlags}
\alias{swSite_ModelFlags}
\title{\code{swSite_ModelFlags}}
\usage{
swSite_ModelFlags(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swSite}} or \code{\linkS4class{swInputData}}.}
}
\description{
\code{swSite_ModelFlags}
}
\seealso{
\code{\linkS4class{swSite}} and \code{\linkS4class{swInputData}}
}
