% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R, R/F_swSite.R,
%   R/K_swContainer.R
\name{swSite_PDFNoPDF}
\alias{swSite_PDFNoPDF}
\alias{swSite_PDFNoPDF<-}
\alias{swSite_PDFNoPDF,swSite-method}
\alias{swSite_PDFNoPDF<-,swSite-method}
\alias{swSite_PDFNoPDF,swInputData-method}
\alias{swSite_PDFNoPDF<-,swInputData-method}
\title{The utilized \code{pdf} name stored for later use}
\usage{
swSite_PDFNoPDF(object)

swSite_PDFNoPDF(object) <- value

\S4method{swSite_PDFNoPDF}{swSite}(object)

\S4method{swSite_PDFNoPDF}{swSite}(object) <- value

\S4method{swSite_PDFNoPDF}{swInputData}(object)

\S4method{swSite_PDFNoPDF}{swInputData}(object) <- value
}
\arguments{
\item{object}{An object of class \linkS4class{swSite} or \linkS4class{swInputData}.}

\item{value}{A character vector of a pedotransfer function name.}
}
\value{
The pedotransfer function name that was stored
with \code{\link[=swSite_PDFNoPDF]{swSite_PDFNoPDF()}} for later usage
in the situation when \code{pdf_name} is set to \code{"NoPDF"} but
the \code{pdf} was actually being for estimating the \code{SWRC} parameters.
}
\description{
The utilized \code{pdf} name stored for later use
}
