% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R, R/F_swSite.R,
%   R/K_swContainer.R
\name{swSite_PDFutilized}
\alias{swSite_PDFutilized}
\alias{swSite_PDFutilized,swSite-method}
\alias{swSite_PDFutilized,swInputData-method}
\title{Identify the actually utilized \code{pdf} name}
\usage{
swSite_PDFutilized(object)

\S4method{swSite_PDFutilized}{swSite}(object)

\S4method{swSite_PDFutilized}{swInputData}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{swSite} or \linkS4class{swInputData}.}
}
\value{
The pedotransfer function name that was utilized for estimating
\code{SWRC} given that it was stored either via \code{\link[=swSite_SWRCflags]{swSite_SWRCflags()}} or
\code{\link[=swSite_PDFNoPDF]{swSite_PDFNoPDF()}}.
}
\description{
Identify the actually utilized \code{pdf} name
}
