% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swSite_RootingProfileMethod}
\alias{swSite_RootingProfileMethod}
\title{\code{swSite_RootingProfileMethod}}
\usage{
swSite_RootingProfileMethod(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swSite}} or
\code{\linkS4class{swInputData}}.}
}
\description{
\code{swSite_RootingProfileMethod}
}
\seealso{
\code{\linkS4class{swSite}} and \code{\linkS4class{swInputData}}
}
