% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swSite_SnowSimulationParams}
\alias{swSite_SnowSimulationParams}
\title{swSite_SnowSimulationParams}
\usage{
swSite_SnowSimulationParams(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swSite}} or \code{\linkS4class{swInputData}}.}
}
\description{
swSite_SnowSimulationParams
}
\seealso{
\code{\linkS4class{swSite}} and \code{\linkS4class{swInputData}}
}
