% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swSite_SWClimits}
\alias{swSite_SWClimits}
\title{\code{swSite_SWClimits}}
\usage{
swSite_SWClimits(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swSite}} or \code{\linkS4class{swInputData}}.}
}
\description{
\code{swSite_SWClimits}
}
\seealso{
\code{\linkS4class{swSite}} and \code{\linkS4class{swInputData}}
}
