% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R, R/F_swSite.R,
%   R/K_swContainer.R
\name{swSite_SWRCflags}
\alias{swSite_SWRCflags}
\alias{swSite_SWRCflags<-}
\alias{swSite_SWRCflags,swSite-method}
\alias{swSite_SWRCflags<-,swSite-method}
\alias{swSite_SWRCflags,swInputData-method}
\alias{swSite_SWRCflags<-,swInputData-method}
\title{Names of \code{SWRC} and \code{PTF}}
\usage{
swSite_SWRCflags(object)

swSite_SWRCflags(object) <- value

\S4method{swSite_SWRCflags}{swSite}(object)

\S4method{swSite_SWRCflags}{swSite}(object) <- value

\S4method{swSite_SWRCflags}{swInputData}(object)

\S4method{swSite_SWRCflags}{swInputData}(object) <- value
}
\arguments{
\item{object}{An object of class \link[=swSite-class]{swSite} or \link[=swInputData-class]{swInputData}.}

\item{value}{A character vector with two elements for
\code{"swrc_name"} and \code{"ptf_name"}.}
}
\value{
A character vector with two elements \code{"swrc_name"} and \code{"ptf_name"}.
}
\description{
Names of \code{SWRC} and \code{PTF}
}
\section{Details}{

The replacement method \code{\link[=swSite_SWRCflags]{swSite_SWRCflags()}} for class \link[=swInputData-class]{swInputData}
resets \code{has_swrcp} to \code{FALSE} if \code{"swrc_name"} or \code{"ptf_name"} are updated.
This is to avoid inconsistency between
\code{SWRCp}, \code{has_swrcp}, and \code{swrc_flags}.


The correct sequence for setting values is
\enumerate{
\item \code{\link[=swSoils_Layers]{swSoils_Layers()}},
\item \code{\link[=swSite_SWRCflags]{swSite_SWRCflags()}}, and
\item \code{\link[=swSoils_SWRCp]{swSoils_SWRCp()}} and \code{\link[=swSite_hasSWRCp]{swSite_hasSWRCp()}}
\item \code{\link[=swSoils_omSWRCp]{swSoils_omSWRCp()}} (and \code{\link[=swSite_depthSapric]{swSite_depthSapric()}})
}
}

