% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swSite_TranspCoefficients}
\alias{swSite_TranspCoefficients}
\title{swSite_TranspCoefficients}
\usage{
swSite_TranspCoefficients(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swSite}} or \code{\linkS4class{swInputData}}.}
}
\description{
swSite_TranspCoefficients
}
\seealso{
\code{\linkS4class{swSite}} and \code{\linkS4class{swInputData}}
}
