% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_swSoils.R
\docType{class}
\name{swSoils-class}
\alias{swSoils-class}
\alias{initialize,swSoils-method}
\alias{get_swSoils,swSoils-method}
\alias{swSoils_Layers,swSoils-method}
\alias{set_swSoils<-,swSoils,swSoils-method}
\alias{swSoils_Layers<-,swSoils-method}
\alias{swReadLines,swSoils,character-method}
\title{Class \code{"swSoils"}}
\usage{
\S4method{initialize}{swSoils}(.Object, ...)

\S4method{get_swSoils}{swSoils}(object)

\S4method{swSoils_Layers}{swSoils}(object)

\S4method{set_swSoils}{swSoils,swSoils}(object) <- value


\S4method{swSoils_Layers}{swSoils}(object) <- value

\S4method{swReadLines}{swSoils,character}(object, file)
}
\arguments{
\item{.Object}{An object of class \code{\linkS4class{swSoils}}.}

\item{...}{Further arguments to methods.}

\item{object}{An object of class \code{\linkS4class{swSoils}}.}

\item{value}{A value to assign to a specific slot of the object.}

\item{file}{A character string. The file name from which to read.}
}
\description{
The methods listed below work on this class and the proper slot of the class
  \code{\linkS4class{swInputData}}.
}
\examples{
showClass("swSoils")
x <- new("swSoils")

}
\seealso{
\code{\linkS4class{swInputData}} \code{\linkS4class{swFiles}}
\code{\linkS4class{swWeather}} \code{\linkS4class{swCloud}}
\code{\linkS4class{swMarkov}} \code{\linkS4class{swProd}}
\code{\linkS4class{swSite}} \code{\linkS4class{swInputData}}
\code{\linkS4class{swEstab}} \code{\linkS4class{swOUT}}
\code{\linkS4class{swSWC}} \code{\linkS4class{swLog}}
}
