% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swSoils_Layers<-}
\alias{swSoils_Layers<-}
\title{swSoils_Layers<-}
\usage{
swSoils_Layers(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swSoils}} or \code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
swSoils_Layers<-
}
\seealso{
\code{\linkS4class{swSoils}} and \code{\linkS4class{swInputData}}
}
