% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R, R/F_swSoils.R,
%   R/K_swContainer.R
\name{swSoils_Layers}
\alias{swSoils_Layers}
\alias{swSoils_Layers<-}
\alias{swSoils_Layers,swSoils-method}
\alias{swSoils_Layers<-,swSoils-method}
\alias{swSoils_Layers,swInputData-method}
\alias{swSoils_Layers<-,swInputData-method}
\title{Interact with the soil layer data frame}
\usage{
swSoils_Layers(object)

swSoils_Layers(object) <- value

\S4method{swSoils_Layers}{swSoils}(object)

\S4method{swSoils_Layers}{swSoils}(object) <- value

\S4method{swSoils_Layers}{swInputData}(object)

\S4method{swSoils_Layers}{swInputData}(object) <- value
}
\arguments{
\item{object}{An object of class \code{\link{swSoils}} or \link[=swInputData-class]{swInputData}.}

\item{value}{An object that can be converted to a data matrix and represents
required soil layer information.}
}
\description{
Interact with the soil layer data frame
}
\section{Details}{

The replacement method \verb{swSoils_Layers<-} for class \link[=swInputData-class]{swInputData}
resizes \code{SWRCp} to match number of new soil layers
(and reset \code{SWRCp} values to \code{NA}) if \code{"has_swrcp"} is \code{FALSE}.
This is to avoid inconsistency between
soil properties and \code{SWRCp}.


The correct sequence for setting values is
\enumerate{
\item \code{\link[=swSoils_Layers]{swSoils_Layers()}},
\item \code{\link[=swSite_SWRCflags]{swSite_SWRCflags()}}, and
\item \code{\link[=swSoils_SWRCp]{swSoils_SWRCp()}} and \code{\link[=swSite_hasSWRCp]{swSite_hasSWRCp()}}
\item \code{\link[=swSoils_omSWRCp]{swSoils_omSWRCp()}} (and \code{\link[=swSite_depthSapric]{swSite_depthSapric()}})
}
}

