% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swSoils_SWRCp<-}
\alias{swSoils_SWRCp<-}
\title{\code{swSoils_SWRCp<-}}
\usage{
swSoils_SWRCp(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swSoils}} or
\code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
\code{swSoils_SWRCp<-}
}
\seealso{
\code{\linkS4class{swSoils}} and \code{\linkS4class{swInputData}}
}
