% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R, R/F_swSoils.R,
%   R/K_swContainer.R
\name{swSoils_SWRCp}
\alias{swSoils_SWRCp}
\alias{swSoils_SWRCp<-}
\alias{swSoils_SWRCp,swSoils-method}
\alias{swSoils_SWRCp<-,swSoils-method}
\alias{swSoils_SWRCp,swInputData-method}
\alias{swSoils_SWRCp<-,swInputData-method}
\title{\code{SWRC} parameters of the mineral soil}
\usage{
swSoils_SWRCp(object)

swSoils_SWRCp(object) <- value

\S4method{swSoils_SWRCp}{swSoils}(object)

\S4method{swSoils_SWRCp}{swSoils}(object) <- value

\S4method{swSoils_SWRCp}{swInputData}(object)

\S4method{swSoils_SWRCp}{swInputData}(object) <- value
}
\arguments{
\item{object}{An object of class \code{\link{swSoils}} or \link[=swInputData-class]{swInputData}.}

\item{value}{An object that can be converted to a data matrix and represents
required `SWRC` parameters.}
}
\value{
A data matrix.
}
\description{
\code{SWRC} parameters of the mineral soil
}
\section{Details}{

The correct sequence for setting values is
\enumerate{
\item \code{\link[=swSoils_Layers]{swSoils_Layers()}},
\item \code{\link[=swSite_SWRCflags]{swSite_SWRCflags()}}, and
\item \code{\link[=swSoils_SWRCp]{swSoils_SWRCp()}} and \code{\link[=swSite_hasSWRCp]{swSite_hasSWRCp()}}
\item \code{\link[=swSoils_omSWRCp]{swSoils_omSWRCp()}} (and \code{\link[=swSite_depthSapric]{swSite_depthSapric()}})
}
}

