% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swSpinup_SpinupDuration<-}
\alias{swSpinup_SpinupDuration<-}
\title{\code{swSpinup_SpinupDuration<-}}
\usage{
swSpinup_SpinupDuration(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swSpinup}} or
\code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
\code{swSpinup_SpinupDuration<-}
}
\seealso{
\code{\linkS4class{swSpinup}} and \code{\linkS4class{swInputData}}
}
