% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swSpinup_SpinupScope}
\alias{swSpinup_SpinupScope}
\title{\code{swSpinup_SpinupScope}}
\usage{
swSpinup_SpinupScope(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swSpinup}} or
\code{\linkS4class{swInputData}}.}
}
\description{
\code{swSpinup_SpinupScope}
}
\seealso{
\code{\linkS4class{swSpinup}} and \code{\linkS4class{swInputData}}
}
