% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_swSWC.R
\docType{class}
\name{swSWC-class}
\alias{swSWC-class}
\alias{initialize,swSWC-method}
\alias{swSWC_use,swSWC-method}
\alias{swSWC_prefix,swSWC-method}
\alias{swSWC_FirstYear,swSWC-method}
\alias{swSWC_Method,swSWC-method}
\alias{swSWC_HistoricList,swSWC-method}
\alias{swSWC_HistoricData,swSWC-method}
\alias{swSWC_use<-,swSWC,logical-method}
\alias{swSWC_prefix<-,swSWC,character-method}
\alias{swSWC_FirstYear<-,swSWC,integer-method}
\alias{swSWC_Method<-,swSWC,integer-method}
\alias{swSWC_HistoricList<-,swSWC,list-method}
\alias{swSWC_HistoricData<-,swSWC,swSWC_hist-method}
\alias{swReadLines,swSWC,character-method}
\title{Class \code{"swSWC"}}
\usage{
\S4method{initialize}{swSWC}(.Object, ...)

\S4method{swSWC_use}{swSWC}(object)

\S4method{swSWC_prefix}{swSWC}(object)

\S4method{swSWC_FirstYear}{swSWC}(object)

\S4method{swSWC_Method}{swSWC}(object)

\S4method{swSWC_HistoricList}{swSWC}(object)

\S4method{swSWC_HistoricData}{swSWC}(object, year)

\S4method{swSWC_use}{swSWC,logical}(object) <- value

\S4method{swSWC_prefix}{swSWC,character}(object) <- value

\S4method{swSWC_FirstYear}{swSWC,integer}(object) <- value

\S4method{swSWC_Method}{swSWC,integer}(object) <- value

\S4method{swSWC_HistoricList}{swSWC,list}(object) <- value

\S4method{swSWC_HistoricData}{swSWC,swSWC_hist}(object) <- value

\S4method{swReadLines}{swSWC,character}(object, file)
}
\arguments{
\item{.Object}{An object of class \code{\linkS4class{swSWC}}.}

\item{...}{Further arguments to methods.}

\item{object}{An object of class \code{\linkS4class{swSWC}}.}

\item{year}{An integer value. The calendar year of the \var{SWC}
\code{data} object.}

\item{value}{A value to assign to a specific slot of the object.}

\item{file}{A character string. The file name from which to read.}
}
\description{
The methods listed below work on this class and the proper slot of the class
  \code{\linkS4class{swInputData}}.
}
\examples{
showClass("swSWC")
x <- new("swSWC")

}
\seealso{
\code{\linkS4class{swInputData}} \code{\linkS4class{swFiles}}
\code{\linkS4class{swWeather}} \code{\linkS4class{swCloud}}
\code{\linkS4class{swMarkov}} \code{\linkS4class{swProd}}
\code{\linkS4class{swSite}} \code{\linkS4class{swSoils}}
\code{\linkS4class{swEstab}} \code{\linkS4class{swOUT}}
\code{\linkS4class{swInputData}} \code{\linkS4class{swLog}}
}
