% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swSWC_HistoricData<-}
\alias{swSWC_HistoricData<-}
\title{\code{swSWC_HistoricData<-}}
\usage{
swSWC_HistoricData(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swSWC}} or \code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
\code{swSWC_HistoricData<-}
}
\seealso{
\code{\linkS4class{swSWC}} and \code{\linkS4class{swInputData}}
}
