% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swSWC_HistoricList}
\alias{swSWC_HistoricList}
\title{swSWC_HistoricList}
\usage{
swSWC_HistoricList(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swSWC}} or \code{\linkS4class{swInputData}}.}
}
\description{
swSWC_HistoricList
}
\seealso{
\code{\linkS4class{swSWC}} and \code{\linkS4class{swInputData}}
}
