% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swSWC_prefix<-}
\alias{swSWC_prefix<-}
\title{\code{swSWC_prefix<-}}
\usage{
swSWC_prefix(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swSWC}} or
\code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
\code{swSWC_prefix<-}
}
\seealso{
\code{\linkS4class{swSWC}} and \code{\linkS4class{swInputData}}
}
