% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSOILWAT2-package.R
\docType{class}
\name{swWeather-class}
\alias{swWeather-class}
\alias{swWeather_DaysRunningAverage<-,swWeather-method}
\alias{swWeather_DaysRunningAverage,swWeather-method}
\alias{swWeather_FirstYearHistorical<-,swWeather-method}
\alias{swWeather_FirstYearHistorical,swWeather-method}
\alias{swWeather_MonScalingParams<-,swWeather-method}
\alias{swWeather_MonScalingParams,swWeather-method}
\alias{swWeather_pct_SnowDrift<-,swWeather-method}
\alias{swWeather_pct_SnowDrift,swWeather-method}
\alias{swWeather_pct_SnowRunoff<-,swWeather-method}
\alias{swWeather_pct_SnowRunoff,swWeather-method}
\alias{swWeather_UseMarkov<-,swWeather-method}
\alias{swWeather_UseMarkov,swWeather-method}
\alias{swWeather_UseSnow<-,swWeather-method}
\alias{swWeather_UseSnow,swWeather-method}
\alias{swClear,swWeather-method}
\alias{swReadLines,swWeather-method}
\title{Class \code{"swWeather"}}
\description{
The generic functions listed work on this and the proper sw Class in the container's slots.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("swWeather", ...)}.
}

\examples{

showClass("swWeather")

}
\references{
https://github.com/Burke-Lauenroth-Lab/rSOILWAT2
}
\seealso{
\code{\linkS4class{swInputData}} \code{\linkS4class{swFiles}}
\code{\linkS4class{swInputData}} \code{\linkS4class{swCloud}}
\code{\linkS4class{swMarkov}} \code{\linkS4class{swProd}}
\code{\linkS4class{swSite}} \code{\linkS4class{swSoils}}
\code{\linkS4class{swEstab}} \code{\linkS4class{swOUT}}
\code{\linkS4class{swSWC}} \code{\linkS4class{swLog}}
}
\author{
Ryan Murphy
}
\keyword{classes}
