% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swWeather_FirstYearHistorical}
\alias{swWeather_FirstYearHistorical}
\title{\code{swWeather_FirstYearHistorical}}
\usage{
swWeather_FirstYearHistorical(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swWeather}} or
\code{\linkS4class{swInputData}}.}
}
\description{
\code{swWeather_FirstYearHistorical}
}
\seealso{
\code{\linkS4class{swWeather}} and \code{\linkS4class{swInputData}}
}
