% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swWeather_pct_SnowDrift}
\alias{swWeather_pct_SnowDrift}
\title{\code{swWeather_pct_SnowDrift}}
\usage{
swWeather_pct_SnowDrift(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swWeather}} or \code{\linkS4class{swInputData}}.}
}
\description{
\code{swWeather_pct_SnowDrift}
}
\seealso{
\code{\linkS4class{swWeather}} and \code{\linkS4class{swInputData}}
}
