% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swWeather_UseMarkovOnly<-}
\alias{swWeather_UseMarkovOnly<-}
\title{\code{swWeather_UseMarkovOnly<-}}
\usage{
swWeather_UseMarkovOnly(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swWeather}} or
\code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
\code{swWeather_UseMarkovOnly<-}
}
\seealso{
\code{\linkS4class{swWeather}} and \code{\linkS4class{swInputData}}
}
