% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsoilwat31-package.R
\docType{class}
\name{swYears-class}
\alias{swClear,swYears-method}
\alias{swReadLines,swYears-method}
\alias{swWriteLines,swYears-method}
\alias{swYears-class}
\alias{swYears_EDOEY,swYears-method}
\alias{swYears_EDOEY<-,swYears-method}
\alias{swYears_EndYear,swYears-method}
\alias{swYears_EndYear<-,swYears-method}
\alias{swYears_FDOFY,swYears-method}
\alias{swYears_FDOFY<-,swYears-method}
\alias{swYears_StartYear,swYears-method}
\alias{swYears_StartYear<-,swYears-method}
\alias{swYears_isNorth,swYears-method}
\alias{swYears_isNorth<-,swYears-method}
\title{Class \code{"swYears"}}
\description{
The generic functions listed work on this and the proper sw Class in the container's slots.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("swYears", ...)}.
}
\examples{

showClass("swYears")

}
\author{
Ryan Murphy
}
\references{
https://github.com/Burke-Lauenroth-Lab/Rsoilwat
}
\seealso{
\code{\linkS4class{swInputData}} \code{\linkS4class{swFiles}}
\code{\linkS4class{swWeather}} \code{\linkS4class{swCloud}}
\code{\linkS4class{swMarkov}} \code{\linkS4class{swProd}}
\code{\linkS4class{swSite}} \code{\linkS4class{swSoils}}
\code{\linkS4class{swEstab}} \code{\linkS4class{swOUT}}
\code{\linkS4class{swSWC}} \code{\linkS4class{swLog}}
}
\keyword{classes}

