% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swYears_FDOFY<-}
\alias{swYears_FDOFY<-}
\title{\code{swYears_FDOFY<-}}
\usage{
swYears_FDOFY(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swYears}} or \code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
\code{swYears_FDOFY<-}
}
\seealso{
\code{\linkS4class{swYears}} and \code{\linkS4class{swInputData}}
}
