% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swYears_isNorth<-}
\alias{swYears_isNorth<-}
\title{\code{swYears_isNorth<-}}
\usage{
swYears_isNorth(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swYears}} or \code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
\code{swYears_isNorth<-}
}
\seealso{
\code{\linkS4class{swYears}} and \code{\linkS4class{swInputData}}
}
