% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swYears_StartYear}
\alias{swYears_StartYear}
\title{\code{swYears_StartYear}}
\usage{
swYears_StartYear(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swYears}} or \code{\linkS4class{swInputData}}.}
}
\description{
\code{swYears_StartYear}
}
\seealso{
\code{\linkS4class{swYears}} and \code{\linkS4class{swInputData}}
}
