% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_OutputDerived_Functions.R
\name{time_columns}
\alias{time_columns}
\title{Output column indices with time information}
\usage{
time_columns(timestep = c("Day", "Week", "Month", "Season", "Year"))
}
\arguments{
\item{timestep}{A character string. One of the \pkg{rSOILWAT2} time steps.}
}
\description{
Output column indices with time information
}
\examples{
time_columns("Month")

}
