% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Vegetation.R
\name{TranspCoeffByVegType}
\alias{TranspCoeffByVegType}
\title{Lookup transpiration coefficients}
\usage{
TranspCoeffByVegType(
  tr_input_code,
  tr_input_coeff,
  soillayer_no,
  trco_type,
  layers_depth,
  adjustType = c("positive", "inverse", "allToLast")
)
}
\arguments{
\item{tr_input_code}{The \code{"desc"} component of \code{\link{sw2_trco_table}}.}

\item{tr_input_coeff}{The \code{"data"} component of \code{\link{sw2_trco_table}}.}

\item{soillayer_no}{An integer value. The number of soil layers.}

\item{trco_type}{A character string. A column name of \code{tr_input_code}.}

\item{layers_depth}{An integer vector. The lower depths of soil layers \code{\link{cm}}}

\item{adjustType}{A character string. The method to adjust prescribed
coefficient profile onto provided depth profile \code{layers_depth}.}
}
\description{
Lookup transpiration coefficients for grasses, shrubs, and trees per soil
layer or per soil depth increment of 1 cm per distribution type for
each simulation run and copy values to \var{\sQuote{datafile.soils}}
}
\details{
\itemize{
\item first row of datafile is label for per soil layer
\var{\dQuote{Layer}} or per soil depth increment of 1 cm
\var{\dQuote{DepthCM}}
\item second row of datafile is source of data
\item the other rows contain the data for each distribution type = columns
}
}
\seealso{
\code{\link[=estimate_PotNatVeg_roots]{estimate_PotNatVeg_roots()}} with example code
}
