% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Time_SimulationWorld.R
\name{update_requested_years}
\alias{update_requested_years}
\title{Check requested years}
\usage{
update_requested_years(
  start_year,
  end_year,
  has_start_year,
  has_end_year,
  temp_call = NULL,
  verbose = FALSE
)
}
\arguments{
\item{start_year}{An integer value. The requested first year to extract
weather data.}

\item{end_year}{An integer value. The requested last year to extract weather
data.}

\item{has_start_year}{An integer value. The available first year of the
weather data.}

\item{has_end_year}{An integer value. The available last year of the weather
data.}

\item{temp_call}{A character string. An identifier of the calling function
used for printing.}

\item{verbose}{A logical value. If \code{TRUE} prints statements if first or
last year were updated.}
}
\value{
A list with two named elements \itemize{ \item \code{start_year} to
  updated first year no smaller than \code{has_start_year} \item
  \code{end_year} to updated last year no larger than \code{has_end_year} }
}
\description{
Check requested years
}
