% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_swContainer.R
\name{upgrade_swInputData}
\alias{upgrade_swInputData}
\title{Attempt to upgrade essential slots if input object is from an older version}
\usage{
upgrade_swInputData(x)
}
\arguments{
\item{x}{An object of class \code{\linkS4class{swInputData}}.}
}
\value{
An object of class \code{\linkS4class{swInputData}},
  potentially upgraded.
}
\description{
Missing slots are added and take the new default values from `SOILWAT2`.
}
