% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_swProd.R
\name{VegTypes}
\alias{VegTypes}
\alias{namesVegTypes}
\alias{mapVegTypes}
\title{Vegetation (plant functional) types}
\usage{
namesVegTypes(
  version = c("v2", "v1"),
  order = c("rSOILWAT2", "SOILWAT2"),
  shortened = FALSE
)

mapVegTypes(request = "2from1", order = c("rSOILWAT2", "SOILWAT2"))
}
\arguments{
\item{version}{A character string indicating the vegetation type version.}

\item{order}{A character string indicating the order of vegetation types
as relevant for \code{SOILWAT2} or \code{rSOILWAT2} (impacts only version \code{"v1"}).}

\item{shortened}{A logical value. Shortened or full vegetation type names
(impacts only version \code{"v1"}).}

\item{request}{A character string defining the request to map one version
of vegetation types from a previous version.}
}
\value{
\code{namesVegTypes()}: a named character vector that contains the
requested vegetation types.

\code{mapVegTypes()}: indices that (approximately) map a new
from the old vegetation types, i.e.,
a named vector where 0 indicates no match.
}
\description{
Vegetation (plant functional) types
}
\section{Details}{

Vegetation type version \code{"v2"} was introduced with \code{rSOILWAT2} \code{v6.5.0}.
}

\section{Notes}{

Names of vegetation type version \code{"v1"} varied and the order of types was
inconsistent.
}

\examples{
namesVegTypes("v1")
namesVegTypes("v1", order = "SOILWAT2")
namesVegTypes("v1", shortened = TRUE)
namesVegTypes("v2")

mapVegTypes()
stopifnot(names(mapVegTypes("2from1")) == namesVegTypes("v2"))

cbind(
  v2 = namesVegTypes("v2")[mapVegTypes("2from1") > 0L],
  v1 = namesVegTypes("v1")[mapVegTypes("2from1")]
)

}
