% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_swWeatherData.R
\name{weather_dataAggFun}
\alias{weather_dataAggFun}
\title{Functions to summarize currently implemented daily weather variables}
\usage{
weather_dataAggFun()
}
\value{
A named vector of functions that summarize
daily weather variables across time.
}
\description{
Functions to summarize currently implemented daily weather variables
}
