% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_swWeatherData.R
\name{weather_dataBounds}
\alias{weather_dataBounds}
\title{List valid ranges of weather variables}
\usage{
weather_dataBounds()
}
\value{
A named list of length-two numeric vectors.
}
\description{
List valid ranges of weather variables
}
