% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_swWeatherData.R
\name{weather_dataColumns}
\alias{weather_dataColumns}
\title{List names of currently implemented daily weather variables}
\usage{
weather_dataColumns()
}
\value{
A vector of daily weather variable names.
}
\description{
List names of currently implemented daily weather variables
}
