% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swWeatherGenerator.R
\name{weatherGenerator_dataColumns}
\alias{weatherGenerator_dataColumns}
\title{List daily weather variables incorporated in the weather generator}
\usage{
weatherGenerator_dataColumns()
}
\description{
List daily weather variables incorporated in the weather generator
}
