% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{activate_swOUT_OutKey-set}
\alias{activate_swOUT_OutKey-set}
\alias{activate_swOUT_OutKey}
\alias{deactivate_swOUT_OutKey}
\alias{deactivate_swOUT_OutKey-set}
\alias{activate_swOUT_OutKey<-}
\alias{deactivate_swOUT_OutKey<-}
\title{Activate/deactivate an output slot (\var{swOUT_OutKey})}
\usage{
activate_swOUT_OutKey(object) <- value

deactivate_swOUT_OutKey(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swOUT}} or
\code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
Activate/deactivate an output slot (\var{swOUT_OutKey})
}
\seealso{
\code{\linkS4class{swOUT}}, \code{\linkS4class{swInputData}}, and
  \code{\link{sw_exec}}
}
