% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{check_content}
\alias{check_content}
\alias{dbW_has_sites}
\alias{dbW_has_siteIDs}
\alias{dbW_has_scenarioIDs}
\alias{dbW_has_scenarios}
\alias{dbW_has_weatherData}
\alias{dbW_have_sites_all_weatherData}
\title{Check availability of content in registered weather database}
\usage{
dbW_has_sites(Labels, ignore.case = FALSE)

dbW_has_siteIDs(Site_ids)

dbW_has_scenarioIDs(Scenario_ids)

dbW_has_scenarios(Scenarios, ignore.case = FALSE)

dbW_has_weatherData(Site_ids, Scenario_ids)

dbW_have_sites_all_weatherData(
  site_labels = NULL,
  site_ids = NULL,
  scen_labels = NULL,
  scen_ids = NULL,
  chunk_size = 1500L,
  verbose = FALSE
)
}
\arguments{
\item{Labels}{A vector of character strings. The names/labels of the queried
sites.}

\item{ignore.case}{A logical value.}

\item{Site_ids}{An integer vector. The IDs/database keys of the queried site.}

\item{Scenario_ids}{An integer vector. The IDs/database keys of the queried
scenario.}

\item{Scenarios}{A vector of character strings. The names/labels of the
queried scenarios.}
}
\value{
\code{dbW_has_siteIDs} returns a logical vector of the length of
  queried sites.

\code{dbW_has_siteIDs} returns a logical vector of the length of
  queried sites.

\code{dbW_has_scenarios} returns a logical vector of the length of
  queried Scenarios.

\code{dbW_has_scenarios} returns a logical vector of the length of
  queried Scenarios.

\code{dbW_has_weatherData} returns a logical matrix with rows
  corresponding to queried sites and columns to queried scenarios.

\code{dbW_have_sites_all_weatherData} returns a logical vector
  of length of queried sites;
  a \code{TRUE} element indicates that weather data
  for each queried scenarios is available for that queried site.
}
\description{
Check availability of content in registered weather database
}
\section{Details}{
 \code{dbW_has_siteIDs} checks whether sites are available.

 \code{dbW_has_siteIDs} checks whether sites are available.

 \code{dbW_has_scenarioIDs} checks whether scenarios are
  available.

 \code{dbW_has_scenarios} checks whether scenarios are
  available.

 \code{dbW_has_weatherData} checks whether weather data are
  available but ignores \code{start_year} and \code{end_year}.


  \code{dbW_have_sites_all_weatherData} checks whether weather data are
  available but ignores \code{start_year} and \code{end_year}.
}

