% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Pedotransfer_Functions.R
\name{check_swrcp}
\alias{check_swrcp}
\title{Check Soil Water Retention Curve parameters}
\usage{
check_swrcp(swrc_name, swrcp)
}
\arguments{
\item{swrc_name}{An character string or vector.
The selected \code{SWRC} name
(one of \code{\link[=swrc_names]{swrc_names()}}, with default \code{"Campbell1974"}).}

\item{swrcp}{A numeric vector or matrix.
The parameters of a selected \code{SWRC};
each row represents one \code{SWRC}, e.g., one per soil layer.}
}
\description{
Check Soil Water Retention Curve parameters
}
\section{Notes}{

The argument selecting \code{SWRC} (\code{swrc_name}) is recycled
for multiple parameter sets, i.e., rows of \code{swrcp}.
}

\section{Details}{

\code{\link[=swrc_names]{swrc_names()}} lists implemented \code{SWRCs}.
}

\examples{
swrc_name <- "Campbell1974"
ptf_name <- "Cosby1984AndOthers"
swrcp <- ptf_estimate(
  sand = c(0.5, 0.3),
  clay = c(0.2, 0.1),
  fcoarse = c(0, 0),
  swrc_name = swrc_name,
  ptf_name = ptf_name
)

check_swrcp(swrc_name, swrcp)
check_swrcp(swrc_name, swrcp[1, ])

swrcp2 <- swrcp
swrcp2[1, 1] <- -10
check_swrcp(swrc_name, swrcp2)

}
\seealso{
\code{\link[=ptf_estimate]{ptf_estimate()}}
}
