% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_upgrade.R
\name{check_updatedDB}
\alias{check_updatedDB}
\title{Check an updated weather database}
\usage{
check_updatedDB(con)
}
\arguments{
\item{con}{A connection to an updated weather database.}
}
\description{
Check an updated weather database
}
