% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{check_version}
\alias{check_version}
\title{Check that version of an input or output object is up-to-date}
\usage{
check_version(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swInputData}} or
\code{\linkS4class{swOutput}}.}
}
\value{
A logical value.
}
\description{
Check that version of an input or output object is up-to-date
}
