% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_addFromFolders}
\alias{dbW_addFromFolders}
\title{Read \code{SOILWAT2}-style weather data from disk and store in weather database}
\usage{
dbW_addFromFolders(
  MetaData = NULL,
  FoldersPath = ".",
  ScenarioName = "Current",
  weather_tag = "weath"
)
}
\arguments{
\item{MetaData}{A data frame. If not missing, then must contain columns
(in that order) (name of site weather data) \code{folder},
\code{lat} (site latitude),
\code{long} (site longitude),
\code{label} (name of site),
\code{elevation} (site elevation).}

\item{FoldersPath}{A character string. The path to the folder that contains
the site weather data folders.}

\item{ScenarioName}{A character string. The scenario name represented by
the data.}

\item{weather_tag}{A character string. The file name tag that identifies
the individual weather data files inside the site weather data folders.}
}
\description{
Read \code{SOILWAT2}-style weather data from disk and store in weather database
}
