% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSOILWAT2_deprecated.R
\name{dbW_addSite}
\alias{dbW_addSite}
\title{Add one site to the active weather database}
\usage{
dbW_addSite(Site_id = NULL, lat = NULL, long = NULL, Label = NULL)
}
\arguments{
\item{Site_id}{An integer value. The identification number of the site.}

\item{lat}{A numeric value. The latitude of the site.}

\item{long}{A numeric value. The longitude of the site.}

\item{Label}{A character string. The name of the site.}
}
\description{
Add one site to the active weather database
}
