% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_addWeatherData}
\alias{dbW_addWeatherData}
\title{Adds daily weather data to a registered weather database}
\usage{
dbW_addWeatherData(
  Site_id = NULL,
  lat = NULL,
  long = NULL,
  tol_xy = 1e-04,
  weatherFolderPath = NULL,
  weatherData = NULL,
  Label = NULL,
  Scenario_id = NULL,
  Scenario = "Current",
  weather_tag = "weath",
  ignore.case = FALSE,
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{Site_id}{Numeric. Used to identify site and extract weather data.}

\item{lat}{Numeric. Latitude used with longitude to identify site id if
\code{Site_id} is missing.}

\item{long}{Numeric. Longitude and Latitude are used to identify site if
\code{Site_id} is missing.}

\item{tol_xy}{A numeric value. The tolerance used to match requested
longitude and latitude values.}

\item{Label}{A character string. A site label.}

\item{Scenario}{A character string.}

\item{ignore.case}{A logical value.}
}
\value{
An invisible logical value indicating success with \code{TRUE} and
 failure with \code{FALSE}.
}
\description{
Adds daily weather data to a registered weather database
}
