% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_delete_duplicated_weatherData}
\alias{dbW_delete_duplicated_weatherData}
\title{Remove duplicated weather data records}
\usage{
dbW_delete_duplicated_weatherData(check_values = TRUE)
}
\arguments{
\item{check_values}{A logical value.}
}
\value{
The number of deleted records
}
\description{
Remove duplicated weather data records
}
\section{Details}{

A weather data record is considered a duplicate if \code{site_id}, \code{scenario_id},
\code{start_year}, and \code{end_year} agree; if \code{check_values} is \code{TRUE}, then
the daily weather values must also exactly agree (checked as blobs)
to be considered a duplicate entry.
}

