% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swFunctions.R
\name{dbW_deleteSiteData}
\alias{dbW_deleteSiteData}
\title{Delete a weather data record from a registered weather database}
\usage{
dbW_deleteSiteData(Site_id, Scenario_id = NULL)
}
\value{
An invisible logical value indicating success with \code{TRUE} and
  failure with \code{FALSE}.
}
\description{
Delete a weather data record from a registered weather database
}
