% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swFunctions.R
\name{dbW_disconnectConnection}
\alias{dbW_disconnectConnection}
\title{Disconnects a SQLite weather database from the package}
\usage{
dbW_disconnectConnection()
}
\value{
An invisible logical value indicating success with \code{TRUE} and
 failure with \code{FALSE}.
}
\description{
Disconnects a SQLite weather database from the package
}
