% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swWeatherGenerator.R
\name{dbW_estimate_WGen_coefs}
\alias{dbW_estimate_WGen_coefs}
\title{Estimate coefficients of the \code{SOILWAT2} weather generator}
\usage{
dbW_estimate_WGen_coefs(
  weatherData,
  WET_limit_cm = 0,
  type_cftemp = c("all", "dry", "wet", "none"),
  propagate_NAs = FALSE,
  valNA = NULL,
  imputation_type = c("none", "mean", "locf"),
  imputation_span = 5L
)
}
\arguments{
\item{weatherData}{A list of elements of class \code{\link{swWeatherData}}
or a data frame as returned by \code{\link[=dbW_weatherData_to_dataframe]{dbW_weatherData_to_dataframe()}}.}

\item{WET_limit_cm}{A numeric value. A day with more precipitation than
this value is considered \code{wet} instead of \code{dry}. Default is 0.
This values should be equal to the corresponding value used in
\code{SOILWAT2}'s function \code{SW_MKV_today()}.}

\item{type_cftemp}{A character string. Indicator for applying
temperature correction factors:
\itemize{
\item \code{"all"}, then correct min/max daily temperature on wet and dry days
(default);
\item \code{"dry"}, then correct min/max daily temperature only on dry days;
\item \code{"wet"}, then correct min/max daily temperature only on wet days;
\item \code{"none"}, then do not correct min/max daily temperature for
wet/dry days.
}}

\item{propagate_NAs}{A logical value. If \code{TRUE}, then missing weather
values in the input \code{weatherData} are excluded; if \code{FALSE}, then
missing values are propagated to the estimation. See Details.}

\item{valNA}{The (numerical) value of missing weather data.
If \code{NULL}, then default values are interpreted as missing.}

\item{imputation_type}{A text string. Passed to \code{\link[rSW2utils:impute_df]{rSW2utils::impute_df()}}
for imputation of missing values in estimated weather generator coefficients.}

\item{imputation_span}{An integer value. Passed to \code{\link[rSW2utils:impute_df]{rSW2utils::impute_df()}}
for imputation of missing values in estimated weather generator coefficients.}
}
\value{
A list with two named elements:
\itemize{
\item \code{"mkv_doy"}: A data frame with 366 rows (day of year) and 5 columns:
\itemize{
\item \code{"DOY"}: Rows represent day of year.
\item \code{"p_W_W"}: Probability that \code{DOY} is wet if the previous day
\code{(doy - 1)} was wet.
\item \code{"p_W_D"}: Probability that \code{DOY} is wet if the previous day
\code{(doy - 1)} was dry.
\item \code{"PPT_avg"}: Average amount of precipitation \verb{[cm]} on \code{DOY} if wet.
\item \code{"PPT_sd"}: Standard deviation of amount of precipitation \verb{[cm]}
on \code{DOY} if wet.
}
\item \code{"mkv_woy"}: A data frame with 53 rows \code{SOILWAT2} weeks of year
(i.e., consecutive \code{heptads} of days) and 11 columns
\itemize{
\item \code{"WEEK"}: Rows represent week of year.
\item \code{"wTmax_C"}: Average daily maximum temperature \verb{[C]} for week.
\item \code{"wTmin_C"}: Average daily minimum temperature \verb{[C]} for week.
\item \code{"var_MAX"}: Variance of daily maximum temperature for week.
\item \code{"cov_MAXMIN"}: Covariance of daily maximum and minimum
temperatures for week.
\item \code{"cov_MINMAX"}: Identical to \code{"cov_MAXMIN"}.
\item \code{"var_MIN"}: Variance of daily minimum temperature for week.
\item \code{"CF_Tmax_wet"}: Difference between average daily maximum
temperature \verb{[C]} for wet days of week and \code{"wTmax_C"}.
\item \code{"CF_Tmax_dry"}: Difference between average daily maximum
temperature \verb{[C]} for dry days of week and \code{"wTmax_C"}.
\item \code{"CF_Tmin_wet"}: Same as \code{"CF_Tmax_wet"} but for daily
minimum temperature.
\item \code{"CF_Tmin_dry"}: Same as \code{"CF_Tmax_dry"} but for daily
minimum temperature.
}
}
}
\description{
Estimated coefficients correspond to what is required by the two files
\code{"mkv_covar.in"} and \code{"mkv_prob.in"} for the first-order
Markov weather generator in \code{SOILWAT2} \verb{> v4.2.5}.
}
\section{Notes}{
 This code is a complete overhaul compared to the version
from \href{https://github.com/DrylandEcology/rSFSTEP2/commit/cd9e161971136e1e56d427a4f76062bbb0f3d03a}{\code{rSFSTEP2} on \code{2019-Feb-10}}.

 This function will produce \code{NA}s in the output if there
are insufficient weather observation in the input data \code{weatherData}
for a specific day or week of the year. Such \code{NA}s will cause a
\code{SOILWAT2} run to fail (potentially non-graciously and
with non-obvious error messages). To avoid that, this function offers
imputation approaches in order to fill in those failed coefficient
estimates; see \code{\link[rSW2utils:impute_df]{rSW2utils::impute_df()}}, but please note that
any such imputation likely introduces biases in the generated weather.
}

\section{Details}{

Most users will likely want to set \code{propagate_NAs} to \code{FALSE}.
Note: \code{propagate_NAs} corresponds to \code{!na.rm}
from previous versions of this function with a different default value.
Consider an example: a the 30-year long input \code{weatherData} is
complete except for missing values on Jan 1, 2018.
\itemize{
\item If \code{propagate_NAs} is set to \code{TRUE}, then the
coefficients for day 1 and week 1 of year will be \code{NA} --
despite all the available data. In this case, the missing coefficients
for day 1 and week 1 of year will be imputed.
\item If \code{propagate_NAs} is set to \code{FALSE}, then the coefficients
for day 1 and week 1 of year will be calculated based on the non-missing
values for that day respectively that week of year. No imputation occurs.
}
}

\examples{
res1 <- dbW_estimate_WGen_coefs(rSOILWAT2::weatherData)
wdata <- data.frame(dbW_weatherData_to_dataframe(rSOILWAT2::weatherData))
res2 <- dbW_estimate_WGen_coefs(wdata)

sw_in <- rSOILWAT2::sw_exampleData
swMarkov_Prob(sw_in) <- res2[["mkv_doy"]]
swMarkov_Conv(sw_in) <- res2[["mkv_woy"]]

}
\seealso{
\code{\link[=print_mkv_files]{print_mkv_files()}} to print values to \code{SOILWAT2}
compatible files. \code{\link[=swMarkov_Prob]{swMarkov_Prob()}} and
\code{\link[=swMarkov_Conv]{swMarkov_Conv()}} to extract/replace values in a \code{rSOILWAT2}
input object of class \code{\link{swInputData}}.
}
