% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_getScenarioId}
\alias{dbW_getScenarioId}
\title{Extract table keys to connect scenario(s) with weather data in the registered weather
database}
\usage{
dbW_getScenarioId(Scenario, ignore.case = FALSE, verbose = FALSE)
}
\arguments{
\item{ignore.case}{A logical value.}
}
\value{
An integer vector with the values of the keys or \code{NA} if not located.
}
\description{
Extract table keys to connect scenario(s) with weather data in the registered weather
database
}
