% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_getWeatherData}
\alias{dbW_getWeatherData}
\title{Extracts daily weather data from a registered weather database}
\usage{
dbW_getWeatherData(
  Site_id = NULL,
  lat = NULL,
  long = NULL,
  tol_xy = 1e-04,
  Label = NULL,
  startYear = NULL,
  endYear = NULL,
  Scenario = "Current",
  Scenario_id = NULL,
  ignore.case = FALSE,
  stop_if_missing = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{Site_id}{Numeric. Used to identify site and extract weather data.}

\item{lat}{Numeric. Latitude used with longitude to identify site id if
\code{Site_id} is missing.}

\item{long}{Numeric. Longitude and Latitude are used to identify site if
\code{Site_id} is missing.}

\item{tol_xy}{A numeric value. The tolerance used to match requested
longitude and latitude values.}

\item{Label}{A character string. A site label.}

\item{startYear}{Numeric. Extracted weather data will start with this year.}

\item{endYear}{Numeric. Extracted weather data will end with this year.}

\item{Scenario}{A character string.}

\item{ignore.case}{A logical value.}

\item{stop_if_missing}{A logical value. If \code{TRUE}, then throws an
error if at least one requested weather data object is not available
in the current weather database. If \code{FALSE}, then returns \code{NULL}
for those requested site scenario combinations.}
}
\value{
If one site and one scenario were requested, then returns
 weather data as list. Each element is an object of class
 \code{\linkS4class{swWeatherData}} and contains data for one year.
 If more than one site or more than scenario were requested, then returns
 a list of such weather data lists.
 Elements of the returned list may be \code{NULL} if there is no
 weather data object for the requested site scenario combination and if
 \code{stop_if_missing} is \code{FALSE}.
}
\description{
Weather data for the soil water simulation run can be stored in the input
data or it can be separate to keep the input data size down for multiple
variations of the same site. This function is used to return the weather
data from a predefined weather database. Using the database was faster then
reading in multiple weather files from disk.
}
\details{
If there is missing data, then impute or use the built-in Markov
weather generator (see examples for \code{\link{sw_exec}}).
}
\section{Notes}{

  This function returns the first record of weather data for a
  site x scenario combination even if duplicate entries match the query.
}

\seealso{
\code{\link{getWeatherData_folders}}
}
