% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_getWeatherData}
\alias{dbW_getWeatherData}
\title{Extracts daily weather data from a registered weather database}
\usage{
dbW_getWeatherData(
  Site_id = NULL,
  lat = NULL,
  long = NULL,
  tol_xy = 1e-04,
  Label = NULL,
  startYear = NULL,
  endYear = NULL,
  Scenario = "Current",
  Scenario_id = NULL,
  ignore.case = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{Site_id}{Numeric. Used to identify site and extract weather data.}

\item{lat}{Numeric. Latitude used with longitude to identify site id if
\code{Site_id} is missing.}

\item{long}{Numeric. Longitude and Latitude are used to identify site if
\code{Site_id} is missing.}

\item{tol_xy}{A numeric value. The tolerance used to match requested
longitude and latitude values.}

\item{Label}{A character string. A site label.}

\item{startYear}{Numeric. Extracted weather data will start with this year.}

\item{endYear}{Numeric. Extracted weather data will end with this year.}

\item{Scenario}{A character string.}

\item{ignore.case}{A logical value.}
}
\value{
Returns weather data as list. Each element is an object of class
  \code{\linkS4class{swWeatherData}} and contains data for one year.
}
\description{
Weather data for the soil water simulation run can be stored in the input
data or it can be separate to keep the input data size down for multiple
variations of the same site. This function is used to return the weather
data from a predefined weather database. Using the database was faster then
reading in multiple weather files from disk.
}
\details{
\pkg{SOILWAT2} does not handle missing weather data. If you have missing
data, then you have to impute yourself or use the built-in Markov
weather generator (see examples for \code{\link{sw_exec}}).
}
\seealso{
\code{\link{getWeatherData_folders}}
}
