% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_setConnection}
\alias{dbW_setConnection}
\alias{.dbW_setConnection}
\title{Registers/connects a SQLite weather database with the package}
\usage{
dbW_setConnection(
  dbFilePath,
  create_if_missing = FALSE,
  check_version = FALSE,
  verbose = FALSE
)

.dbW_setConnection(dbFilePath)
}
\arguments{
\item{dbFilePath}{A character string. The weather database file path.}

\item{create_if_missing}{A logical value. If \code{TRUE} and now file
\code{dbFilePath} exists then create a new file.}

\item{check_version}{A logical value. If \code{TRUE} then check database
version against currently implemented version by the package.}

\item{verbose}{A logical value.}
}
\value{
An invisible logical value indicating success/failure.
}
\description{
Registers/connects a SQLite weather database with the package
}
\section{Details}{

\code{.dbW_setConnection()} is a bare-bones version of \code{dbW_setConnection()}.
It doesn't carry out any checks that make sure the database works
correctly.
}

