% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swFunctions.R
\name{dbW_setConnection}
\alias{dbW_setConnection}
\title{Registers/connects a SQLite weather database with the package}
\usage{
dbW_setConnection(dbFilePath, create_if_missing = FALSE,
  check_version = FALSE, verbose = FALSE)
}
\arguments{
\item{dbFilePath}{A character string. The weather database file path.}

\item{create_if_missing}{A logical value. If \code{TRUE} and now file \code{dbFilePath}
exists then create a new file.}

\item{check_version}{A logical value. If \code{TRUE} then check database version
against currently implemented version by the package.}

\item{verbose}{A logical value.}
}
\value{
An invisible logical value indicating success/failure.
}
\description{
Registers/connects a SQLite weather database with the package
}
