% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_temporal_summaries}
\alias{dbW_temporal_summaries}
\alias{dbW_weatherData_to_monthly}
\alias{dbW_dataframe_aggregate}
\alias{dbW_dataframe_to_monthly}
\title{Summarize daily weather to weekly, monthly, or yearly values}
\usage{
dbW_weatherData_to_monthly(
  dailySW,
  na.rm = FALSE,
  valNA = NULL,
  funs = weather_dataAggFun()
)

dbW_dataframe_aggregate(
  dailySW,
  time_step = c("Year", "Month", "Week", "Day"),
  na.rm = FALSE,
  funs = weather_dataAggFun()
)

dbW_dataframe_to_monthly(dailySW, na.rm = FALSE)
}
\arguments{
\item{dailySW}{A list of elements of class \code{\linkS4class{swWeatherData}}
that each hold daily weather data for one calendar year.}

\item{na.rm}{A logical value. Should missing daily values be removed before
calculating monthly temperature and precipitation.}

\item{valNA}{The (numerical) value of missing weather data.
If \code{NULL}, then default values are interpreted as missing.}

\item{funs}{A named vector of functions. The names must match column names
in \code{dailySW} and the function are used to summarize daily weather values.}

\item{time_step}{A character string.}
}
\description{
Summarize daily weather to weekly, monthly, or yearly values
}
