% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_weather_to_SOILWATfiles}
\alias{dbW_weather_to_SOILWATfiles}
\title{Conversion: object of class \code{\linkS4class{swWeatherData}} or
data.frame to \pkg{SOILWAT} input text files}
\usage{
dbW_weather_to_SOILWATfiles(path, site.label, weatherData = NULL,
  weatherDF = NULL, years = NULL, weatherDF_dataColumns = c("DOY",
  "Tmax_C", "Tmin_C", "PPT_cm"))
}
\description{
Conversion: object of class \code{\linkS4class{swWeatherData}} or
data.frame to \pkg{SOILWAT} input text files
}
