% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_weatherData_round}
\alias{dbW_weatherData_round}
\title{Round weather data in a list class \code{\linkS4class{swWeatherData}}}
\usage{
dbW_weatherData_round(
  weatherData,
  digits = 4L,
  weatherDF_dataColumns = weather_dataColumns()
)
}
\arguments{
\item{weatherData}{A list of elements of class \code{\linkS4class{swWeatherData}}
that each hold daily weather data for one calendar year.}

\item{digits}{An integer value. The number of decimal places for rounding
weather values.}

\item{weatherDF_dataColumns}{A vector of character strings. The column
names of \code{weatherDF} in the correct order for \code{SOILWAT2} including
calendar year \code{year} (optional) and day of year \code{DOY}, see
\code{\link[=weather_dataColumns]{weather_dataColumns()}}.}
}
\value{
A list with \code{\linkS4class{swWeatherData}} elements.
}
\description{
Round weather data in a list class \code{\linkS4class{swWeatherData}}
}
\section{Notes}{

\code{weatherDF_dataColumns} lists the columns of \code{weatherData} to be rounded.
}

