% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_weatherData_round}
\alias{dbW_weatherData_round}
\title{Round weather data in a list class \code{\linkS4class{swWeatherData}}}
\usage{
dbW_weatherData_round(
  weatherData,
  digits = 4L,
  weatherDF_dataColumns = c("Tmax_C", "Tmin_C", "PPT_cm")
)
}
\arguments{
\item{weatherData}{A list with \code{\linkS4class{swWeatherData}} elements.}

\item{digits}{An integer value. The returned values will be rounded to
the specified number of decimal places.}

\item{weatherDF_dataColumns}{A character vector. The column names
that should be rounded.}
}
\value{
A list with \code{\linkS4class{swWeatherData}} elements.
}
\description{
Round weather data in a list class \code{\linkS4class{swWeatherData}}
}
