% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{dbW_weatherData_to_dataframe}
\alias{dbW_weatherData_to_dataframe}
\title{Convert an object of class \code{\linkS4class{swWeatherData}} to a data.frame}
\usage{
dbW_weatherData_to_dataframe(weatherData, valNA = NULL)
}
\arguments{
\item{valNA}{The (numerical) value of missing weather data.
If \code{NULL}, then default values are interpreted as missing.}
}
\description{
Convert an object of class \code{\linkS4class{swWeatherData}} to a data.frame
}
