% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swFunctions.R
\name{dbW_weatherData_to_dataframe}
\alias{dbW_weatherData_to_dataframe}
\title{Convert an object of class \code{\linkS4class{swWeatherData}} to a data.frame}
\usage{
dbW_weatherData_to_dataframe(weatherData)
}
\description{
Convert an object of class \code{\linkS4class{swWeatherData}} to a data.frame
}
