% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_OutputAccess_Functions.R
\name{get_derived_output}
\alias{get_derived_output}
\title{Calculate derived variables from outputs}
\arguments{
\item{x}{An object of class \code{\linkS4class{swOutput}}.}

\item{timestep}{A character string. One of the \pkg{rSOILWAT2} time steps.}
}
\description{
Calculate derived variables from outputs
}
