% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_OutputAccess_Functions.R
\name{get_evaporation}
\alias{get_evaporation}
\title{Calculate evaporation from outputs}
\usage{
get_evaporation(x, timestep = c("Day", "Week", "Month", "Year"))
}
\arguments{
\item{x}{An object of class \code{\linkS4class{swOutput}}.}

\item{timestep}{A character string. One of the \pkg{rSOILWAT2} time steps.}
}
\value{
A numeric vector of evaporation [mm] for each time step.
}
\description{
Calculate evaporation from outputs
}
\examples{
sw_out <- sw_exec(inputData = rSOILWAT2::sw_exampleData)
get_evaporation(sw_out, "Month")

}
