% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_OutputDerived_Functions.R
\name{get_soilmoisture}
\alias{get_soilmoisture}
\title{Extract or calculate soil moisture}
\usage{
get_soilmoisture(
  x,
  timestep = c("Day", "Week", "Month", "Year"),
  type = c("swc", "vwc_bulk", "vwc_matric"),
  swInput = NULL,
  widths_cm = NULL,
  fcoarse = NULL,
  keep_time = FALSE
)
}
\arguments{
\item{x}{An object of class \code{\linkS4class{swOutput}}.}

\item{timestep}{A character string. One of the \pkg{rSOILWAT2} time steps.}

\item{type}{A character string selecting type of soil moisture.}

\item{swInput}{An object of class \link[=swInputData-class]{swInputData}.}

\item{widths_cm}{A numeric vector of soil layer widths (units \verb{[cm]}).}

\item{fcoarse}{A numeric vector of coarse fragments per soil layer
(units \verb{[volume fraction]}).}

\item{keep_time}{A logical value. Include time information in the returned
object.}
}
\value{
A data frame with requested soil moisture;
rows represent time steps and columns represent soil layers.
}
\description{
Extract or calculate soil moisture
}
\section{Details}{

Information on soil layer \code{widths} and coarse fragments \code{fcoarse}
are only used if requested type of soil moisture is
not available and has to be calculated from a different type.
\code{widths} and \code{fcoarse} may be provided directly or via \code{swInput}
from which the information is extracted (see examples).
}

\examples{
sw_in <- rSOILWAT2::sw_exampleData

sw_out <- sw_exec(inputData = sw_in)
res1 <- get_soilmoisture(sw_out, "Month", type = "swc")

deactivate_swOUT_OutKey(sw_in) <- sw_out_flags()[["sw_swcbulk"]]
sw_out <- sw_exec(inputData = sw_in)
res2 <- get_soilmoisture(sw_out, "Month", type = "swc", swInput = sw_in)
all.equal(res1, res2)

res3 <- get_soilmoisture(
  sw_out,
  timestep = "Month",
  type = "swc",
  widths = diff(c(0., swSoils_Layers(sw_in)[, "depth_cm"])),
  fcoarse = swSoils_Layers(sw_in)[, "gravel_content"]
)
all.equal(res1, res3)

}
