% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{get_timestamp}
\alias{get_timestamp}
\alias{get_timestamp,ANY-method}
\title{Retrieve time stamp of an object}
\usage{
get_timestamp(object)

\S4method{get_timestamp}{ANY}(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swInputData}} or
\code{\linkS4class{swOutput}}.}
}
\description{
Retrieve time stamp of an object
}
\examples{
get_timestamp(rSOILWAT2::sw_exampleData)
get_timestamp(sw_exec(rSOILWAT2::sw_exampleData))

}
\seealso{
\code{\link{format_timestamp}}
}
