% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R, R/K_swContainer.R,
%   R/L_swOutput.R
\name{get_version}
\alias{get_version}
\alias{get_version,ANY-method}
\alias{get_version,swInputData-method}
\alias{get_version,swOutput-method}
\title{Retrieve version of \pkg{rSOILWAT2} that was used to create object}
\usage{
get_version(object)

\S4method{get_version}{ANY}(object)

\S4method{get_version}{swInputData}(object)

\S4method{get_version}{swOutput}(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swInputData}} or
\code{\linkS4class{swOutput}}.}
}
\description{
Retrieve version of \pkg{rSOILWAT2} that was used to create object
}
\examples{
get_version(rSOILWAT2::sw_exampleData)
get_version(sw_exec(rSOILWAT2::sw_exampleData))

}
\seealso{
\code{\link{check_version}}
}
