% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swFunctions.R
\name{get_years_from_weatherDF}
\alias{get_years_from_weatherDF}
\title{Assign years to weather data.frame}
\usage{
get_years_from_weatherDF(weatherDF, years, weatherDF_dataColumns)
}
\arguments{
\item{weatherDF}{A data.frame. data.frame containing weather information for site.}

\item{years}{A numeric or integer vector or \code{NULL}. Vector of year data where length is equal to either the number of years in the weather data.frame OR the number of rows in the data.frame.}

\item{weatherDF_dataColumns}{A vector of string values. Column names of the weather data.frame.}
}
\value{
A named list of length 2.
\itemize{
\item \code{years} a vector of unique year values.
 \item \code{year_ts} a vector of time series values for each row/day of the data.frame.
}
}
\description{
Assign years to weather data.frame
}
