% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Pedotransfer_Functions.R
\name{list_matched_swrcs_pdfs}
\alias{list_matched_swrcs_pdfs}
\title{Matching pairs of implemented `SWRCs` and `PDFs`}
\usage{
list_matched_swrcs_pdfs(swrc_name)
}
\arguments{
\item{swrc_name}{An character string or vector.
The selected \code{SWRC} name
(one of \code{\link[=swrc_names]{swrc_names()}}, with default \code{"Campbell1974"}).}
}
\value{
A `data.frame` with two columns `SWRC` and `PDF` where each
  row contains a matching pair of `SWRC` and `PDF` that are implemented.
}
\description{
Matching pairs of implemented `SWRCs` and `PDFs`
}
\section{Details}{

  The argument `swrc_name` is optional. If missing, then all implemented
  `SWRCs` are listed.
  `"NoPDF` is not included in the list.
}

