% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_OutputDerived_Functions.R
\name{nrow_output}
\alias{nrow_output}
\title{Number of time steps in output}
\usage{
nrow_output(x, timestep = c("Day", "Week", "Month", "Year"))
}
\arguments{
\item{x}{An object of class \code{\linkS4class{swOutput}}.}

\item{timestep}{A character string. One of the \pkg{rSOILWAT2} time steps.}
}
\description{
Number of time steps in output
}
\examples{
nrow_output(sw_exec(rSOILWAT2::sw_exampleData), "Month")

}
