% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swWeatherGenerator.R
\name{print_mkv_files}
\alias{print_mkv_files}
\title{Print Markov weather generator files as required by \var{SOILWAT2}}
\usage{
print_mkv_files(mkv_doy, mkv_woy, path, digits = 5)
}
\arguments{
\item{mkv_doy}{A data.frame. The same named output element of
\code{\link{dbW_estimate_WGen_coefs}}.}

\item{mkv_woy}{A data.frame. The same named output element of
\code{\link{dbW_estimate_WGen_coefs}}.}

\item{path}{A character string. The path on disk to the location
where output files should be created.}

\item{digits}{An integer value. The number of digits with which to write
the values to disk.}
}
\value{
This function is called for its side effect, i.e., writing values
  to files \var{\dQuote{mkv_prob.in}} and \var{\dQuote{mkv_covar.in}}.
}
\description{
Print Markov weather generator files as required by \var{SOILWAT2}
}
\examples{
path <- tempdir()

res <- dbW_estimate_WGen_coefs(rSOILWAT2::weatherData)
print_mkv_files(
  mkv_doy = res[["mkv_doy"]],
  mkv_woy = res[["mkv_woy"]],
  path = path)

## Cleanup
unlink(list.files(path), force = TRUE)

}
\seealso{
\code{\link{dbW_estimate_WGen_coefs}} to
  calculate the necessary values based on daily weather data.
}
