% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{set_swProd}
\alias{set_swProd}
\title{\code{set_swProd}
Need to define and export this generic method -- otherwise,
\code{\link{set_swProd<-}} doesn't work.}
\usage{
set_swProd(object, value)
}
\description{
\code{set_swProd}
Need to define and export this generic method -- otherwise,
\code{\link{set_swProd<-}} doesn't work.
}
