% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{set_swSite}
\alias{set_swSite}
\title{set_swSite
Need to define and export this generic method -- otherwise, \code{\link{set_swSite<-}} doesn't work.}
\usage{
set_swSite(object, value)
}
\description{
set_swSite
Need to define and export this generic method -- otherwise, \code{\link{set_swSite<-}} doesn't work.
}
