% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{set_swSite}
\alias{set_swSite}
\title{\code{set_swSite}}
\usage{
set_swSite(object, value)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swSite}} or
\code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
\code{set_swSite}
}
