% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{set_swSWC}
\alias{set_swSWC}
\title{set_swSWC
Need to define and export this generic method -- otherwise, \code{\link{set_swSWC<-}} doesn't work.}
\usage{
set_swSWC(object, value)
}
\description{
set_swSWC
Need to define and export this generic method -- otherwise, \code{\link{set_swSWC<-}} doesn't work.
}
