% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{set_swYears}
\alias{set_swYears}
\title{\code{set_swYears}}
\usage{
set_swYears(object, value)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swYears}} or
\code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
\code{set_swYears}
}
