% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSOILWAT2_deprecated.R
\name{sw_dailyC4_TempVar}
\alias{sw_dailyC4_TempVar}
\title{Calculate variables required to estimate percent C4 species in North America}
\usage{
sw_dailyC4_TempVar(dailyTempMin, dailyTempMean, simTime2)
}
\arguments{
\item{dailyTempMin}{A numeric vector. Time series of daily minimum air
temperature \verb{[C]}.}

\item{dailyTempMean}{A numeric vector. Time series of daily mean air
temperature \verb{[C]}.}

\item{simTime2}{A list with two named elements. The elements are numeric
vectors \var{month_ForEachUsedDay_NSadj} and
\var{year_ForEachUsedDay_NSadj}; they are calculated internally
if \code{NULL}; alternatively, they can be generated by a call to the
function \code{\link[rSW2data]{simTiming_ForEachUsedTimeUnit}}.}
}
\value{
A named numeric vector of length 6.
}
\description{
Calculate variables required to estimate percent C4 species in North America
}
\references{
Teeri J.A., Stowe L.G. (1976) Climatic patterns and the
distribution of C4 grasses in North America. Oecologia, 23, 1-12.
}
