% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsw.R
\name{sw_exec}
\alias{sw_exec}
\title{Execute a SOILWAT simulation run}
\usage{
sw_exec(inputData = NULL, weatherList = NULL, dir = "",
  files.in = "files.in", echo = FALSE, quiet = FALSE)
}
\arguments{
\item{inputData}{an object of the S4 class \code{swInputData} which is
generated from \code{\link{sw_inputData}} or
\code{\link{sw_inputDataFromFiles}}.}

\item{weatherList}{a list of weather data generated via
\code{\link{dbW_getWeatherData}} or \code{\link{getWeatherData_folders}}.}

\item{dir}{a character vector that represents the path to the input data.
Use with files.in}

\item{files.in}{a character vector that represents the partial path of the
'files.in' file}

\item{echo}{logical. This option will echo the inputs to the Log Data.
Helpful for debugging.}

\item{quiet}{logical. Quiet mode doesn't print messages to the logfile.}
}
\value{
A S4 Object Containing 28 outputkey slots.  Each of those has slots
for Yr Mo Wk Dy.

NOTE: WTHR ALLH2O ET ALLVEG do not work.

The output also contains slots for the number of rows for each time period.
Individual outputKeys also have slots for time periods used. Title from
outputsetup. They also have the number of data columns.

These elements contain the output data from soilwat based on the
outputSetup.  Output columns should all be labeled.
}
\description{
Run the simulation and get the output data.  Executes Soil Water simulator
and returns Soil Water data via List.  Uses .Call to pass data to/from C
library.
}
\details{
The input data for a simulation run can be passed to the function
\code{\link{sw_exec}} either as \code{swInputData} and \code{weatherList} or
as text files organized in a folder \code{dir} and explained in
\code{files.in}.

The weather data can be a part of the input data S4 class or separate as a
list to reduce the input object size.  The option 'weatherList' if set will
use the weatherList provided and not the inputdata object's weather data if
it has it. Weather data can also be read in when 'dir' and 'files.in' are
set.

SOILWAT does not handle missing data. If you have missing data, then you
have to impute yourself or use the built-in Markov weather generator (see
examples section). If you use the weather generator, then you have to
provide appropriate values for the input (files) 'mkv_covar.in' and
'mkv_prob.in' for your simulation run - currently, SOILWAT does not contain
code to estimate these values.
}
\references{
Bradford, J. B., D. R. Schlaepfer, and W. K. Lauenroth (2014)
Ecohydrology of adjacent sagebrush and lodgepole pine ecosystems: The
consequences of climate change and disturbance.  \emph{Ecosystems}
\bold{17}:590--605.

Schlaepfer, D. R., W. K. Lauenroth, and J. B. Bradford (2012)
Ecohydrological niche of sagebrush ecosystems.  \emph{Ecohydrology}
\bold{5}:453--466.

Parton, W.J. (1978).  \emph{Abiotic section of ELM}. In: Grassland
simulation model (ed. Innis, G.S.).  Springer New York, NY, pp. 31--53.

Sala, O.E., Lauenroth, W.K. & Parton, W.J. (1992) Long-term soil-water
dynamics in the shortgrass steppe.  \emph{Ecology} \bold{73}:1175--1181.
}
\seealso{
\itemize{ \item \code{\link{sw_exec}} for running a simulation
\item \code{\link{sw_inputData}} and \code{\link{sw_inputDataFromFiles}} for
data input \item \code{\link{dbW_getWeatherData}} and
\code{\link{getWeatherData_folders}} for weather data input }
}
\author{
Ryan Murphy
}
