% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsw.R
\name{sw_inputData}
\alias{sw_inputData}
\title{Rsoilwat sw_inputData}
\usage{
sw_inputData()
}
\value{
Returns a S4 class of type swInputData. This is a container for the
input S4 objects used to run soilwat. The data for a test project is loaded.
}
\description{
Default data Set for SoilWat.  Use this as a template or for testing the
package.
}
\details{
swInputData Object is a S4 object containing slots for all input data.
Calling sw_inputData() generates a new swInputData object with test project
data loaded into them. This can be used for testing or as a template to
build another project off of.

The data returned can be directly used to run the simulation. Use the
function sw_exec and pass it the output from this function to run SW.

\code{showMethods(class="swInputData")} # Will show all generic functions
that be applied to the inputData. These include getters/setters for slots.

swInputData consists of slots for each file that is read in. These slots can
be accessed via the following functions: \tabular{ll}{ \code{get_Markov}
\tab #get markov prop and conv\cr \code{get_swCloud} \tab #get cloud\cr
\code{get_swFiles} \tab #get files.in\cr \code{get_swOUT} \tab #get
outpusetup.in\cr \code{get_swProd} \tab #get prod.in\cr \code{get_swSite}
\tab #get site.in\cr \code{get_swSoils} \tab #get soils.in\cr
\code{get_swSWC} \tab #get swcSetup.in\cr \code{get_swWeather} \tab #get
weatherSetup.in\cr \code{get_swWeatherData} \tab #get individual year\cr
\code{get_swYears} \tab #get years.in data object\cr
\code{get_WeatherHistory} \tab #get S4 of WeatherHistory\cr }

generic functions to get/set individual elements follow a format: 'sw' +
file name + '_' + option, e.g.  \itemize{ \item \code{swFiles_Cloud} \item
\code{swProd_Albedo} } Remember tab complete is your friend.

SOILWAT does not handle missing data. If you have missing data, then you
have to impute yourself or use the built-in Markov weather generator (see
examples for \code{\link{sw_exec}}).
}
\examples{

## ------ Simulation with demonstration data ------------
## Access demonstration data (including daily weather forcing)
sw_in <- sw_inputData()

## Slots of the input object of class 'swInputData'
str(sw_in, max.level=2)

## Execute the simulation run
sw_out <- sw_exec(inputData = sw_in)

}
\author{
Ryan Murphy
}
\seealso{
\itemize{ \item \code{\link{sw_exec}} for running a simulation
\item \code{\link{sw_inputData}} and \code{\link{sw_inputDataFromFiles}} for
data input \item \code{\link{dbW_getWeatherData}} and
\code{\link{getWeatherData_folders}} for weather data input }
}

