% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsw.R
\name{sw_inputData}
\alias{sw_inputData}
\title{Default input object for \pkg{rSOILWAT2}. Use this as a template or for
testing the package.}
\usage{
sw_inputData()
}
\value{
An object of class \code{\linkS4class{swInputData}}.
}
\description{
Default input object for \pkg{rSOILWAT2}. Use this as a template or for
testing the package.
}
\examples{

## ------ Simulation with demonstration data ------------
## Access demonstration data (including daily weather forcing)
sw_in <- rSOILWAT2::sw_exampleData

## Slots of the input object of class \code{\linkS4class{swInputData}}
utils::str(sw_in, max.level = 2)

## Execute the simulation run
sw_out <- sw_exec(inputData = sw_in)

}
\seealso{
\itemize{
  \item \code{\link{sw_exec}} for running a simulation
  \item \code{\link{sw_inputData}} and \code{\link{sw_inputDataFromFiles}}
        for data input
  \item \code{\link{dbW_getWeatherData}} and
        \code{\link{getWeatherData_folders}} for weather data input
}
}
