% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsw.R
\name{sw_inputDataFromFiles}
\alias{sw_inputDataFromFiles}
\title{Read simulation input data from files on disk}
\usage{
sw_inputDataFromFiles(dir = "", files.in = "files.in")
}
\arguments{
\item{dir}{A character string. The path to the simulation project directory.}

\item{files.in}{A character string. The file name (and path relative to
\code{dir}) of the \var{files} input file that contains information
about the remaining input files.}
}
\value{
An object of class \code{\linkS4class{swInputData}}.
}
\description{
Read simulation input data from files on disk
}
\examples{

path_demo <- system.file("extdata", "example1", package = "rSOILWAT2")


## ------ Simulation with data prepared beforehand ------------
## Read inputs from files on disk (including daily weather forcing)
sw_in2 <- sw_inputDataFromFiles(dir = path_demo, files.in = "files.in")

## Slots of the input object of \code{\linkS4class{swInputData}}
utils::str(sw_in2, max.level = 2)

## Execute the simulation run
sw_out2 <- sw_exec(inputData = sw_in2)


}
\seealso{
\itemize{
  \item \code{\link{sw_exec}} for running a simulation
  \item \code{\link{sw_inputData}} and \code{\link{sw_inputDataFromFiles}}
        for data input
  \item \code{\link{dbW_getWeatherData}} and
        \code{\link{getWeatherData_folders}} for weather data input.
}
}
