% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSOILWAT2_deprecated.R
\name{sw_outputData}
\alias{sw_outputData}
\title{Return output data}
\usage{
sw_outputData(inputData)
}
\arguments{
\item{inputData}{an object of the \var{S4} class
\code{\linkS4class{swInputData}} which is generated from
\code{\link{sw_inputData}} or \code{\link{sw_inputDataFromFiles}}.}
}
\value{
An object of class \code{\linkS4class{swOutput}}.
}
\description{
Return output data
}
