% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R, R/B_swFiles.R,
%   R/D_swCarbon.R, R/D_swWeather.R, R/D_swWeatherData.R, R/E_swProd.R,
%   R/E_swProd2.R, R/F_swSite.R, R/F_swSoils.R, R/G_swOut.R, R/I_swEstab.R,
%   R/K_swContainer.R
\name{sw_upgrade}
\alias{sw_upgrade}
\alias{sw_upgrade,ANY-method}
\alias{sw_upgrade,swFiles-method}
\alias{sw_upgrade,swCarbon-method}
\alias{sw_upgrade,swMonthlyScalingParams-method}
\alias{sw_upgrade,swWeather-method}
\alias{upgrade_weatherColumns}
\alias{upgrade_weatherDF}
\alias{upgrade_weatherHistory}
\alias{sw_upgrade,swProd-method}
\alias{sw_upgrade,swProd2-method}
\alias{sw_upgrade,swSite-method}
\alias{upgrade_soilLayers}
\alias{sw_upgrade,swSoils-method}
\alias{sw_upgrade,swOUT-method}
\alias{sw_upgrade,swEstabSpecies-method}
\alias{sw_upgrade,swEstab-method}
\alias{sw_upgrade,swInputData-method}
\title{Upgrade a \code{rSOILWAT2}-classed object from an older package version}
\usage{
sw_upgrade(object, verbose = FALSE)

\S4method{sw_upgrade}{ANY}(object, verbose = FALSE)

\S4method{sw_upgrade}{swFiles}(object, verbose = FALSE)

\S4method{sw_upgrade}{swCarbon}(object, verbose = FALSE)

\S4method{sw_upgrade}{swMonthlyScalingParams}(object, verbose = FALSE)

\S4method{sw_upgrade}{swWeather}(object, verbose = FALSE)

upgrade_weatherColumns(
  weatherDF,
  template_weatherColumns = c("Year", "DOY", weather_dataColumns())
)

upgrade_weatherDF(
  weatherDF,
  template_weatherColumns = c("Year", "DOY", weather_dataColumns())
)

upgrade_weatherHistory(object, verbose = FALSE)

\S4method{sw_upgrade}{swProd}(object, verbose = FALSE)

\S4method{sw_upgrade}{swProd2}(object, verbose = FALSE)

\S4method{sw_upgrade}{swSite}(object, verbose = FALSE)

upgrade_soilLayers(
  soilLayers,
  template_soilLayerProperties = soilLayer_dataColumns()
)

\S4method{sw_upgrade}{swSoils}(object, verbose = FALSE)

\S4method{sw_upgrade}{swOUT}(object, verbose = FALSE)

\S4method{sw_upgrade}{swEstabSpecies}(object, verbose = FALSE)

\S4method{sw_upgrade}{swEstab}(object, verbose = FALSE)

\S4method{sw_upgrade}{swInputData}(object, verbose = FALSE)
}
\arguments{
\item{object}{An object of a \code{rSOILWAT2} class.}

\item{verbose}{A logical value.}

\item{weatherDF}{A data frame with weather variables.}

\item{template_weatherColumns}{A vector with requested weather variables.}

\item{soilLayers}{A two-dimensional object representing soil layers in
rows and soil properties in columns,
see slot \code{"Layers"} of \verb{["swSoils-class"]}.}

\item{template_soilLayerProperties}{A vector of standard names of
soil properties, see \code{\link[=soilLayer_dataColumns]{soilLayer_dataColumns()}}.}
}
\value{
The upgraded \code{object}, if needed, to match the current version
with missing slots and elements filled with default values.

For \code{\link[=upgrade_weatherColumns]{upgrade_weatherColumns()}}:
an updated \code{weatherDF} with requested column name changes.

For \code{\link[=upgrade_weatherDF]{upgrade_weatherDF()}}:
an updated \code{weatherDF} with requested columns.

For \code{\link[=upgrade_soilLayers]{upgrade_soilLayers()}}:
an updated \code{soilLayers} matrix with requested columns
(a new \code{"som_frac"} is initialized to the default value of 0).
}
\description{
Missing slots and elements are added and
take the new default values from \code{SOILWAT2}.
}
\section{Details}{

List of changes:
\itemize{
\item Changes with \code{v6.6.0}:
\itemize{
\item class \code{\link[=swSite-class]{swSite}}: new slots \code{"PotSoilEvCoMethod"},
\code{"RootingProfileMethod"}, \code{"RoughnessLengthGroundSurface"},
\code{"AlbedoMethod"}, \code{"AlbedoSnowMax"}, \code{"AlbedoSoilDry"},
\code{"AlbedoSoilSaturated"}, \code{"AlbedoSoilDarkeningParameter"},
\code{"SnowFractionalCoverMeltingFactor"}
\item class \code{\link[=swProd2-class]{swProd2}}: new slots \code{"RootProfileParameters"},
\code{"kExtVegAlbedo"}
\item class \code{\link[=swOUT-class]{swOUT}}: one new \code{outkey} \code{"ENERGYAVG"},
for a new total of 35 \code{outkey}
}
\item Changes with \code{v6.5.0}:
\itemize{
\item class \code{\link[=swSite-class]{swSite}}: new slot \code{"SoilTemperatureBoundaryMethod"}
\item class \code{\link[=swProd-class]{swProd}}: new slots \code{"nYearsDynamicShort"} and
\code{"nYearsDynamicLong"}
\item new class \code{\link[=swProd2-class]{swProd2}} that supersedes \code{\link[=swProd-class]{swProd}};
an upgrade crosswalks old to the new vegetation types
(retaining old values) if an old \code{"swProd"} slot but no
new \code{"swProd2"} slot is present
\item class \code{\link[=swSoils-class]{swSoils}}: slot \code{"Layers"} reorganized several columns
and replaced rooting profile columns from old with new vegetation
types
\item class \code{\link[=swCarbon-class]{swCarbon}}: new slot \code{"CO2ppmVegRef"}
}
\item Changes with \code{v6.4.0}:
\itemize{
\item class \code{\link[=swProd-class]{swProd}}: new slots \code{"vegYear"} and
\code{"isBiomAsIf100Cover"}
\item class \code{\link[=swWeather-class]{swWeather}}: new slot \code{"correctWeatherValues"}
\item class \code{\link[=swOUT-class]{swOUT}}:
two new \code{outkey}s \code{"DERIVEDSUM"} and \code{"DERIVEDAVG"},
for a new total of 34 \code{outkey}s
}
\item Changes with \code{v6.3.0}:
\itemize{
\item class \code{\link[=swSoils-class]{swSoils}}:
\itemize{
\item new slot \code{"omSWRCp"}; slot \code{"SWRCp"} refers to mineral soil
\item slot \code{"Layers"} gained column \code{"som_frac"}
}
\item class \code{\link[=swSite-class]{swSite}}: new slots \code{"depth_sapric"} and
\code{"SurfaceTemperatureMethod"}
}
\item Changes with \code{v6.2.0}:
\itemize{
\item class \code{\link{swWeatherData}}: slot \code{"data"} changed column name
(\code{"specHavg_pct"} to \code{"specHavg_gPERkg"}) and
units (\code{"\%"} to \code{"g kg-1"}).
}
\item Changes with \code{v6.1.0}:
\itemize{
\item class \code{\link[=swInputData-class]{swInputData}}:
new slot \code{"spinup"} of new class \code{\link[=swSpinup-class]{swSpinup}}
\item class \code{\link[=swFiles-class]{swFiles}}: new total of 27 input files
}
\item Changes with \code{v6.0.0}:
\itemize{
\item class \code{\link[=swSite-class]{swSite}}:
new slots \code{"swrc_flags"}, \code{"has_swrcp"}, and
\code{"SoilDensityInputType"}
\item class \code{\link[=swSoils-class]{swSoils}}: new slot \code{"SWRCp"}
\item class \code{\link[=swFiles-class]{swFiles}}:
\code{SWRC} parameter input file added as file 6 for a new total of 23
\item class \code{\link[=swProd-class]{swProd}}: new slot \code{"veg_method"}
\item class \code{\link{swWeatherData}}: new slots \code{"use_cloudCoverMonthly"},
\code{"use_windSpeedMonthly"}, \code{"use_humidityMonthly"},
\code{"dailyInputFlags"}, and \code{"desc_rsds"}
\item class \code{\link{swWeatherData}}: slot \code{"data"} gained 11 new columns:
\code{"cloudCov_pct"}, \code{"windSpeed_mPERs"}, \code{"windSpeed_east_mPERs"},
\code{"windSpeed_north_mPERs"}, \code{"rHavg_pct"}, \code{"rHmax_pct"},
\code{"rHmin_pct"}, \code{"specHavg_pct"}, \code{"Tdewpoint_C"}, \code{"actVP_kPa"},
and \code{"shortWR"}
}
\item Changes with \code{v5.4.0}:
\itemize{
\item classes \code{\link[=swEstabSpecies-class]{swEstabSpecies}} and \code{\link[=swEstab-class]{swEstab}}:
new slot \code{"vegType"}
}
\item Changes with \code{v5.2.0}:
\itemize{
\item class \code{\link[=swOUT-class]{swOUT}}:
\code{"FROZEN"} added as \code{outkey} 28 for a new total of 32
}
\item Changes with \code{v3.1.2}:
\itemize{
\item class \code{\link[=swOUT-class]{swOUT}}:
\code{"BIOMASS"} added as \code{outkey} 31 for a new total of 31
}
\item Changes with \code{v2.3.0}:
\itemize{
\item class \code{\link[=swOUT-class]{swOUT}}:
\code{"SWA"} added as \code{outkey} 8 for a new total of 30
}
}
}

\examples{
x <- sw_upgrade(rSOILWAT2::sw_exampleData, verbose = TRUE)

upgrade_weatherColumns(
  data.frame(DOY = 1:2, Tmax_C = runif(2), dummy = runif(2))
)
upgrade_weatherColumns(
  data.frame(DOY = 1:2, Tmax_C = runif(2), specHavg_pct = NA)
)

upgrade_weatherDF(
  data.frame(DOY = 1:2, Tmax_C = runif(2), dummy = runif(2))
)
upgrade_weatherDF(
  data.frame(DOY = 1:2, Tmax_C = runif(2), specHavg_pct = NA)
)

upgrade_soilLayers(
  data.frame(
    sand_frac = runif(2),
    clay_frac = runif(2),
    dummy = runif(2),
    transpGrass_frac = runif(2)
  )
)
soils <- slot(rSOILWAT2::sw_exampleData, "soils")
upgrade_soilLayers(slot(soils, "Layers"))
upgrade_soilLayers(slot(soils, "Layers")[, 1:8L, drop = FALSE])

}
