% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R, R/B_swFiles.R,
%   R/F_swSite.R, R/F_swSoils.R, R/G_swOut.R, R/K_swContainer.R
\name{sw_upgrade}
\alias{sw_upgrade}
\alias{sw_upgrade,ANY-method}
\alias{sw_upgrade,swFiles-method}
\alias{sw_upgrade,swSite-method}
\alias{sw_upgrade,swSoils-method}
\alias{sw_upgrade,swOUT-method}
\alias{sw_upgrade,swInputData-method}
\title{Upgrade a \code{rSOILWAT2}-classed object from an older package version}
\usage{
sw_upgrade(object, verbose = FALSE)

\S4method{sw_upgrade}{ANY}(object, verbose = FALSE)

\S4method{sw_upgrade}{swFiles}(object, verbose = FALSE)

\S4method{sw_upgrade}{swSite}(object, verbose = FALSE)

\S4method{sw_upgrade}{swSoils}(object, verbose = FALSE)

\S4method{sw_upgrade}{swOUT}(object, verbose = FALSE)

\S4method{sw_upgrade}{swInputData}(object, verbose = FALSE)
}
\arguments{
\item{object}{An object of a \code{rSOILWAT2} class.}

\item{verbose}{A logical value.}
}
\value{
The upgraded \code{object}, if needed, to match the current version
with missing slots and elements filled with default values.
}
\description{
Missing slots and elements are added and
take the new default values from \code{SOILWAT2}.
}
\section{Details}{

List of changes:
\itemize{
\item Changes with \code{v6.0.0}:
\itemize{
\item class \code{\linkS4class{swSite}}:
new slots \code{"swrc_flags"} and \code{"pdf_for_NoPDF"}
\item class \code{\linkS4class{swSoils}}: new slot \code{"SWRCp"}
\item class \code{\linkS4class{swFiles}}:
\code{SWRC} parameter input file added as file 6 for a new total of 23
}
\item Changes with \code{v5.2.0}:
\itemize{
\item class \code{\linkS4class{swOUT}}:
\code{"FROZEN"} added as \code{outkey} 28 for a new total of 32
}
\item Changes with \code{v3.1.0}:
\itemize{
\item class \code{\linkS4class{swOUT}}:
\code{"BIOMASS"} added as \code{outkey} 31 for a new total of 31
}
\item Changes with \code{v2.3.0}:
\itemize{
\item class \code{\linkS4class{swOUT}}:
\code{"SWA"} added as \code{outkey} 8 for a new total of 30
}
}
}

\examples{
  x <- sw_upgrade(rSOILWAT2::sw_exampleData, verbose = TRUE)

}
