% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_dbW_WeatherDatabase.R
\name{sw_weather_data}
\alias{sw_weather_data}
\title{\code{rSOILWAT2} weather data functionality}
\arguments{
\item{wd}{A list of elements of class \code{\linkS4class{swWeatherData}}
that each hold daily weather data for one calendar year.}

\item{weatherData}{A list of elements of class \code{\linkS4class{swWeatherData}}
that each hold daily weather data for one calendar year.}

\item{dailySW}{A list of elements of class \code{\linkS4class{swWeatherData}}
that each hold daily weather data for one calendar year.}

\item{weatherDF}{A \code{data.frame}. Daily weather data where rows represent
days and columns represent the weather variables
(see \code{weatherDF_dataColumns}).}

\item{weatherDF_dataColumns}{A vector of character strings. The column
names of \code{weatherDF} in the correct order for \code{SOILWAT2} including
calendar year \code{year} (optional) and day of year \code{DOY}, see
\code{\link[=weather_dataColumns]{weather_dataColumns()}}.}

\item{years}{A numeric vector. The calendar years.}

\item{digits}{An integer value. The number of decimal places for rounding
weather values (or \code{TRUE} but no rounding if \code{FALSE} or not finite).}

\item{round}{An integer value. The number of decimal places for rounding
weather values (or \code{TRUE} but no rounding if \code{FALSE} or not finite).}

\item{weather_tag}{A character string. The base file name without extension
for \code{SOILWAT2}-formatted input files; default is \code{"weath"}}
}
\description{
\code{rSOILWAT2} weather data functionality
}
