% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R, R/D_swCarbon.R,
%   R/rSOILWAT2-package.R
\docType{class}
\name{swCarbon-class}
\alias{swCarbon-class}
\alias{swCarbon-class}
\alias{swCarbon-class}
\alias{swCarbon-class}
\alias{swCarbon-class}
\alias{swCarbon-class}
\alias{swCarbon-class}
\alias{swCarbon-class}
\alias{swCarbon-class}
\alias{swCarbon-class}
\alias{swCarbon-class}
\alias{swCarbon-class}
\alias{swCarbon_Use_Bio,swCarbon-method}
\alias{swCarbon_Use_WUE,swCarbon-method}
\alias{swCarbon_Scenario,swCarbon-method}
\alias{swCarbon_DeltaYear,swCarbon-method}
\alias{swCarbon_CO2ppm,swCarbon-method}
\alias{swCarbon_Scenario<-,swCarbon-method}
\alias{swCarbon_DeltaYear<-,swCarbon-method}
\alias{swCarbon_CO2ppm<-,swCarbon-method}
\alias{swCarbon_Use_Bio<-,swCarbon-method}
\alias{swCarbon_Use_WUE<-,swCarbon-method}
\title{Class "swCarbon"}
\usage{
swCarbon_Use_Bio(object)

swCarbon_Use_WUE(object)

swCarbon_Scenario(object)

swCarbon_DeltaYear(object)

swCarbon_CO2ppm(object)

swCarbon_Use_Bio(object) <- value

swCarbon_Use_WUE(object) <- value

swCarbon_Scenario(object) <- value

swCarbon_DeltaYear(object) <- value

swCarbon_CO2ppm(object) <- value
}
\description{
Class \code{swCarbon} defines variables that allow \code{SOILWAT2} to simulate the effects of atmospheric carbon dioxide.

The generic functions listed work on this and the proper sw Class in the container's slots.
}
\section{Slots}{

  \describe{
    \item{\code{CarbonUseBio}}{Object of class \code{"integer"}, where a value of 1 enables the CO2 biomass multiplier.}
    \item{\code{CarbonUseWUE}}{Object of class \code{"integer"}, where a value of 1 enables the CO2 WUE multipler.}
    \item{\code{Scenario}}{Object of class \code{"character"}, that represents the name of the scenario that is being simulated.}
    \item{\code{DeltaYear}}{Object of class \code{"integer"}, that represents the number of years in the future that this simulation is being run.}
    \item{\code{CO2ppm}}{Object of class \code{"matrix"}, that holds years in the first column and CO2 ppm concentrations in the second column.}
  }
}

\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("swCarbon", ...)}.
}

\examples{

showClass("swCarbon")

}
\references{
https://github.com/Burke-Lauenroth-Lab/rSOILWAT2
}
\seealso{
\code{\linkS4class{swInputData}} \code{\linkS4class{swFiles}}
}
\author{
Zach Kramer
}
\keyword{classes}
