% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSOILWAT2-package.R
\docType{class}
\name{swCarbon-class}
\alias{swCarbon-class}
\alias{swCarbon_Use_Bio,swCarbon-method}
\alias{swCarbon_Use_WUE,swCarbon-method}
\alias{swCarbon_Scenario,swCarbon-method}
\alias{swCarbon_DeltaYear,swCarbon-method}
\alias{swCarbon_CO2ppm,swCarbon-method}
\alias{swCarbon_Scenario<-,swCarbon-method}
\alias{swCarbon_DeltaYear<-,swCarbon-method}
\alias{swCarbon_CO2ppm<-,swCarbon-method}
\alias{swCarbon_Use_Bio<-,swCarbon-method}
\alias{swCarbon_Use_WUE<-,swCarbon-method}
\title{Class \code{"swCarbon"}}
\description{
The generic functions listed work on this and the proper sw Class in the container's slots.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("swCarbon", ...)}.
}

\examples{

showClass("swCarbon")

}
\references{
https://github.com/Burke-Lauenroth-Lab/rSOILWAT2
}
\seealso{
\code{\linkS4class{swInputData}} \code{\linkS4class{swFiles}}
}
\author{
Zach Kramer
}
\keyword{classes}
