% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSOILWAT2-package.R
\docType{class}
\name{swCloud-class}
\alias{swCloud-class}
\alias{swCloud_Humidity<-,swCloud-method}
\alias{swCloud_Humidity,swCloud-method}
\alias{swCloud_SkyCover<-,swCloud-method}
\alias{swCloud_SkyCover,swCloud-method}
\alias{swCloud_SnowDensity<-,swCloud-method}
\alias{swCloud_SnowDensity,swCloud-method}
\alias{swCloud_Transmissivity<-,swCloud-method}
\alias{swCloud_Transmissivity,swCloud-method}
\alias{swCloud_WindSpeed<-,swCloud-method}
\alias{swCloud_WindSpeed,swCloud-method}
\alias{swClear,swCloud-method}
\alias{swReadLines,swCloud-method}
\title{Class \code{"swCloud"}}
\description{
The generic functions listed work on this and the proper sw Class in the container's slots.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("swCloud", ...)}.
}

\examples{

showClass("swCloud")

}
\references{
https://github.com/Burke-Lauenroth-Lab/rSOILWAT2
}
\seealso{
\code{\linkS4class{swInputData}} \code{\linkS4class{swFiles}}
\code{\linkS4class{swWeather}} \code{\linkS4class{swInputData}}
\code{\linkS4class{swMarkov}} \code{\linkS4class{swProd}}
\code{\linkS4class{swSite}} \code{\linkS4class{swSoils}}
\code{\linkS4class{swEstab}} \code{\linkS4class{swOUT}}
\code{\linkS4class{swSWC}} \code{\linkS4class{swLog}}
}
\author{
Ryan Murphy
}
\keyword{classes}
