% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSOILWAT2-package.R
\docType{class}
\name{swEstab-class}
\alias{swEstab-class}
\alias{swEstab_useEstab,swEstab-method}
\alias{swEstab_useEstab<-,swEstab-method}
\alias{swClear,swEstab-method}
\alias{swReadLines,swEstab-method}
\title{Class \code{"swEstab"}}
\description{
The generic functions listed work on this and the proper sw Class in the container's slots.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("swEstab", ...)}.
}

\examples{

showClass("swEstab")

}
\references{
https://github.com/Burke-Lauenroth-Lab/rSOILWAT2
}
\seealso{
\code{\linkS4class{swInputData}} \code{\linkS4class{swFiles}}
\code{\linkS4class{swWeather}} \code{\linkS4class{swCloud}}
\code{\linkS4class{swMarkov}} \code{\linkS4class{swProd}}
\code{\linkS4class{swSite}} \code{\linkS4class{swSoils}}
\code{\linkS4class{swInputData}} \code{\linkS4class{swOUT}}
\code{\linkS4class{swSWC}} \code{\linkS4class{swLog}}
}
\author{
Ryan Murphy
}
\keyword{classes}
