% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_swEstab.R
\docType{class}
\name{swEstab-class}
\alias{swEstab-class}
\alias{initialize,swEstab-method}
\alias{swEstab_useEstab,swEstab-method}
\alias{swEstab_useEstab<-,swEstab-method}
\alias{swReadLines,swEstab,character-method}
\title{Class \code{"swEstab"}}
\usage{
\S4method{initialize}{swEstab}(.Object, ...)

\S4method{swEstab_useEstab}{swEstab}(object)

\S4method{swEstab_useEstab}{swEstab}(object) <- value

\S4method{swReadLines}{swEstab,character}(object, file)
}
\arguments{
\item{.Object}{An object of class \code{\linkS4class{swEstab}}.}

\item{...}{Further arguments to methods.}

\item{object}{An object of class \code{\linkS4class{swEstab}}.}

\item{value}{A value to assign to a specific slot of the object.}

\item{file}{A character string. The file name from which to read.}
}
\description{
The methods listed below work on this class and the proper slot of the class
  \code{\linkS4class{swInputData}}.
}
\examples{
showClass("swEstab")
x <- new("swEstab")

}
\seealso{
\code{\linkS4class{swInputData}} \code{\linkS4class{swFiles}}
\code{\linkS4class{swWeather}} \code{\linkS4class{swCloud}}
\code{\linkS4class{swMarkov}} \code{\linkS4class{swProd}}
\code{\linkS4class{swSite}} \code{\linkS4class{swSoils}}
\code{\linkS4class{swInputData}} \code{\linkS4class{swOUT}}
\code{\linkS4class{swSWC}} \code{\linkS4class{swLog}}
}
