% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_swEstab.R
\docType{class}
\name{swEstabSpecies-class}
\alias{swEstabSpecies-class}
\alias{initialize,swEstabSpecies-method}
\alias{swReadLines,swEstabSpecies,character-method}
\title{Class \code{"swEstabSpecies"}}
\usage{
\S4method{initialize}{swEstabSpecies}(.Object, ...)

\S4method{swReadLines}{swEstabSpecies,character}(object, file)
}
\arguments{
\item{.Object}{An object of class \code{\linkS4class{swEstabSpecies}}.}

\item{...}{Further arguments to methods.}

\item{object}{An object of class \code{\linkS4class{swEstabSpecies}}.}

\item{file}{A character string. The file name from which to read.}
}
\description{
The methods listed below work on this class and the proper slot of the class
  \code{\linkS4class{swInputData}}.
}
\examples{
showClass("swEstabSpecies")
x <- new("swEstabSpecies")

}
\seealso{
\code{\linkS4class{swInputData}} \code{\linkS4class{swFiles}}
\code{\linkS4class{swWeather}} \code{\linkS4class{swCloud}}
\code{\linkS4class{swMarkov}} \code{\linkS4class{swProd}}
\code{\linkS4class{swSite}} \code{\linkS4class{swSoils}}
\code{\linkS4class{swEstab}} \code{\linkS4class{swInputData}}
\code{\linkS4class{swSWC}} \code{\linkS4class{swLog}}
}
