% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsoilwat31-package.R
\docType{class}
\name{swFiles-class}
\alias{swFiles-class}
\alias{swFiles_Cloud<-,swFiles-method}
\alias{swFiles_Cloud,swFiles-method}
\alias{swFiles_Estab<-,swFiles-method}
\alias{swFiles_Estab,swFiles-method}
\alias{swFiles_filesIn<-,swFiles-method}
\alias{swFiles_filesIn,swFiles-method}
\alias{swFiles_LogFile<-,swFiles-method}
\alias{swFiles_LogFile,swFiles-method}
\alias{swFiles_MarkovCov<-,swFiles-method}
\alias{swFiles_MarkovCov,swFiles-method}
\alias{swFiles_MarkovProbs<-,swFiles-method}
\alias{swFiles_MarkovProbs,swFiles-method}
\alias{swFiles_OutputPrefix<-,swFiles-method}
\alias{swFiles_OutputPrefix,swFiles-method}
\alias{swFiles_Output<-,swFiles-method}
\alias{swFiles_Output,swFiles-method}
\alias{swFiles_Prod<-,swFiles-method}
\alias{swFiles_Prod,swFiles-method}
\alias{swFiles_ProjDir<-,swFiles-method}
\alias{swFiles_ProjDir,swFiles-method}
\alias{swFiles_SiteParams<-,swFiles-method}
\alias{swFiles_SiteParams,swFiles-method}
\alias{swFiles_Soils<-,swFiles-method}
\alias{swFiles_Soils,swFiles-method}
\alias{swFiles_SWCsetup<-,swFiles-method}
\alias{swFiles_SWCsetup,swFiles-method}
\alias{swFiles_WeatherPrefix<-,swFiles-method}
\alias{swFiles_WeatherPrefix,swFiles-method}
\alias{swFiles_WeatherSetup<-,swFiles-method}
\alias{swFiles_WeatherSetup,swFiles-method}
\alias{swFiles_Years<-,swFiles-method}
\alias{swFiles_Years,swFiles-method}
\alias{swClear,swFiles-method}
\alias{swWriteLines,swFiles-method}
\alias{swReadLines,swFiles-method}
\title{Class \code{"swFiles"}}
\description{
The generic functions listed work on this and the proper sw Class in the container's slots.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("swFiles", ...)}.
}

\examples{

showClass("swFiles")

}
\references{
https://github.com/Burke-Lauenroth-Lab/Rsoilwat
}
\seealso{
\code{\linkS4class{swInputData}} \code{\linkS4class{swYears}}
\code{\linkS4class{swWeather}} \code{\linkS4class{swCloud}}
\code{\linkS4class{swMarkov}} \code{\linkS4class{swProd}}
\code{\linkS4class{swSite}} \code{\linkS4class{swSoils}}
\code{\linkS4class{swEstab}} \code{\linkS4class{swOUT}}
\code{\linkS4class{swSWC}} \code{\linkS4class{swLog}}
}
\author{
Ryan Murphy
}
\keyword{classes}
