% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsoilwat31-package.R
\docType{class}
\name{swFiles-class}
\alias{swFiles-class}
\alias{swFiles_Cloud<-}
\alias{swFiles-method}
\alias{swFiles_Cloud}
\alias{swFiles_Estab<-}
\alias{swFiles_Estab}
\alias{swFiles_filesIn<-}
\alias{swFiles_filesIn}
\alias{swFiles_LogFile<-}
\alias{swFiles_LogFile}
\alias{swFiles_MarkovCov<-}
\alias{swFiles_MarkovCov}
\alias{swFiles_MarkovProbs<-}
\alias{swFiles_MarkovProbs}
\alias{swFiles_OutputPrefix<-}
\alias{swFiles_OutputPrefix}
\alias{swFiles_Output<-}
\alias{swFiles_Output}
\alias{swFiles_Prod<-}
\alias{swFiles_Prod}
\alias{swFiles_ProjDir<-}
\alias{swFiles_ProjDir}
\alias{swFiles_SiteParams<-}
\alias{swFiles_SiteParams}
\alias{swFiles_Soils<-}
\alias{swFiles_Soils}
\alias{swFiles_SWCsetup<-}
\alias{swFiles_SWCsetup}
\alias{swFiles_WeatherPrefix<-}
\alias{swFiles_WeatherPrefix}
\alias{swFiles_WeatherSetup<-}
\alias{swFiles_WeatherSetup}
\alias{swFiles_Years<-}
\alias{swFiles_Years}
\alias{swClear}
\alias{swWriteLines}
\alias{swReadLines}
\title{Class \code{"swFiles"}}
\description{
The generic functions listed work on this and the proper sw Class in the container's slots.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("swFiles", ...)}.
}

\examples{

showClass("swFiles")

}
\references{
https://github.com/Burke-Lauenroth-Lab/Rsoilwat
}
\seealso{
\code{\linkS4class{swInputData}} \code{\linkS4class{swYears}}
\code{\linkS4class{swWeather}} \code{\linkS4class{swCloud}}
\code{\linkS4class{swMarkov}} \code{\linkS4class{swProd}}
\code{\linkS4class{swSite}} \code{\linkS4class{swSoils}}
\code{\linkS4class{swEstab}} \code{\linkS4class{swOUT}}
\code{\linkS4class{swSWC}} \code{\linkS4class{swLog}}
}
\author{
Ryan Murphy
}
\keyword{classes}
