% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSOILWAT2-package.R
\docType{class}
\name{swInputData-class}
\alias{swInputData-class}
\alias{get_Markov}
\alias{swInputData-method}
\alias{get_swCloud}
\alias{get_swFiles}
\alias{get_swOUT}
\alias{get_swProd}
\alias{get_swSite}
\alias{get_swSoils}
\alias{get_swSWC}
\alias{get_swWeatherData}
\alias{get_swWeather}
\alias{get_swYears}
\alias{get_WeatherHistory}
\alias{get_swCarbon,swInputData-method}
\alias{set_swCarbon<-,swInputData-method}
\alias{set_Markov<-}
\alias{set_swCloud<-}
\alias{set_swFiles<-}
\alias{set_swOUT<-}
\alias{set_swProd<-}
\alias{set_swSite<-}
\alias{set_swSoils<-}
\alias{swInputData}
\alias{swSoils-method}
\alias{set_swSWC<-}
\alias{swSWC-method}
\alias{set_swWeatherData<-}
\alias{swWeatherData-method}
\alias{set_swWeather<-}
\alias{set_swYears<-}
\alias{set_WeatherHistory<-}
\alias{list-method}
\alias{swClear}
\alias{swCloud_Humidity<-}
\alias{swCloud_Humidity}
\alias{swCloud_SkyCover<-}
\alias{swCloud_SkyCover}
\alias{swCloud_SnowDensity<-}
\alias{swCloud_SnowDensity}
\alias{swCloud_Transmissivity<-}
\alias{swCloud_Transmissivity}
\alias{swCloud_WindSpeed<-}
\alias{swCloud_WindSpeed}
\alias{swFiles_Cloud<-}
\alias{swFiles_Cloud}
\alias{swFiles_Estab<-}
\alias{swFiles_Estab}
\alias{swFiles_Carbon<-,swInputData-method}
\alias{swFiles_Carbon,swInputData-method}
\alias{swCarbon_Use_Bio,swInputData-method}
\alias{swCarbon_Use_WUE,swInputData-method}
\alias{swCarbon_Scenario,swInputData-method}
\alias{swCarbon_DeltaYear,swInputData-method}
\alias{swCarbon_Scenario<-,swInputData-method}
\alias{swCarbon_DeltaYear<-,swInputData-method}
\alias{swCarbon_Use_Bio<-,swInputData-method}
\alias{swCarbon_Use_WUE<-,swInputData-method}
\alias{swFiles_filesIn<-}
\alias{swFiles_filesIn}
\alias{swFiles_LogFile<-}
\alias{swFiles_LogFile}
\alias{swFiles_MarkovCov<-}
\alias{swFiles_MarkovCov}
\alias{swFiles_MarkovProbs<-}
\alias{swFiles_MarkovProbs}
\alias{swFiles_OutputPrefix<-}
\alias{swFiles_OutputPrefix}
\alias{swFiles_Output<-}
\alias{swFiles_Output}
\alias{swFiles_Prod<-}
\alias{swFiles_Prod}
\alias{swFiles_ProjDir<-}
\alias{swFiles_ProjDir}
\alias{swFiles_SiteParams<-}
\alias{swFiles_SiteParams}
\alias{swFiles_Soils<-}
\alias{swFiles_Soils}
\alias{swFiles_SWCsetup<-}
\alias{swFiles_SWCsetup}
\alias{swFiles_WeatherPrefix<-}
\alias{swFiles_WeatherPrefix}
\alias{swFiles_WeatherSetup<-}
\alias{swFiles_WeatherSetup}
\alias{swFiles_Years<-}
\alias{swFiles_Years}
\alias{swLog_setLine<-}
\alias{swMarkov_Conv<-}
\alias{swMarkov_Conv}
\alias{swMarkov_Prob<-}
\alias{swMarkov_Prob}
\alias{swOUT_OutputSeparator<-}
\alias{swOUT_OutputSeparator}
\alias{swOUT_TimeStep<-}
\alias{swOUT_TimeStep}
\alias{swProd_Albedo<-}
\alias{swProd_Albedo}
\alias{swProd_CanopyHeight<-}
\alias{swProd_CanopyHeight}
\alias{swProd_Composition<-}
\alias{swProd_Composition}
\alias{swProd_Cover_stcr<-}
\alias{swProd_Cover_stcr}
\alias{swProd_CritSoilWaterPotential<-}
\alias{swProd_CritSoilWaterPotential}
\alias{swProd_Es_param_limit<-}
\alias{swProd_Es_param_limit}
\alias{swProd_EsTpartitioning_param<-}
\alias{swProd_EsTpartitioning_param}
\alias{swProd_HydrRedstro<-}
\alias{swProd_HydrRedstro}
\alias{swProd_HydrRedstro_use<-}
\alias{swProd_HydrRedstro_use}
\alias{swProd_LitterInterParam<-}
\alias{swProd_LitterInterParam}
\alias{swProd_MonProd_grass<-}
\alias{swProd_MonProd_grass}
\alias{swProd_MonProd_shrub<-}
\alias{swProd_MonProd_shrub}
\alias{swProd_MonProd_tree<-}
\alias{swProd_MonProd_tree}
\alias{swProd_Shade<-}
\alias{swProd_Shade}
\alias{swProd_VegInterParam<-}
\alias{swProd_VegInterParam}
\alias{swReadLines}
\alias{character-method}
\alias{swSite_DrainageCoefficient<-}
\alias{swSite_DrainageCoefficient}
\alias{swSite_EvapCoefficients<-}
\alias{swSite_EvapCoefficients}
\alias{swSite_IntrinsicSiteParams<-}
\alias{swSite_IntrinsicSiteParams}
\alias{swSite_ModelCoefficients<-}
\alias{swSite_ModelCoefficients}
\alias{swSite_ModelFlags<-}
\alias{swSite_ModelFlags}
\alias{swSite_SnowSimulationParams<-}
\alias{swSite_SnowSimulationParams}
\alias{swSite_SoilTemperatureConsts<-}
\alias{swSite_SoilTemperatureConsts}
\alias{swSite_SoilTemperatureFlag<-}
\alias{swSite_SoilTemperatureFlag}
\alias{swSite_SWClimits<-}
\alias{swSite_SWClimits}
\alias{swSite_TranspCoefficients<-}
\alias{swSite_TranspCoefficients}
\alias{swSite_TranspirationRegions<-}
\alias{swSite_TranspirationRegions}
\alias{swSoils_Layers<-}
\alias{matrix-method}
\alias{swSoils_Layers}
\alias{swSWC_FirstYear<-}
\alias{integer-method}
\alias{swSWC_FirstYear}
\alias{swSWC_HistoricData<-}
\alias{swSWC_hist-method}
\alias{swSWC_HistoricData}
\alias{swSWC_HistoricList<-}
\alias{swSWC_HistoricList}
\alias{swSWC_Method<-}
\alias{swSWC_Method}
\alias{swSWC_prefix<-}
\alias{swSWC_prefix}
\alias{swSWC_use<-}
\alias{logical-method}
\alias{swSWC_use}
\alias{swWeather_DaysRunningAverage<-}
\alias{swWeather_DaysRunningAverage}
\alias{swWeather_FirstYearHistorical<-}
\alias{swWeather_FirstYearHistorical}
\alias{swWeather_MonScalingParams<-}
\alias{swWeather_MonScalingParams}
\alias{swWeather_pct_SnowDrift<-}
\alias{swWeather_pct_SnowDrift}
\alias{swWeather_pct_SnowRunoff<-}
\alias{swWeather_pct_SnowRunoff}
\alias{swWeather_UseMarkov<-}
\alias{swWeather_UseMarkov}
\alias{swWeather_UseSnow<-}
\alias{swWeather_UseSnow}
\alias{swWriteLines}
\alias{swYears_EDOEY<-}
\alias{swYears_EDOEY}
\alias{swYears_EndYear<-}
\alias{swYears_EndYear}
\alias{swYears_FDOFY<-}
\alias{swYears_FDOFY}
\alias{swYears_isNorth<-}
\alias{swYears_isNorth}
\alias{swYears_StartYear<-}
\alias{swYears_StartYear}
\title{Class \code{"swInputData"}}
\description{
This class is a container class to the input file S4 objects. The generic functions
listed work on this and the proper sw classes in the container's slots.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("swInputData", ...)}.
}

\examples{

showClass("swInputData")

}
\references{
https://github.com/Burke-Lauenroth-Lab/rSOILWAT2
}
\seealso{
\code{\linkS4class{swFiles}} \code{\linkS4class{swYears}}
\code{\linkS4class{swWeather}} \code{\linkS4class{swCloud}}
\code{\linkS4class{swMarkov}} \code{\linkS4class{swProd}}
\code{\linkS4class{swSite}} \code{\linkS4class{swSoils}}
\code{\linkS4class{swEstab}} \code{\linkS4class{swOUT}}
\code{\linkS4class{swSWC}} \code{\linkS4class{swLog}}
\code{\linkS4class{swCarbon}}
}
\author{
Ryan Murphy
}
\keyword{classes}
