% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_swContainer.R
\docType{class}
\name{swInputData-class}
\alias{swInputData-class}
\alias{initialize,swInputData-method}
\alias{get_version,swInputData-method}
\alias{get_timestamp,swInputData-method}
\alias{get_swFiles,swInputData-method}
\alias{swFiles_ProjDir,swInputData-method}
\alias{swFiles_filesIn,swInputData-method}
\alias{swFiles_Years,swInputData-method}
\alias{swFiles_LogFile,swInputData-method}
\alias{swFiles_SiteParams,swInputData-method}
\alias{swFiles_Soils,swInputData-method}
\alias{swFiles_WeatherSetup,swInputData-method}
\alias{swFiles_MarkovProbs,swInputData-method}
\alias{swFiles_MarkovCov,swInputData-method}
\alias{swFiles_Cloud,swInputData-method}
\alias{swFiles_Prod,swInputData-method}
\alias{swFiles_Estab,swInputData-method}
\alias{swFiles_Carbon,swInputData-method}
\alias{swFiles_SWCsetup,swInputData-method}
\alias{swFiles_Output,swInputData-method}
\alias{swFiles_WeatherPrefix,swInputData-method}
\alias{swFiles_OutputPrefix,swInputData-method}
\alias{set_swFiles<-,swInputData-method}
\alias{swFiles_ProjDir<-,swInputData-method}
\alias{swFiles_filesIn<-,swInputData-method}
\alias{swFiles_Years<-,swInputData-method}
\alias{swFiles_LogFile<-,swInputData-method}
\alias{swFiles_SiteParams<-,swInputData-method}
\alias{swFiles_Soils<-,swInputData-method}
\alias{swFiles_WeatherSetup<-,swInputData-method}
\alias{swFiles_MarkovProbs<-,swInputData-method}
\alias{swFiles_MarkovCov<-,swInputData-method}
\alias{swFiles_Cloud<-,swInputData-method}
\alias{swFiles_Prod<-,swInputData-method}
\alias{swFiles_Estab<-,swInputData-method}
\alias{swFiles_Carbon<-,swInputData-method}
\alias{swFiles_SWCsetup<-,swInputData-method}
\alias{swFiles_Output<-,swInputData-method}
\alias{swFiles_WeatherPrefix<-,swInputData-method}
\alias{swFiles_OutputPrefix<-,swInputData-method}
\alias{get_swYears,swInputData-method}
\alias{swYears_StartYear,swInputData-method}
\alias{swYears_EndYear,swInputData-method}
\alias{swYears_FDOFY,swInputData-method}
\alias{swYears_EDOEY,swInputData-method}
\alias{swYears_isNorth,swInputData-method}
\alias{set_swYears<-,swInputData-method}
\alias{swYears_StartYear<-,swInputData-method}
\alias{swYears_EndYear<-,swInputData-method}
\alias{swYears_FDOFY<-,swInputData-method}
\alias{swYears_EDOEY<-,swInputData-method}
\alias{swYears_isNorth<-,swInputData-method}
\alias{get_swCloud,swInputData-method}
\alias{set_swCloud<-,swInputData-method}
\alias{get_swWeather,swInputData-method}
\alias{swWeather_DaysRunningAverage,swInputData-method}
\alias{swWeather_FirstYearHistorical,swInputData-method}
\alias{swWeather_pct_SnowDrift,swInputData-method}
\alias{swWeather_pct_SnowRunoff,swInputData-method}
\alias{swWeather_UseMarkov,swInputData-method}
\alias{swWeather_UseSnow,swInputData-method}
\alias{swWeather_MonScalingParams,swInputData-method}
\alias{set_swWeather<-,swInputData-method}
\alias{swWeather_DaysRunningAverage<-,swInputData-method}
\alias{swWeather_FirstYearHistorical<-,swInputData-method}
\alias{swWeather_pct_SnowDrift<-,swInputData-method}
\alias{swWeather_pct_SnowRunoff<-,swInputData-method}
\alias{swWeather_UseMarkov<-,swInputData-method}
\alias{swWeather_UseSnow<-,swInputData-method}
\alias{swWeather_MonScalingParams<-,swInputData-method}
\alias{swCloud_SkyCover,swInputData-method}
\alias{swCloud_WindSpeed,swInputData-method}
\alias{swCloud_Humidity,swInputData-method}
\alias{swCloud_Transmissivity,swInputData-method}
\alias{swCloud_SnowDensity,swInputData-method}
\alias{swCloud_RainEvents,swInputData-method}
\alias{swCloud_SkyCover<-,swInputData-method}
\alias{swCloud_WindSpeed<-,swInputData-method}
\alias{swCloud_Humidity<-,swInputData-method}
\alias{swCloud_Transmissivity<-,swInputData-method}
\alias{swCloud_SnowDensity<-,swInputData-method}
\alias{swCloud_RainEvents<-,swInputData-method}
\alias{get_Markov,swInputData-method}
\alias{swMarkov_Prob,swInputData-method}
\alias{swMarkov_Conv,swInputData-method}
\alias{set_Markov<-,swInputData-method}
\alias{swMarkov_Prob<-,swInputData-method}
\alias{swMarkov_Conv<-,swInputData-method}
\alias{get_WeatherHistory,swInputData-method}
\alias{set_WeatherHistory<-,swInputData,list-method}
\alias{get_swWeatherData,swInputData-method}
\alias{get_swWeatherData,list-method}
\alias{set_swWeatherData<-,swInputData,swWeatherData-method}
\alias{set_swWeatherData<-,list,swWeatherData-method}
\alias{get_swProd,swInputData-method}
\alias{swProd_Composition,swInputData-method}
\alias{swProd_Albedo,swInputData-method}
\alias{swProd_CanopyHeight,swInputData-method}
\alias{swProd_VegInterParam,swInputData-method}
\alias{swProd_LitterInterParam,swInputData-method}
\alias{swProd_EsTpartitioning_param,swInputData-method}
\alias{swProd_Es_param_limit,swInputData-method}
\alias{swProd_Shade,swInputData-method}
\alias{swProd_HydrRedstro_use,swInputData-method}
\alias{swProd_HydrRedstro,swInputData-method}
\alias{swProd_CritSoilWaterPotential,swInputData-method}
\alias{swProd_CO2Coefficients,swInputData-method}
\alias{swProd_MonProd_veg,swInputData,ANY-method}
\alias{swProd_MonProd_grass,swInputData-method}
\alias{swProd_MonProd_shrub,swInputData-method}
\alias{swProd_MonProd_tree,swInputData-method}
\alias{swProd_MonProd_forb,swInputData-method}
\alias{set_swProd<-,swInputData-method}
\alias{swProd_Composition<-,swInputData-method}
\alias{swProd_Albedo<-,swInputData-method}
\alias{swProd_CanopyHeight<-,swInputData-method}
\alias{swProd_VegInterParam<-,swInputData-method}
\alias{swProd_LitterInterParam<-,swInputData-method}
\alias{swProd_EsTpartitioning_param<-,swInputData-method}
\alias{swProd_Es_param_limit<-,swInputData-method}
\alias{swProd_Shade<-,swInputData-method}
\alias{swProd_HydrRedstro_use<-,swInputData-method}
\alias{swProd_HydrRedstro<-,swInputData-method}
\alias{swProd_CritSoilWaterPotential<-,swInputData-method}
\alias{swProd_CO2Coefficients<-,swInputData-method}
\alias{swProd_MonProd_veg<-,swInputData,ANY,ANY-method}
\alias{swProd_MonProd_grass<-,swInputData-method}
\alias{swProd_MonProd_shrub<-,swInputData-method}
\alias{swProd_MonProd_tree<-,swInputData-method}
\alias{swProd_MonProd_forb<-,swInputData-method}
\alias{get_swSite,swInputData-method}
\alias{swSite_SWClimits,swInputData-method}
\alias{swSite_ModelFlags,swInputData-method}
\alias{swSite_ModelCoefficients,swInputData-method}
\alias{swSite_SnowSimulationParams,swInputData-method}
\alias{swSite_DrainageCoefficient,swInputData-method}
\alias{swSite_EvapCoefficients,swInputData-method}
\alias{swSite_TranspCoefficients,swInputData-method}
\alias{swSite_IntrinsicSiteParams,swInputData-method}
\alias{swSite_SoilTemperatureFlag,swInputData-method}
\alias{swSite_SoilTemperatureConsts,swInputData-method}
\alias{swSite_TranspirationRegions,swInputData-method}
\alias{set_swSite<-,swInputData-method}
\alias{swSite_SWClimits<-,swInputData-method}
\alias{swSite_ModelFlags<-,swInputData-method}
\alias{swSite_ModelCoefficients<-,swInputData-method}
\alias{swSite_SnowSimulationParams<-,swInputData-method}
\alias{swSite_DrainageCoefficient<-,swInputData-method}
\alias{swSite_EvapCoefficients<-,swInputData-method}
\alias{swSite_TranspCoefficients<-,swInputData-method}
\alias{swSite_IntrinsicSiteParams<-,swInputData-method}
\alias{swSite_SoilTemperatureFlag<-,swInputData-method}
\alias{swSite_SoilTemperatureConsts<-,swInputData-method}
\alias{swSite_TranspirationRegions<-,swInputData-method}
\alias{get_swSoils,swInputData-method}
\alias{swSoils_Layers,swInputData-method}
\alias{set_swSoils<-,swInputData,swSoils-method}
\alias{swSoils_Layers<-,swInputData,matrix-method}
\alias{get_swSWC,swInputData-method}
\alias{swSWC_use,swInputData-method}
\alias{swSWC_prefix,swInputData-method}
\alias{swSWC_FirstYear,swInputData-method}
\alias{swSWC_Method,swInputData-method}
\alias{swSWC_HistoricList,swInputData-method}
\alias{swSWC_HistoricData,swInputData-method}
\alias{set_swSWC<-,swInputData,swSWC-method}
\alias{swSWC_use<-,swInputData,logical-method}
\alias{swSWC_prefix<-,swInputData,character-method}
\alias{swSWC_FirstYear<-,swInputData,integer-method}
\alias{swSWC_Method<-,swInputData,integer-method}
\alias{swSWC_HistoricList<-,swInputData,list-method}
\alias{swSWC_HistoricData<-,swInputData,swSWC_hist-method}
\alias{get_swCarbon,swInputData-method}
\alias{swCarbon_Use_Bio,swInputData-method}
\alias{swCarbon_Use_WUE,swInputData-method}
\alias{swCarbon_Scenario,swInputData-method}
\alias{swCarbon_DeltaYear,swInputData-method}
\alias{swCarbon_CO2ppm,swInputData-method}
\alias{set_swCarbon<-,swInputData-method}
\alias{swCarbon_Use_Bio<-,swInputData-method}
\alias{swCarbon_Use_WUE<-,swInputData-method}
\alias{swCarbon_Scenario<-,swInputData-method}
\alias{swCarbon_DeltaYear<-,swInputData-method}
\alias{swCarbon_CO2ppm<-,swInputData-method}
\alias{get_swOUT,swInputData-method}
\alias{swOUT_TimeStep,swInputData-method}
\alias{swOUT_OutputSeparator,swInputData-method}
\alias{set_swOUT<-,swInputData-method}
\alias{swOUT_TimeStep<-,swInputData-method}
\alias{swOUT_TimeStepsForEveryKey<-,swInputData-method}
\alias{swOUT_OutputSeparator<-,swInputData-method}
\alias{swLog_setLine<-,swInputData-method}
\alias{swReadLines,swInputData,character-method}
\title{Class \code{"swInputData"}}
\usage{
\S4method{initialize}{swInputData}(.Object)

\S4method{get_version}{swInputData}(object)

\S4method{get_timestamp}{swInputData}(object)

\S4method{get_swFiles}{swInputData}(object)

\S4method{swFiles_ProjDir}{swInputData}(object)

\S4method{swFiles_filesIn}{swInputData}(object)

\S4method{swFiles_Years}{swInputData}(object)

\S4method{swFiles_LogFile}{swInputData}(object)

\S4method{swFiles_SiteParams}{swInputData}(object)

\S4method{swFiles_Soils}{swInputData}(object)

\S4method{swFiles_WeatherSetup}{swInputData}(object)

\S4method{swFiles_MarkovProbs}{swInputData}(object)

\S4method{swFiles_MarkovCov}{swInputData}(object)

\S4method{swFiles_Cloud}{swInputData}(object)

\S4method{swFiles_Prod}{swInputData}(object)

\S4method{swFiles_Estab}{swInputData}(object)

\S4method{swFiles_Carbon}{swInputData}(object)

\S4method{swFiles_SWCsetup}{swInputData}(object)

\S4method{swFiles_Output}{swInputData}(object)

\S4method{swFiles_WeatherPrefix}{swInputData}(object)

\S4method{swFiles_OutputPrefix}{swInputData}(object)

\S4method{set_swFiles}{swInputData}(object) <- value

\S4method{swFiles_ProjDir}{swInputData}(object) <- value

\S4method{swFiles_filesIn}{swInputData}(object) <- value

\S4method{swFiles_Years}{swInputData}(object) <- value

\S4method{swFiles_LogFile}{swInputData}(object) <- value

\S4method{swFiles_SiteParams}{swInputData}(object) <- value

\S4method{swFiles_Soils}{swInputData}(object) <- value

\S4method{swFiles_WeatherSetup}{swInputData}(object) <- value

\S4method{swFiles_MarkovProbs}{swInputData}(object) <- value

\S4method{swFiles_MarkovCov}{swInputData}(object) <- value

\S4method{swFiles_Cloud}{swInputData}(object) <- value

\S4method{swFiles_Prod}{swInputData}(object) <- value

\S4method{swFiles_Estab}{swInputData}(object) <- value

\S4method{swFiles_Carbon}{swInputData}(object) <- value

\S4method{swFiles_SWCsetup}{swInputData}(object) <- value

\S4method{swFiles_Output}{swInputData}(object) <- value

\S4method{swFiles_WeatherPrefix}{swInputData}(object) <- value

\S4method{swFiles_OutputPrefix}{swInputData}(object) <- value

\S4method{get_swYears}{swInputData}(object)

\S4method{swYears_StartYear}{swInputData}(object)

\S4method{swYears_EndYear}{swInputData}(object)

\S4method{swYears_FDOFY}{swInputData}(object)

\S4method{swYears_EDOEY}{swInputData}(object)

\S4method{swYears_isNorth}{swInputData}(object)

\S4method{set_swYears}{swInputData}(object) <- value

\S4method{swYears_StartYear}{swInputData}(object) <- value

\S4method{swYears_EndYear}{swInputData}(object) <- value

\S4method{swYears_FDOFY}{swInputData}(object) <- value

\S4method{swYears_EDOEY}{swInputData}(object) <- value

\S4method{swYears_isNorth}{swInputData}(object) <- value

\S4method{get_swCloud}{swInputData}(object)

\S4method{set_swCloud}{swInputData}(object) <- value

\S4method{get_swWeather}{swInputData}(object)

\S4method{swWeather_DaysRunningAverage}{swInputData}(object)

\S4method{swWeather_FirstYearHistorical}{swInputData}(object)

\S4method{swWeather_pct_SnowDrift}{swInputData}(object)

\S4method{swWeather_pct_SnowRunoff}{swInputData}(object)

\S4method{swWeather_UseMarkov}{swInputData}(object)

\S4method{swWeather_UseSnow}{swInputData}(object)

\S4method{swWeather_MonScalingParams}{swInputData}(object)

\S4method{set_swWeather}{swInputData}(object) <- value

\S4method{swWeather_DaysRunningAverage}{swInputData}(object) <- value

\S4method{swWeather_FirstYearHistorical}{swInputData}(object) <- value

\S4method{swWeather_pct_SnowDrift}{swInputData}(object) <- value

\S4method{swWeather_pct_SnowRunoff}{swInputData}(object) <- value

\S4method{swWeather_UseMarkov}{swInputData}(object) <- value

\S4method{swWeather_UseSnow}{swInputData}(object) <- value

\S4method{swWeather_MonScalingParams}{swInputData}(object) <- value

\S4method{get_swCloud}{swInputData}(object)

\S4method{swCloud_SkyCover}{swInputData}(object)

\S4method{swCloud_WindSpeed}{swInputData}(object)

\S4method{swCloud_Humidity}{swInputData}(object)

\S4method{swCloud_Transmissivity}{swInputData}(object)

\S4method{swCloud_SnowDensity}{swInputData}(object)

\S4method{swCloud_RainEvents}{swInputData}(object)

\S4method{set_swCloud}{swInputData}(object) <- value

\S4method{swCloud_SkyCover}{swInputData}(object) <- value

\S4method{swCloud_WindSpeed}{swInputData}(object) <- value

\S4method{swCloud_Humidity}{swInputData}(object) <- value

\S4method{swCloud_Transmissivity}{swInputData}(object) <- value

\S4method{swCloud_SnowDensity}{swInputData}(object) <- value

\S4method{swCloud_RainEvents}{swInputData}(object) <- value

\S4method{get_Markov}{swInputData}(object)

\S4method{swMarkov_Prob}{swInputData}(object)

\S4method{swMarkov_Conv}{swInputData}(object)

\S4method{set_Markov}{swInputData}(object) <- value

\S4method{swMarkov_Prob}{swInputData}(object) <- value

\S4method{swMarkov_Conv}{swInputData}(object) <- value

\S4method{get_WeatherHistory}{swInputData}(object)

\S4method{set_WeatherHistory}{swInputData,list}(object) <- value

\S4method{get_swWeatherData}{swInputData}(object, year)

\S4method{get_swWeatherData}{list}(object, year)

\S4method{set_swWeatherData}{swInputData,swWeatherData}(object) <- value

\S4method{set_swWeatherData}{list,swWeatherData}(object) <- value

\S4method{get_swProd}{swInputData}(object)

\S4method{swProd_Composition}{swInputData}(object)

\S4method{swProd_Albedo}{swInputData}(object)

\S4method{swProd_CanopyHeight}{swInputData}(object)

\S4method{swProd_VegInterParam}{swInputData}(object)

\S4method{swProd_LitterInterParam}{swInputData}(object)

\S4method{swProd_EsTpartitioning_param}{swInputData}(object)

\S4method{swProd_Es_param_limit}{swInputData}(object)

\S4method{swProd_Shade}{swInputData}(object)

\S4method{swProd_HydrRedstro_use}{swInputData}(object)

\S4method{swProd_HydrRedstro}{swInputData}(object)

\S4method{swProd_CritSoilWaterPotential}{swInputData}(object)

\S4method{swProd_CO2Coefficients}{swInputData}(object)

\S4method{swProd_MonProd_veg}{swInputData,ANY}(object, vegtype)

\S4method{swProd_MonProd_grass}{swInputData}(object)

\S4method{swProd_MonProd_shrub}{swInputData}(object)

\S4method{swProd_MonProd_tree}{swInputData}(object)

\S4method{swProd_MonProd_forb}{swInputData}(object)

\S4method{set_swProd}{swInputData}(object) <- value

\S4method{swProd_Composition}{swInputData}(object) <- value

\S4method{swProd_Albedo}{swInputData}(object) <- value

\S4method{swProd_CanopyHeight}{swInputData}(object) <- value

\S4method{swProd_VegInterParam}{swInputData}(object) <- value

\S4method{swProd_LitterInterParam}{swInputData}(object) <- value

\S4method{swProd_EsTpartitioning_param}{swInputData}(object) <- value

\S4method{swProd_Es_param_limit}{swInputData}(object) <- value

\S4method{swProd_Shade}{swInputData}(object) <- value

\S4method{swProd_HydrRedstro_use}{swInputData}(object) <- value

\S4method{swProd_HydrRedstro}{swInputData}(object) <- value

\S4method{swProd_CritSoilWaterPotential}{swInputData}(object) <- value

\S4method{swProd_CO2Coefficients}{swInputData}(object) <- value

\S4method{swProd_MonProd_veg}{swInputData,ANY,ANY}(object, vegtype) <- value

\S4method{swProd_MonProd_grass}{swInputData}(object) <- value

\S4method{swProd_MonProd_shrub}{swInputData}(object) <- value

\S4method{swProd_MonProd_tree}{swInputData}(object) <- value

\S4method{swProd_MonProd_forb}{swInputData}(object) <- value

\S4method{get_swSite}{swInputData}(object)

\S4method{swSite_SWClimits}{swInputData}(object)

\S4method{swSite_ModelFlags}{swInputData}(object)

\S4method{swSite_ModelCoefficients}{swInputData}(object)

\S4method{swSite_SnowSimulationParams}{swInputData}(object)

\S4method{swSite_DrainageCoefficient}{swInputData}(object)

\S4method{swSite_EvapCoefficients}{swInputData}(object)

\S4method{swSite_TranspCoefficients}{swInputData}(object)

\S4method{swSite_IntrinsicSiteParams}{swInputData}(object)

\S4method{swSite_SoilTemperatureFlag}{swInputData}(object)

\S4method{swSite_SoilTemperatureConsts}{swInputData}(object)

\S4method{swSite_TranspirationRegions}{swInputData}(object)

\S4method{set_swSite}{swInputData}(object) <- value

\S4method{swSite_SWClimits}{swInputData}(object) <- value

\S4method{swSite_ModelFlags}{swInputData}(object) <- value

\S4method{swSite_ModelCoefficients}{swInputData}(object) <- value

\S4method{swSite_SnowSimulationParams}{swInputData}(object) <- value

\S4method{swSite_DrainageCoefficient}{swInputData}(object) <- value

\S4method{swSite_EvapCoefficients}{swInputData}(object) <- value

\S4method{swSite_TranspCoefficients}{swInputData}(object) <- value

\S4method{swSite_IntrinsicSiteParams}{swInputData}(object) <- value

\S4method{swSite_SoilTemperatureFlag}{swInputData}(object) <- value

\S4method{swSite_SoilTemperatureConsts}{swInputData}(object) <- value

\S4method{swSite_TranspirationRegions}{swInputData}(object) <- value

\S4method{get_swSoils}{swInputData}(object)

\S4method{swSoils_Layers}{swInputData}(object)

\S4method{set_swSoils}{swInputData,swSoils}(object) <- value

\S4method{swSoils_Layers}{swInputData,matrix}(object) <- value

\S4method{get_swSWC}{swInputData}(object)

\S4method{swSWC_use}{swInputData}(object)

\S4method{swSWC_prefix}{swInputData}(object)

\S4method{swSWC_FirstYear}{swInputData}(object)

\S4method{swSWC_Method}{swInputData}(object)

\S4method{swSWC_HistoricList}{swInputData}(object)

\S4method{swSWC_HistoricData}{swInputData}(object, year)

\S4method{set_swSWC}{swInputData,swSWC}(object) <- value

\S4method{swSWC_use}{swInputData,logical}(object) <- value

\S4method{swSWC_prefix}{swInputData,character}(object) <- value

\S4method{swSWC_FirstYear}{swInputData,integer}(object) <- value

\S4method{swSWC_Method}{swInputData,integer}(object) <- value

\S4method{swSWC_HistoricList}{swInputData,list}(object) <- value

\S4method{swSWC_HistoricData}{swInputData,swSWC_hist}(object) <- value

\S4method{get_swCarbon}{swInputData}(object)

\S4method{swCarbon_Use_Bio}{swInputData}(object)

\S4method{swCarbon_Use_WUE}{swInputData}(object)

\S4method{swCarbon_Scenario}{swInputData}(object)

\S4method{swCarbon_DeltaYear}{swInputData}(object)

\S4method{swCarbon_CO2ppm}{swInputData}(object)

\S4method{set_swCarbon}{swInputData}(object) <- value

\S4method{swCarbon_Use_Bio}{swInputData}(object) <- value

\S4method{swCarbon_Use_WUE}{swInputData}(object) <- value

\S4method{swCarbon_Scenario}{swInputData}(object) <- value

\S4method{swCarbon_DeltaYear}{swInputData}(object) <- value

\S4method{swCarbon_CO2ppm}{swInputData}(object) <- value

\S4method{get_swOUT}{swInputData}(object)

\S4method{swOUT_TimeStep}{swInputData}(object)

\S4method{swOUT_OutputSeparator}{swInputData}(object)

\S4method{set_swOUT}{swInputData}(object) <- value

\S4method{swOUT_TimeStep}{swInputData}(object) <- value

\S4method{swOUT_TimeStepsForEveryKey}{swInputData}(object) <- value

\S4method{swOUT_OutputSeparator}{swInputData}(object) <- value

\S4method{swLog_setLine}{swInputData}(object) <- value

\S4method{swReadLines}{swInputData,character}(object, file)
}
\arguments{
\item{.Object}{An object of class \code{\linkS4class{swInputData}}.}

\item{object}{An object of class \code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the object.}

\item{year}{An integer value. The calendar year of the weather or
\var{SWC} \code{data} object.}

\item{vegtype}{The name or index of the vegetation type.}

\item{file}{A character string. The file name from which to read.}

\item{...}{Further arguments to methods.}
}
\description{
This class is a container class to the input file \var{S4} objects. The
generic functions listed work on this and the proper \pkg{rSOILWAT2}
classes in the container's slots.
}
\details{
\code{\linkS4class{swInputData}} consists of slots for each file that is
read in. These slots can be accessed via the following functions: \itemize{
  \item \code{\link{get_Markov}}
  \item \code{\link{get_swCloud}}
  \item \code{\link{get_swFiles}}
  \item \code{\link{get_swOUT}}
  \item \code{\link{get_swProd}}
  \item \code{\link{get_swSite}}
  \item \code{\link{get_swSoils}}
  \item \code{\link{get_swSWC}}
  \item \code{\link{get_swWeather}}
  \item \code{\link{get_swWeatherData}}
  \item \code{\link{get_swYears}}
  \item \code{\link{get_WeatherHistory}}
}

Generic methods to get/set individual elements follow a format:
\code{"sw" + filename + "_" + option}, e.g.
\itemize{
  \item \code{\link{swFiles_Cloud}}
  \item \code{\link{swProd_Albedo}}
}
}
\examples{
showClass("swInputData")
x <- new("swInputData")

}
\seealso{
\code{\linkS4class{swFiles}} \code{\linkS4class{swYears}}
\code{\linkS4class{swWeather}} \code{\linkS4class{swCloud}}
\code{\linkS4class{swMarkov}} \code{\linkS4class{swProd}}
\code{\linkS4class{swSite}} \code{\linkS4class{swSoils}}
\code{\linkS4class{swEstab}} \code{\linkS4class{swOUT}}
\code{\linkS4class{swSWC}} \code{\linkS4class{swLog}}
\code{\linkS4class{swCarbon}}
}
