% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J_swLog.R
\docType{class}
\name{swLog-class}
\alias{swLog-class}
\alias{initialize,swLog-method}
\title{Class \code{"swLog"}}
\usage{
\S4method{initialize}{swLog}(.Object, ...)
}
\arguments{
\item{.Object}{An object of class \code{\linkS4class{swLog}}.}

\item{...}{Further arguments to methods.}
}
\description{
The methods listed below work on this class and the proper slot of the class
  \code{\linkS4class{swInputData}}.
}
\examples{
showClass("swLog")
x <- new("swLog")

}
\seealso{
\code{\linkS4class{swInputData}} \code{\linkS4class{swFiles}}
\code{\linkS4class{swWeather}} \code{\linkS4class{swCloud}}
\code{\linkS4class{swMarkov}} \code{\linkS4class{swProd}}
\code{\linkS4class{swSite}} \code{\linkS4class{swSoils}}
\code{\linkS4class{swEstab}} \code{\linkS4class{swOUT}}
\code{\linkS4class{swSWC}} \code{\linkS4class{swInputData}}
}
