% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsoilwat31-package.R
\docType{class}
\name{swMarkov-class}
\alias{swMarkov-class}
\alias{swMarkov_Conv<-,swMarkov-method}
\alias{swMarkov_Conv,swMarkov-method}
\alias{swMarkov_Prob<-,swMarkov-method}
\alias{swMarkov_Prob,swMarkov-method}
\alias{swClear,swMarkov-method}
\alias{swWriteLines,swMarkov-method}
\alias{swReadLines,swMarkov-method}
\title{Class \code{"swMarkov"}}
\description{
The generic functions listed work on this and the proper sw Class in the container's slots.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("swMarkov", ...)}.
}

\examples{

showClass("swMarkov")

}
\references{
https://github.com/Burke-Lauenroth-Lab/Rsoilwat
}
\seealso{
\code{\linkS4class{swInputData}} \code{\linkS4class{swFiles}}
\code{\linkS4class{swWeather}} \code{\linkS4class{swCloud}}
\code{\linkS4class{swInputData}} \code{\linkS4class{swProd}}
\code{\linkS4class{swSite}} \code{\linkS4class{swSoils}}
\code{\linkS4class{swEstab}} \code{\linkS4class{swOUT}}
\code{\linkS4class{swSWC}} \code{\linkS4class{swLog}}
}
\author{
Ryan Murphy
}
\keyword{classes}
