% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_swWeather.R
\docType{class}
\name{swMonthlyScalingParams-class}
\alias{swMonthlyScalingParams-class}
\alias{initialize,swMonthlyScalingParams-method}
\title{Class \code{"swMonthlyScalingParams"}}
\usage{
\S4method{initialize}{swMonthlyScalingParams}(.Object, ...)
}
\arguments{
\item{.Object}{An object of class \code{\linkS4class{swMonthlyScalingParams}}.}

\item{...}{Further arguments to methods.}
}
\description{
The methods listed below work on this class and the proper slot of the class
  \code{\linkS4class{swInputData}}.
}
\examples{
showClass("swMonthlyScalingParams")
x <- new("swMonthlyScalingParams")

}
\seealso{
\code{\linkS4class{swInputData}} \code{\linkS4class{swFiles}}
\code{\linkS4class{swWeather}} \code{\linkS4class{swCloud}}
\code{\linkS4class{swMarkov}} \code{\linkS4class{swProd}}
\code{\linkS4class{swSite}} \code{\linkS4class{swSoils}}
\code{\linkS4class{swEstab}} \code{\linkS4class{swOUT}}
\code{\linkS4class{swSWC}} \code{\linkS4class{swLog}}
}
