% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_swOut.R
\docType{class}
\name{swOUT-class}
\alias{swOUT-class}
\alias{initialize,swOUT-method}
\alias{get_swOUT,swOUT-method}
\alias{swOUT_TimeStep,swOUT-method}
\alias{swOUT_OutputSeparator,swOUT-method}
\alias{set_swOUT<-,swOUT-method}
\alias{activate_swOUT_OutKey<-,swOUT-method}
\alias{deactivate_swOUT_OutKey<-,swOUT-method}
\alias{swOUT_TimeStep<-,swOUT-method}
\alias{swOUT_TimeStepsForEveryKey<-,swOUT-method}
\alias{swOUT_OutputSeparator<-,swOUT-method}
\alias{swReadLines,swOUT,character-method}
\title{Class \code{swOUT}}
\usage{
\S4method{initialize}{swOUT}(.Object, ...)

\S4method{get_swOUT}{swOUT}(object)

\S4method{swOUT_TimeStep}{swOUT}(object)

\S4method{swOUT_OutputSeparator}{swOUT}(object)

\S4method{set_swOUT}{swOUT}(object) <- value

\S4method{activate_swOUT_OutKey}{swOUT}(object) <- value

\S4method{deactivate_swOUT_OutKey}{swOUT}(object) <- value

\S4method{swOUT_TimeStep}{swOUT}(object) <- value

\S4method{swOUT_TimeStepsForEveryKey}{swOUT}(object) <- value

\S4method{swOUT_OutputSeparator}{swOUT}(object) <- value

\S4method{swReadLines}{swOUT,character}(object, file)
}
\arguments{
\item{.Object}{An object of class \code{\linkS4class{swOUT}}.}

\item{...}{Further arguments to methods.}

\item{object}{An object of class \code{\linkS4class{swOUT}}.}

\item{value}{A value to assign to a specific slot of the object.}

\item{file}{A character string. The file name from which to read.}
}
\description{
The methods listed below work on this class and the proper slot of the class
  \code{\linkS4class{swInputData}}.
}
\details{
Output can be generated for four different time steps:
  daily (\var{DY}), weekly (\var{WK}), monthly (\var{MO}), and yearly
  (\var{YR}) periods.
 We have two options to specify time steps: \itemize{
   \item The same time step(s) for every output; this option corresponds
       to specifying a line with \code{TIMESTEP ...} in the \pkg{SOILWAT2}
       input file \var{outsetup.in}. The matrix in slot \var{timeSteps}
       should have \var{SW_OUTNKEYS} rows and \var{SW_OUTNPERIODS}
       columns of which \var{use_OUTNPERIODS} contain identical values in
       each used row.
   \item A different time step for each output; however, only one time
       step per output variable can be specified. this option corresponds to
       specifying the time step in the column \var{PERIOD} in the
       \pkg{SOILWAT2} input file \var{outsetup.in}.
}
Elements that are turned off have the value \code{eSW_NoTime = 999L}.
}
\section{Slots}{

\describe{
\item{\code{outputSeparator}}{A character string. Currently, only \var{"\\t"} is
functional.}

\item{\code{timeSteps}}{An integer matrix with rows for each output key and columns
for each possible output time period. See details.}
}}

\examples{
showClass("swOUT")
x <- new("swOUT")

x <- new("swOUT")
activate_swOUT_OutKey(x) <- c("VWCMATRIC", "HYDRED")

x <- new("swOUT")
deactivate_swOUT_OutKey(x) <- c("VWCMATRIC", "HYDRED")

x <- new("swOUT")
swOUT_TimeStepsForEveryKey(x) <- c(2, 3)
identical(as.vector(unique(swOUT_TimeStep(x))), as.integer(c(2, 3)))
}
\seealso{
\code{\linkS4class{swInputData}}
}
