% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSOILWAT2-package.R
\docType{class}
\name{swOutput_KEY-class}
\alias{swOutput_KEY-class}
\alias{swOutput_KEY_Period<-,swOUT-method}
\alias{swOutput_KEY_Period,swOUT-method}
\alias{swOutput_KEY_TimeStep<-,swOUT-method}
\alias{swOutput_KEY_TimeStep,swOUT-method}
\alias{swOutput_KEY_Columns<-,swOUT-method}
\alias{swOutput_KEY_Columns,swOUT-method}
\title{Class \code{"swOutput_KEY"}}
\description{
The generic functions listed work on this and the proper sw Class in the container's slots.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("swOutput_KEY", ...)}.
}

\examples{

showClass("swOutput_KEY")

}
\references{
https://github.com/DrylandEcology/rSOILWAT2
}
\seealso{
\code{\linkS4class{swInputData}} \code{\linkS4class{swFiles}}
\code{\linkS4class{swWeather}} \code{\linkS4class{swCloud}}
\code{\linkS4class{swMarkov}} \code{\linkS4class{swProd}}
\code{\linkS4class{swSite}} \code{\linkS4class{swSoils}}
\code{\linkS4class{swEstab}} \code{\linkS4class{swInputData}}
\code{\linkS4class{swSWC}} \code{\linkS4class{swLog}}
}
\author{
Ryan Murphy
}
\keyword{classes}
