% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swOutput_setKEY<-}
\alias{swOutput_setKEY<-}
\title{\code{swOutput_setKEY<-}}
\usage{
swOutput_setKEY(object, index) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swOutput}} or
\code{\linkS4class{swInputData}}.}

\item{index}{An integer value. The "key" (slot) position.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
\code{swOutput_setKEY<-}
}
\seealso{
\code{\linkS4class{swOutput}} and \code{\linkS4class{swInputData}}
}
