% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Pedotransfer_Functions.R
\name{swrc_conversion}
\alias{swrc_conversion}
\alias{swrc_swp_to_vwc}
\alias{swrc_vwc_to_swp}
\title{Conversion between soil water content and soil water potential}
\usage{
swrc_conversion(
  direction = c("swp_to_vwc", "vwc_to_swp"),
  x,
  fcoarse,
  layer_width,
  swrc,
  sand = NULL,
  clay = NULL
)

swrc_swp_to_vwc(
  swp_MPa,
  fcoarse,
  layer_width,
  swrc = list(swrc_type = 1L, pdf_type = 1L, swrcp = NULL),
  sand = NULL,
  clay = NULL
)

swrc_vwc_to_swp(
  vwcBulk,
  fcoarse,
  layer_width,
  swrc = list(swrc_type = 1L, pdf_type = 1L, swrcp = NULL),
  sand = NULL,
  clay = NULL
)
}
\arguments{
\item{direction}{A character string. Indicates the direction of
soil water conversion.}

\item{x}{A numeric value or vector. The soil water values to be converted,
either soil water potential (units \verb{[MPa]}) or
volumetric water content (units \verb{[cm/cm]}).}

\item{fcoarse}{A numeric value or vector.
Coarse fragments, e.g., gravel, (> 2 mm; units of \verb{[m3/m3]})
relative to the whole soil of each soil layer.
\code{fcoarse} is required, for instance, to translate between
values relative to the matric soil component (< 2 mm fraction) and
relative to the whole soil (matric soil plus coarse fragments).}

\item{layer_width}{A numeric value or vector.
Depth interval, width, of each soil layer (units of \code{cm}).
\code{layer_width} is required to translate between
soil water content of a soil layer and volumetric water content.}

\item{swrc}{A named list.
Contains all necessary elements of a \code{SWRC},
i.e., \code{type} (short for \code{swrc_type}) and \code{swrcp},
or all necessary elements to estimate parameters of a \code{SWRC} given
soil parameters, i.e., \code{swrc_type} and \code{pdf_type}.}

\item{sand}{A numeric value or vector.
Sand content of the matric soil component
(< 2 mm fraction; units of \verb{[g/g]}) of each soil layer.}

\item{clay}{A numeric value or vector.
Clay content of the matric soil component
(< 2 mm fraction; units of \verb{[g/g]}) of each soil layer.}

\item{swp_MPa}{A numeric object. The soil water potential values
(units `[MPa]`) to be converted to
volumetric water content (relative to the whole soil; units `[cm/cm]`).}

\item{vwcBulk}{A numeric object. The volumetric water content values
(relative to the whole soil; units `[cm/cm]`)
to be converted to soil water potential values (units `[MPa]`).}
}
\description{
Conversion between soil water content and soil water potential
}
\section{Functions}{
\itemize{
\item \code{swrc_swp_to_vwc}: Convenience wrapper
to convert from `SWP` to `VWC` with selected `SWRC`

\item \code{swrc_vwc_to_swp}: Convenience wrapper
to convert from `VWC` to `SWP` with selected `SWRC`
}}

\section{Details}{

See \link{SWRCs} for suitable values of \code{swrc_type}.


For backward compatibility, \code{fcoarse} and \code{layer_width} may be missing.
If they are missing, then the soils are assumed to contain
\verb{0\%} coarse fragments and be represented by \verb{1 cm} wide soil layers.


Arguments \code{sand} and \code{clay} are only required if \code{SWRC} parameter values
need to be estimated on the fly, i.e., if \code{swrc} does not contain
the element \code{swrcp} (with suitable \code{SWRC} parameter values).
This is handled by \code{\link[=pdf_estimate]{pdf_estimate()}} and additionally requires
the element \code{pdf_type} for argument \code{swrc}.
}

\examples{
fsand <- c(0.5, 0.3)
fclay <- c(0.2, 0.1)
fcrs1 <- c(0, 0)
fcrs2 <- c(0.4, 0.1)

p1 <- pdf_estimate(sand = fsand, clay = fclay, fcoarse = fcrs1)
swrc_swp_to_vwc(-1.5, fcoarse = fcrs1, swrc = list(type = 1, swrcp = p1))
swrc_swp_to_vwc(-1.5, fcoarse = fcrs1, sand = fsand, clay = fclay)
swrc_vwc_to_swp(
  c(0.10, 0.15, 0.20),
  fcoarse = fcrs1,
  swrc = list(type = 1, swrcp = p1)
)

p2 <- pdf_estimate(sand = fsand, clay = fclay, fcoarse = fcrs2)
swrc_swp_to_vwc(-1.5, fcoarse = fcrs2, swrc = list(type = 1, swrcp = p2))
(1 - fcrs2) * swrc_swp_to_vwc(-1.5, swrc = list(type = 1, swrcp = p2))
swrc_swp_to_vwc(-1.5, fcoarse = fcrs2, sand = fsand, clay = fclay)
swrc_vwc_to_swp(
  c(0.10, 0.15, 0.20),
  fcoarse = fcrs2,
  swrc = list(type = 1, swrcp = p2)
)


# Shape of SWRC


}
\seealso{
\code{\link[=pdf_estimate]{pdf_estimate()}},
\code{\link[=check_swrcp]{check_swrcp()}}
}
