% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Pedotransfer_Functions.R
\name{SWRCs}
\alias{SWRCs}
\title{Functionality for Soil Water Retention Curves (\code{SWRC})}
\arguments{
\item{sand}{A numeric value or vector.
Sand content of the matric soil component
(< 2 mm fraction; units of \verb{[g/g]}) of each soil layer.}

\item{clay}{A numeric value or vector.
Clay content of the matric soil component
(< 2 mm fraction; units of \verb{[g/g]}) of each soil layer.}

\item{fcoarse}{A numeric value or vector.
Coarse fragments, e.g., gravel, (> 2 mm; units of \verb{[m3/m3]})
relative to the whole soil of each soil layer.
\code{fcoarse} is required, for instance, to translate between
values relative to the matric soil component (< 2 mm fraction) and
relative to the whole soil (matric soil plus coarse fragments).}

\item{bdensity}{A numeric value or vector.
Density of the whole soil
(matric soil plus coarse fragments; units \verb{[g/cm3]}).}

\item{layer_width}{A numeric value or vector.
Depth interval, width, of each soil layer (units of \code{cm}).
\code{layer_width} is required to translate between
soil water content of a soil layer and volumetric water content.}

\item{swrc_name}{An character string or vector.
The selected \code{SWRC} name
(one of \code{\link[=swrc_names]{swrc_names()}}, with default \code{"Campbell1974"}).}

\item{pdf_name}{An character string or vector.
The selected \code{PDF} name
(one of \code{\link[=pdf_names]{pdf_names()}}, with default \code{"Cosby1984AndOthers"}).}

\item{swrcp}{A numeric vector or matrix.
The parameters of a selected \code{SWRC};
each row represents one \code{SWRC}, e.g., one per soil layer.}

\item{swrc}{A named list.
Contains all necessary elements of a \code{SWRC},
i.e., \code{name} (short for \code{swrc_name}) and \code{swrcp},
or all necessary elements to estimate parameters of a \code{SWRC} given
soil parameters, i.e., \code{swrc_name} and \code{pdf_name}.}

\item{fail}{A logical value.
Issue a warning (\code{FALSE}, default) or throw an error (\code{TRUE})
if request fails.}

\item{verbose}{A logical value. If \code{TRUE}, then display
\code{SOILWAT2} internal warnings and other messages.}

\item{...}{Additional function arguments passed on or ignored.}
}
\description{
\code{SWRCs} convert between soil water content and soil water potential
using a set of parameters, see \code{\link[=swrc_swp_to_vwc]{swrc_swp_to_vwc()}} and \code{\link[=swrc_vwc_to_swp]{swrc_vwc_to_swp()}}.

The \code{SWRC} parameters may be estimated from soil properties with suitable
pedotransfer functions \code{PDFs}, see \code{\link[=pdf_estimate]{pdf_estimate()}}.

The \code{SWRC} parameters can be checked for consistency with \code{\link[=check_swrcp]{check_swrcp()}}.
}
\section{Details}{

\code{\link[=swrc_names]{swrc_names()}} lists implemented \code{SWRCs};
\code{\link[=pdf_names]{pdf_names()}} lists implemented \code{PDFs}.
}

\references{
Cosby, B. J., G. M. Hornberger, R. B. Clapp, & T. R. Ginn. 1984.
A statistical exploration of the relationships of soil moisture
characteristics to the physical properties of soils.
Water Resources Research, 20:682-690, \doi{10.1029/WR020i006p00682}
}
\seealso{
\code{\link[=swrc_names]{swrc_names()}},
\code{\link[=pdf_names]{pdf_names()}},
\code{\link[=check_pdf_availability]{check_pdf_availability()}},
\code{\link[=pdf_estimate]{pdf_estimate()}},
\code{\link[=check_swrcp]{check_swrcp()}},
\code{\link[=swrc_swp_to_vwc]{swrc_swp_to_vwc()}},
\code{\link[=swrc_vwc_to_swp]{swrc_vwc_to_swp()}}
}
