% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swReadLines}
\alias{swReadLines}
\title{\code{swReadLines}}
\usage{
swReadLines(object, file)
}
\arguments{
\item{object}{An object of a class such \code{\linkS4class{swInputData}}.}
}
\description{
\code{swReadLines}
}
\seealso{
\code{\linkS4class{swInputData}}
}
