% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_swSite.R
\docType{class}
\name{swSite-class}
\alias{swSite-class}
\alias{initialize,swSite-method}
\alias{get_swSite,swSite-method}
\alias{swSite_SWRCflags,swSite-method}
\alias{swSite_SWClimits,swSite-method}
\alias{swSite_ModelFlags,swSite-method}
\alias{swSite_ModelCoefficients,swSite-method}
\alias{swSite_SnowSimulationParams,swSite-method}
\alias{swSite_DrainageCoefficient,swSite-method}
\alias{swSite_EvapCoefficients,swSite-method}
\alias{swSite_TranspCoefficients,swSite-method}
\alias{swSite_IntrinsicSiteParams,swSite-method}
\alias{swSite_SoilTemperatureFlag,swSite-method}
\alias{swSite_SoilTemperatureConsts,swSite-method}
\alias{swSite_TranspirationRegions,swSite-method}
\alias{set_swSite<-,swSite-method}
\alias{swSite_SWRCflags<-,swSite-method}
\alias{swSite_SWClimits<-,swSite-method}
\alias{swSite_ModelFlags<-,swSite-method}
\alias{swSite_ModelCoefficients<-,swSite-method}
\alias{swSite_SnowSimulationParams<-,swSite-method}
\alias{swSite_DrainageCoefficient<-,swSite-method}
\alias{swSite_EvapCoefficients<-,swSite-method}
\alias{swSite_TranspCoefficients<-,swSite-method}
\alias{swSite_IntrinsicSiteParams<-,swSite-method}
\alias{swSite_SoilTemperatureFlag<-,swSite-method}
\alias{swSite_SoilTemperatureConsts<-,swSite-method}
\alias{swSite_TranspirationRegions<-,swSite-method}
\alias{swReadLines,swSite,character-method}
\title{Class \code{"swSite"}}
\usage{
\S4method{initialize}{swSite}(.Object, ...)

\S4method{get_swSite}{swSite}(object)

\S4method{swSite_SWRCflags}{swSite}(object)

\S4method{swSite_SWClimits}{swSite}(object)

\S4method{swSite_ModelFlags}{swSite}(object)

\S4method{swSite_ModelCoefficients}{swSite}(object)

\S4method{swSite_SnowSimulationParams}{swSite}(object)

\S4method{swSite_DrainageCoefficient}{swSite}(object)

\S4method{swSite_EvapCoefficients}{swSite}(object)

\S4method{swSite_TranspCoefficients}{swSite}(object)

\S4method{swSite_IntrinsicSiteParams}{swSite}(object)

\S4method{swSite_SoilTemperatureFlag}{swSite}(object)

\S4method{swSite_SoilTemperatureConsts}{swSite}(object)

\S4method{swSite_TranspirationRegions}{swSite}(object)

\S4method{set_swSite}{swSite}(object) <- value

\S4method{swSite_SWRCflags}{swSite}(object) <- value

\S4method{swSite_SWClimits}{swSite}(object) <- value

\S4method{swSite_ModelFlags}{swSite}(object) <- value

\S4method{swSite_ModelCoefficients}{swSite}(object) <- value

\S4method{swSite_SnowSimulationParams}{swSite}(object) <- value

\S4method{swSite_DrainageCoefficient}{swSite}(object) <- value

\S4method{swSite_EvapCoefficients}{swSite}(object) <- value

\S4method{swSite_TranspCoefficients}{swSite}(object) <- value

\S4method{swSite_IntrinsicSiteParams}{swSite}(object) <- value

\S4method{swSite_SoilTemperatureFlag}{swSite}(object) <- value

\S4method{swSite_SoilTemperatureConsts}{swSite}(object) <- value

\S4method{swSite_TranspirationRegions}{swSite}(object) <- value

\S4method{swReadLines}{swSite,character}(object, file)
}
\arguments{
\item{.Object}{An object of class \code{\linkS4class{swSite}}.}

\item{...}{Further arguments to methods.}

\item{object}{An object of class \code{\linkS4class{swSite}}.}

\item{value}{A value to assign to a specific slot of the object.}

\item{file}{A character string. The file name from which to read.}
}
\description{
The methods listed below work on this class and the proper slot of the class
  \code{\linkS4class{swInputData}}.
}
\examples{
showClass("swSite")
x <- new("swSite")

}
\seealso{
\code{\linkS4class{swInputData}} \code{\linkS4class{swFiles}}
\code{\linkS4class{swWeather}} \code{\linkS4class{swCloud}}
\code{\linkS4class{swMarkov}} \code{\linkS4class{swProd}}
\code{\linkS4class{swInputData}} \code{\linkS4class{swSoils}}
\code{\linkS4class{swEstab}} \code{\linkS4class{swOUT}}
\code{\linkS4class{swSWC}} \code{\linkS4class{swLog}}
}
