% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsoilwat31-package.R
\docType{class}
\name{swSite-class}
\alias{swSite-class}
\alias{swSite_DrainageCoefficient<-,swSite-method}
\alias{swSite_DrainageCoefficient,swSite-method}
\alias{swSite_EvapCoefficients<-,swSite-method}
\alias{swSite_EvapCoefficients,swSite-method}
\alias{swSite_IntrinsicSiteParams<-,swSite-method}
\alias{swSite_IntrinsicSiteParams,swSite-method}
\alias{swSite_ModelCoefficients<-,swSite-method}
\alias{swSite_ModelCoefficients,swSite-method}
\alias{swSite_ModelFlags<-,swSite-method}
\alias{swSite_ModelFlags,swSite-method}
\alias{swSite_SnowSimulationParams<-,swSite-method}
\alias{swSite_SnowSimulationParams,swSite-method}
\alias{swSite_SoilTemperatureConsts<-,swSite-method}
\alias{swSite_SoilTemperatureConsts,swSite-method}
\alias{swSite_SoilTemperatureFlag<-,swSite-method}
\alias{swSite_SoilTemperatureFlag,swSite-method}
\alias{swSite_SWClimits<-,swSite-method}
\alias{swSite_SWClimits,swSite-method}
\alias{swSite_TranspCoefficients<-,swSite-method}
\alias{swSite_TranspCoefficients,swSite-method}
\alias{swSite_TranspirationRegions<-,swSite-method}
\alias{swSite_TranspirationRegions,swSite-method}
\alias{swClear,swSite-method}
\alias{swWriteLines,swSite-method}
\alias{swReadLines,swSite-method}
\title{Class \code{"swSite"}}
\description{
The generic functions listed work on this and the proper sw Class in the container's slots.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("swSite", ...)}.
}

\examples{

showClass("swSite")

}
\references{
https://github.com/Burke-Lauenroth-Lab/Rsoilwat
}
\seealso{
\code{\linkS4class{swInputData}} \code{\linkS4class{swFiles}}
\code{\linkS4class{swWeather}} \code{\linkS4class{swCloud}}
\code{\linkS4class{swMarkov}} \code{\linkS4class{swProd}}
\code{\linkS4class{swInputData}} \code{\linkS4class{swSoils}}
\code{\linkS4class{swEstab}} \code{\linkS4class{swOUT}}
\code{\linkS4class{swSWC}} \code{\linkS4class{swLog}}
}
\author{
Ryan Murphy
}
\keyword{classes}
