% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R, R/F_swSite.R,
%   R/K_swContainer.R
\name{swSite_hasSWRCp}
\alias{swSite_hasSWRCp}
\alias{swSite_hasSWRCp<-}
\alias{swSite_hasSWRCp,swSite-method}
\alias{swSite_hasSWRCp<-,swSite-method}
\alias{swSite_hasSWRCp,swInputData-method}
\alias{swSite_hasSWRCp<-,swInputData-method}
\title{Are mineral soil \code{SWRC} parameters provided in \code{SWRCp}?}
\usage{
swSite_hasSWRCp(object)

swSite_hasSWRCp(object) <- value

\S4method{swSite_hasSWRCp}{swSite}(object)

\S4method{swSite_hasSWRCp}{swSite}(object) <- value

\S4method{swSite_hasSWRCp}{swInputData}(object)

\S4method{swSite_hasSWRCp}{swInputData}(object) <- value
}
\arguments{
\item{object}{An object of class \linkS4class{swSite} or \linkS4class{swInputData}.}

\item{value}{A logical value.}
}
\value{
A logical value.
\code{TRUE} if mineral soil \code{SWRC} parameters are provided in \code{SWRCp};
\code{FALSE} if \code{SWRCp} should be estimated during a simulation run
via specified pedotransfer function
(see \code{"ptf_name"} of \code{\link[=swSite_SWRCflags]{swSite_SWRCflags()}}).
}
\description{
Are mineral soil \code{SWRC} parameters provided in \code{SWRCp}?
}
