% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swSite_SWRCflags}
\alias{swSite_SWRCflags}
\title{\code{swSite_SWRCflags}}
\usage{
swSite_SWRCflags(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swSite}} or
\code{\linkS4class{swInputData}}.}
}
\description{
\code{swSite_SWRCflags}
}
\seealso{
\code{\linkS4class{swSite}} and \code{\linkS4class{swInputData}}
}
