% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swSoils_Layers}
\alias{swSoils_Layers}
\title{\code{swSoils_Layers}}
\usage{
swSoils_Layers(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swSoils}} or \code{\linkS4class{swInputData}}.}
}
\description{
\code{swSoils_Layers}
}
\seealso{
\code{\linkS4class{swSoils}} and \code{\linkS4class{swInputData}}
}
