% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R, R/F_swSoils.R,
%   R/K_swContainer.R
\name{swSoils_omSWRCp}
\alias{swSoils_omSWRCp}
\alias{swSoils_omSWRCp<-}
\alias{swSoils_omSWRCp,swSoils-method}
\alias{swSoils_omSWRCp<-,swSoils-method}
\alias{swSoils_omSWRCp,swInputData-method}
\alias{swSoils_omSWRCp<-,swInputData-method}
\title{\code{SWRC} parameters of fibric and sapric organic matter}
\usage{
swSoils_omSWRCp(object)

swSoils_omSWRCp(object) <- value

\S4method{swSoils_omSWRCp}{swSoils}(object)

\S4method{swSoils_omSWRCp}{swSoils}(object) <- value

\S4method{swSoils_omSWRCp}{swInputData}(object)

\S4method{swSoils_omSWRCp}{swInputData}(object) <- value
}
\arguments{
\item{object}{An object of class \code{\link{swSoils}} or \linkS4class{swInputData}.}

\item{value}{An object that can be converted to a data matrix and represents
required `SWRC` parameters.}
}
\value{
A data matrix.
}
\description{
\code{SWRC} parameters of fibric and sapric organic matter
}
\section{Details}{

The correct sequence for setting values is
\enumerate{
\item \code{\link[=swSoils_Layers]{swSoils_Layers()}},
\item \code{\link[=swSite_SWRCflags]{swSite_SWRCflags()}}, and
\item \code{\link[=swSoils_SWRCp]{swSoils_SWRCp()}} and \code{\link[=swSite_hasSWRCp]{swSite_hasSWRCp()}}
\item \code{\link[=swSoils_omSWRCp]{swSoils_omSWRCp()}} (and \code{\link[=swSite_depthSapric]{swSite_depthSapric()}})
}
}

\examples{
  swin <- rSOILWAT2::sw_exampleData
  swSoils_omSWRCp(swin) <- rSOILWAT2::sw2_list_omSWRCp[["Campbell1974"]]

}
