% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_swSWC.R
\docType{class}
\name{swSWC_hist-class}
\alias{swSWC_hist-class}
\alias{swSWC_hist}
\title{Class \code{"swSWC_hist"}}
\usage{
swSWC_hist(...)
}
\arguments{
\item{...}{Arguments to the helper constructor function.
Dots can either contain objects to copy into slots of that class
(must be named identical to the corresponding slot) or
be one object of that class (in which case it will be copied and
any missing slots will take their default values).
If dots are missing, then corresponding values of
\code{rSOILWAT2::sw_exampleData}
(i.e., the \pkg{SOILWAT2} "testing" defaults) are copied.}
}
\description{
The methods listed below work on this class and the proper slot of the class
  \code{\linkS4class{swInputData}}.
}
\section{Slots}{

\describe{
\item{\code{year}}{An integer value. The calendar year of the \var{SWC}
\code{data} object.}

\item{\code{data}}{A 365 x 4 or 366 x 4 matrix representing daily \var{SWC}
data for one calendar \code{year} with columns \var{doy}, \var{lyr},
\var{swc}, \var{st_err}.}
}}

