% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsoilwat31-package.R
\docType{class}
\name{swSWC_hist-class}
\alias{swClear,swSWC_hist-method}
\alias{swReadLines,swSWC_hist-method}
\alias{swSWC_hist-class}
\alias{swWriteLines,swSWC_hist-method}
\title{Class \code{"swSWC_hist"}}
\description{
The generic functions listed work on this and the proper sw Class in the container's slots.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("swSWC_hist", ...)}.
}
\examples{

showClass("swSWC_hist")

}
\author{
Ryan Murphy
}
\references{
https://github.com/Burke-Lauenroth-Lab/Rsoilwat
}
\seealso{
\code{\linkS4class{swInputData}} \code{\linkS4class{swFiles}}
\code{\linkS4class{swWeather}} \code{\linkS4class{swCloud}}
\code{\linkS4class{swMarkov}} \code{\linkS4class{swProd}}
\code{\linkS4class{swSite}} \code{\linkS4class{swSoils}}
\code{\linkS4class{swEstab}} \code{\linkS4class{swInputData}}
\code{\linkS4class{swSWC}} \code{\linkS4class{swLog}}
}
\keyword{classes}

