% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_swWeatherData.R
\docType{class}
\name{swWeatherData-class}
\alias{swWeatherData-class}
\alias{initialize,swWeatherData-method}
\alias{swReadLines,swWeatherData,character-method}
\title{Class \code{"swWeatherData"}}
\usage{
\S4method{initialize}{swWeatherData}(.Object, ..., year = 0L,
  data = NULL)

\S4method{swReadLines}{swWeatherData,character}(object, file)
}
\arguments{
\item{.Object}{An object of class \code{\linkS4class{swWeatherData}}.}

\item{...}{Further arguments to methods.}

\item{year}{An integer value. The calendar year of the weather \code{data}
object.}

\item{data}{A 365 x 4 or 366 x 4 matrix representing daily weather data for
one calendar \code{year} with columns \var{DOY}, \var{Tmax_C},
\var{Tmin_C}, and \var{PPT_cm}.}

\item{object}{An object of class \code{\linkS4class{swWeatherData}}.}

\item{file}{A character string. The file name from which to read.}
}
\description{
The methods listed below work on this class and the proper slot of the class
  \code{\linkS4class{swInputData}}.
}
\examples{
showClass("swWeatherData")
x <- new("swWeatherData")

}
\seealso{
\code{\linkS4class{swInputData}} \code{\linkS4class{swFiles}}
\code{\linkS4class{swWeather}} \code{\linkS4class{swCloud}}
\code{\linkS4class{swMarkov}} \code{\linkS4class{swProd}}
\code{\linkS4class{swSite}} \code{\linkS4class{swSoils}}
\code{\linkS4class{swEstab}} \code{\linkS4class{swInputData}}
\code{\linkS4class{swSWC}} \code{\linkS4class{swLog}}
}
