% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_swYears.R
\docType{class}
\name{swYears-class}
\alias{swYears-class}
\alias{initialize,swYears-method}
\alias{swYears_StartYear,swYears-method}
\alias{swYears_EndYear,swYears-method}
\alias{swYears_FDOFY,swYears-method}
\alias{swYears_EDOEY,swYears-method}
\alias{swYears_isNorth,swYears-method}
\alias{swYears_StartYear<-,swYears-method}
\alias{swYears_EndYear<-,swYears-method}
\alias{swYears_FDOFY<-,swYears-method}
\alias{swYears_EDOEY<-,swYears-method}
\alias{swYears_isNorth<-,swYears-method}
\alias{swReadLines,swYears,character-method}
\title{Class \code{"swYears"}}
\usage{
\S4method{initialize}{swYears}(.Object, ...)

\S4method{swYears_StartYear}{swYears}(object)

\S4method{swYears_EndYear}{swYears}(object)

\S4method{swYears_FDOFY}{swYears}(object)

\S4method{swYears_EDOEY}{swYears}(object)

\S4method{swYears_isNorth}{swYears}(object)

\S4method{swYears_StartYear}{swYears}(object) <- value

\S4method{swYears_EndYear}{swYears}(object) <- value

\S4method{swYears_FDOFY}{swYears}(object) <- value

\S4method{swYears_EDOEY}{swYears}(object) <- value

\S4method{swYears_isNorth}{swYears}(object) <- value

\S4method{swReadLines}{swYears,character}(object, file)
}
\arguments{
\item{.Object}{An object of class \code{\linkS4class{swYears}}.}

\item{...}{Further arguments to methods.}

\item{object}{An object of class \code{\linkS4class{swYears}}.}

\item{value}{A value to assign to a specific slot of the object.}

\item{file}{A character string. The file name from which to read.}
}
\description{
The methods listed below work on this class and the proper slot of the class
  \code{\linkS4class{swInputData}}.
}
\examples{
showClass("swYears")
x <- new("swYears")

}
\seealso{
\code{\linkS4class{swInputData}} \code{\linkS4class{swFiles}}
\code{\linkS4class{swWeather}} \code{\linkS4class{swCloud}}
\code{\linkS4class{swMarkov}} \code{\linkS4class{swProd}}
\code{\linkS4class{swSite}} \code{\linkS4class{swSoils}}
\code{\linkS4class{swEstab}} \code{\linkS4class{swOUT}}
\code{\linkS4class{swSWC}} \code{\linkS4class{swLog}}
}
