% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_Vegetation.R
\name{update_biomass}
\alias{update_biomass}
\title{Replace selected biomass values of a
\link[rSOILWAT2:swProd-class]{rSOILWAT2::swProd} object}
\usage{
update_biomass(fg = c("Grass", "Shrub", "Tree", "Forb"), use, prod_input,
  prod_default)
}
\arguments{
\item{fg}{A character string. One of the functional groups represented by
\pkg{rSOILWAT2}}

\item{use}{A logical vector.}
}
\description{
Replace selected biomass values of a
\link[rSOILWAT2:swProd-class]{rSOILWAT2::swProd} object
}
